Module Ppx_typed_match

match%ty x with ... matches on the runtime type of the typed value x (via Typed.get_ty) and, in each branch, rebinds x to the value cast at the static phantom type implied by the matched runtime-type constructor.

For example:

match%ty x with
| TBitVector _ -> f x (* x : [< T.sint ] t *)
| TExtension FullPtr -> g x (* x : [< T.sptr_f ] t *)
| _ -> h x (* x unchanged *)

desugars to:

match Typed.get_ty x with
| TBitVector _ ->
    let x = (Typed.cast x : Typed.([< T.sint ] t)) in
    f x
| TExtension FullPtr ->
    let x = (Typed.cast x : Typed.([< T.sptr_f ] t)) in
    g x
| _ -> h x

The casts are unchecked: correctness rests on the table below matching how values are typed. Defining it once here removes the per-site Typed.cast boilerplate and the risk of casting to the wrong type.

Several scrutinees are also supported; only the first is the get_ty subject, the rest are matched as-is:

match%ty (v, to_ty) with
| TBitVector _, TLiteral _ -> f v (* v : [< T.sint ] t; to_ty untouched *)
| _, _ -> g v

val key_of_ty_ctor : string -> string option

The phantom-type table: runtime-type constructor -> phantom tag key. The keys index into Typed.T; "seq" is special-cased (it is the applied type T.any T.sseq).

val key_of_ext_ctor : string -> string option

Same as key_of_ty_ctor but for extension constructors.

val typed : string

val lid_of_path : loc:Ppxlib.location -> string list -> Ppxlib.Longident.t Ppxlib.loc

val phantom_ty : loc:Ppxlib__.Location.t -> string -> Ppxlib__.Import.core_type

val combine : loc:Ppxlib.Location.t -> Stdlib.String.t option -> Stdlib.String.t option -> Stdlib.String.t option

val classify : Ppxlib.pattern -> Stdlib.String.t option

val classify_ext : Ppxlib.pattern -> Stdlib.String.t

val classify_case : tuple:bool -> Ppxlib.pattern -> Stdlib.String.t option

val rewrite_case : var:string -> tuple:bool -> Ppxlib.case -> Ppxlib.case

val expand : loc:Ppxlib.location -> Ppxlib.expression -> Ppxlib.expression

val ext : Ppxlib.Extension.t