Module StateKey.Adt

val mk_tuple : [< Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t list -> [> Soteria_rust_lib__Svalue__Typed.T.tuple ] Soteria_rust_lib__Svalue__Typed.t

Creates a tuple value with the given fields. This may be an array, struct, or tuple.

val mk_array : Charon.Types.ty -> [< Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t iarray -> [> Soteria_rust_lib__Svalue__Typed.T.tuple ] Soteria_rust_lib__Svalue__Typed.t

Creates an array value with the given element type and elements.

val mk_enum : Charon.Types.type_decl_ref -> [< Soteria_rust_lib__Svalue__Typed.T.sint ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t list -> [> Soteria_rust_lib__Svalue__Typed.T.enum ] Soteria_rust_lib__Svalue__Typed.t

Creates an enum value with the given discriminant and fields.

val mk_union : Charon.Types.type_decl_ref -> ([< Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t * [< Soteria_rust_lib__Svalue__Typed.T.sint ] Soteria_rust_lib__Svalue__Typed.t * [< Soteria_rust_lib__Svalue__Typed.T.nonzero ] Soteria_rust_lib__Svalue__Typed.t) list -> [> Soteria_rust_lib__Svalue__Typed.T.union ] Soteria_rust_lib__Svalue__Typed.t

Creates a union value with the given value blocks.

val mk_poly : Charon.Types.type_var_id -> [> Soteria_rust_lib__Svalue__Typed.T.poly ] Soteria_rust_lib__Svalue__Typed.t

Creates an unknown polymorphic value. HACK: what does this even mean?

val unit : [> Soteria_rust_lib__Svalue__Typed.T.tuple ] Soteria_rust_lib__Svalue__Typed.t

The unit tuple value.

val as_union : [< Soteria_rust_lib__Svalue__Typed.T.union ] Soteria_rust_lib__Svalue__Typed.t -> ([> Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t * [> Soteria_rust_lib__Svalue__Typed.T.sint ] Soteria_rust_lib__Svalue__Typed.t * [> Soteria_rust_lib__Svalue__Typed.T.nonzero ] Soteria_rust_lib__Svalue__Typed.t) list

Gets the value blocks of a union.

val as_tuple : [< Soteria_rust_lib__Svalue__Typed.T.tuple ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t list

val as_array : [< Soteria_rust_lib__Svalue__Typed.T.tuple ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t iarray

val as_enum_of_variant : Charon.Types.variant_id -> [< Soteria_rust_lib__Svalue__Typed.T.enum ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t list

val as_tuple1 : [< Soteria_rust_lib__Svalue__Typed.T.tuple ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t

Like as_tuple, but casts to a fixed-arity tuple, failing if the tuple does not have exactly that many fields. Avoids partial list pattern matches at call sites.

val as_tuple2 : [< Soteria_rust_lib__Svalue__Typed.T.tuple ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t * [> Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t

val as_tuple3 : [< Soteria_rust_lib__Svalue__Typed.T.tuple ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t * [> Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t * [> Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t

val as_type_var : [< Soteria_rust_lib__Svalue__Typed.T.poly ] Soteria_rust_lib__Svalue__Typed.t -> Charon.Types.type_var_id

val discriminant_of : [< Soteria_rust_lib__Svalue__Typed.T.enum ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.T.sint ] Soteria_rust_lib__Svalue__Typed.t

val field_of : int -> [< Soteria_rust_lib__Svalue__Typed.T.tuple ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t

Get the field of a tuple

val array_field_of : int -> [< Soteria_rust_lib__Svalue__Typed.T.tuple ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t

Get the field of an array

val field_of_variant : Charon.Types.variant_id -> int -> [< Soteria_rust_lib__Svalue__Typed.T.enum ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t

Get the field of a enum with the given variant

val set_field : int -> [< Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.T.tuple ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.T.tuple ] Soteria_rust_lib__Svalue__Typed.t

val set_field_of_variant : Charon.Types.variant_id -> int -> [< Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.T.enum ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.T.enum ] Soteria_rust_lib__Svalue__Typed.t

val update_field : int -> ([< Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t) -> [< Soteria_rust_lib__Svalue__Typed.T.tuple ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.T.tuple ] Soteria_rust_lib__Svalue__Typed.t

val set_array_field : int -> [< Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.T.tuple ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.T.tuple ] Soteria_rust_lib__Svalue__Typed.t

val update_array_field : int -> ([< Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t) -> [< Soteria_rust_lib__Svalue__Typed.T.tuple ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.T.tuple ] Soteria_rust_lib__Svalue__Typed.t

val update_field_of_variant : Charon.Types.variant_id -> int -> ([< Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.T.any ] Soteria_rust_lib__Svalue__Typed.t) -> [< Soteria_rust_lib__Svalue__Typed.T.enum ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.T.enum ] Soteria_rust_lib__Svalue__Typed.t

module Checked : sig ... end