Module StateKey.Float

val mk_bits : Soteria__Bv_values__Svalue.F.precision -> Z.t -> [> Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t

val to_bits_opt : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sint ] Soteria_rust_lib__Svalue__Typed.t option

We cannot represent a symbolic float as a bitvector, for lack of an SMT-LIB function for it. However, we can return the bitvector representation of a float if it is concrete.

val to_float_opt : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> float option

val sign_bit_opt : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> bool option

val approx : (float -> float) -> [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t option

val approx2 : (float -> float -> float) -> [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t option

val eq : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.sbool ] Soteria_rust_lib__Svalue__Typed.t

val geq : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.sbool ] Soteria_rust_lib__Svalue__Typed.t

val gt : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.sbool ] Soteria_rust_lib__Svalue__Typed.t

val leq : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.sbool ] Soteria_rust_lib__Svalue__Typed.t

val lt : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.sbool ] Soteria_rust_lib__Svalue__Typed.t

val add : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t

val sub : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t

val mul : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t

val div : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t

val rem : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t

val fmod : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t

val fmod_of_rem : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t

For fmod_of_rem res x y, given res is some representation of rem x y, returns mod. We receive res as a parameter, to allow optimisations where it is computed not through rem but through more optimal representation (e.g. using fma).

val fma : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t

fma x y z is the exact result of x * y + z, rounded once to the destination precision (rather than with intermediate rounding).

val min : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t

The IEEE 754-2019 min and max: unlike minimum and maximum, a NaN is ignored, and -0.0 is considered equal to +0.0. This is equivalent to:

let min x y =
  if is_nan x then y
  else if is_nan y then x
  else if lt x y then x
  else if gt x y then y
  else if is_negative x then x
  else y

val max : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t

IEEE 754-2019 max. See min for the differences with maximum.

val minimum : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t

The IEEE 754-2019 minimum: unlike min a NaN propagates, and -0.0 is strictly below +0.0. This is equivalent to:

let minimum x y =
  if is_nan x || is_nan y then nan
  else if lt x y then x
  else if gt x y then y
  else if is_negative x then x
  else y

val maximum : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t

IEEE-754-2019 maximum. See minimum for the differences with max.

val abs : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t

val neg : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t

val sqrt : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t

val is_normal : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.sbool ] Soteria_rust_lib__Svalue__Typed.t

val is_subnormal : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.sbool ] Soteria_rust_lib__Svalue__Typed.t

val is_zero : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.sbool ] Soteria_rust_lib__Svalue__Typed.t

val is_infinite : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.sbool ] Soteria_rust_lib__Svalue__Typed.t

val is_nan : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.sbool ] Soteria_rust_lib__Svalue__Typed.t

val is_negative : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.sbool ] Soteria_rust_lib__Svalue__Typed.t

val is_positive : [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.sbool ] Soteria_rust_lib__Svalue__Typed.t

val round : Floatml.rounding_mode -> [< Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.{T}1/shadowed/(37e963b4a9666faa8f0e337aefbcca66).sfloat ] Soteria_rust_lib__Svalue__Typed.t

val mk : Charon.Types.float_type -> string -> [> Soteria_rust_lib__Svalue__Typed.T.sfloat ] Soteria_rust_lib__Svalue__Typed.t

val zero : Charon.Types.float_type -> [> Soteria_rust_lib__Svalue__Typed.T.sfloat ] Soteria_rust_lib__Svalue__Typed.t

val neg_zero : Charon.Types.float_type -> [> Soteria_rust_lib__Svalue__Typed.T.sfloat ] Soteria_rust_lib__Svalue__Typed.t

val one : Charon.Types.float_type -> [> Soteria_rust_lib__Svalue__Typed.T.sfloat ] Soteria_rust_lib__Svalue__Typed.t

val infinity : Charon.Types.float_type -> [> Soteria_rust_lib__Svalue__Typed.T.sfloat ] Soteria_rust_lib__Svalue__Typed.t

val neg_infinity : Charon.Types.float_type -> [> Soteria_rust_lib__Svalue__Typed.T.sfloat ] Soteria_rust_lib__Svalue__Typed.t

val nan : Charon.Types.float_type -> [> Soteria_rust_lib__Svalue__Typed.T.sfloat ] Soteria_rust_lib__Svalue__Typed.t

val of_z : Charon.Types.float_type -> Soteria.Soteria_std.Z.t -> [> Soteria_rust_lib__Svalue__Typed.T.sfloat ] Soteria_rust_lib__Svalue__Typed.t

val fp_of : [< Soteria_rust_lib__Svalue__Typed.T.sfloat ] Soteria_rust_lib__Svalue__Typed.t -> Charon.Types.float_type

val cast : rounding:Floatml.rounding_mode -> fp:Charon.Types.float_type -> [< Soteria_rust_lib__Svalue__Typed.T.sfloat ] Soteria_rust_lib__Svalue__Typed.t -> [> Soteria_rust_lib__Svalue__Typed.T.sfloat ] Soteria_rust_lib__Svalue__Typed.t