Module State.Tree_block

module Encoder : sig ... end

type rust_val = Sptr_base.t Rust_val.t

val pp_rust_val : (Soteria_rust_lib.Typed.T.sint Typed.t, Soteria_rust_lib.Typed.T.sfloat Typed.t, Sptr_base.t) Rust_val.raw Fmt.t

module MemVal : sig ... end

module Expr : sig ... end

type nonrec sint = MemVal.S_bounded_int.t Sptr.DecayMap.SM.Value.t

type nonrec (!'a, !'sint) tree = {

node : 'a Soteria__Sym_states__Tree_block.node;

range : 'sint * 'sint;

children : (('a, 'sint) Soteria__Sym_states__Tree_block.tree * ('a, 'sint) Soteria__Sym_states__Tree_block.tree) option;

}

module Range : sig ... end

module Split_tree : sig ... end

A Split_tree is a simplified representation of a tree, that has no offset. It however indicates, on Nodes, at what offset the split occurs, relative to that node's start.

module Node : sig ... end

type t = {

root : (MemVal.t, sint) tree;

bound : sint option;

}

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

module SM : sig ... end

val is_empty : t -> bool

val pp_pretty : Stdlib.Format.formatter -> t -> unit

type syn =

| MemVal of {

offset : Expr.t;

len : Expr.t;

v : MemVal.syn;

}

| Bound of Expr.t

val pp_syn : Ppx_deriving_runtime.Format.formatter -> syn -> Ppx_deriving_runtime.unit

val show_syn : syn -> Ppx_deriving_runtime.string

val ins_outs : syn -> Expr.t list * Expr.t list

val lift_fixes : offset:'a Sptr.DecayMap.SM.Value.t -> len:'b Sptr.DecayMap.SM.Value.t -> MemVal.syn list -> syn list

val lift_miss : offset:'a Sptr.DecayMap.SM.Value.t -> len:'b Sptr.DecayMap.SM.Value.t -> ('c, 'd, MemVal.syn list) Sptr.DecayMap.SM.Result.t -> ('c, 'd, syn list) Sptr.DecayMap.SM.Result.t

val lift_miss_c : offset:'a Sptr.DecayMap.SM.Value.t -> len:'b Sptr.DecayMap.SM.Value.t -> ('c, MemVal.syn list) Sptr.DecayMap.SM.Consumer.t -> ('c, syn list) Sptr.DecayMap.SM.Consumer.t

val of_opt : ?mk_fixes:(unit -> 'a list Sptr.DecayMap.SM.t) -> 'b option -> ('b, 'c, 'a) Sptr.DecayMap.SM.Result.t

val to_opt : t -> t option

val with_bound_check : ?mk_fixes:(unit -> syn list list Sptr.DecayMap.SM.t) -> sint -> ((MemVal.t, sint) tree -> ('a * (MemVal.t, sint) tree, [> `OutOfBounds ] as 'b, syn list) Sptr.DecayMap.SM.Result.t) -> ('a, 'b, syn list) SM.Result.t

val assert_exclusively_owned : unit -> (unit, 'a, syn list) Soteria.Symex.Compo_res.t SM.t

val get_raw_tree_owned : MemVal.S_bounded_int.t Sptr.DecayMap.SM.Value.t -> MemVal.S_bounded_int.t Sptr.DecayMap.SM.Value.t -> ((MemVal.t, MemVal.S_bounded_int.t Sptr.DecayMap.SM.Value.t) Soteria__Sym_states__Tree_block.tree, [> `OutOfBounds ], syn list) SM.Result.t

val put_raw_tree : MemVal.S_bounded_int.t Sptr.DecayMap.SM.Value.t -> (MemVal.t, sint) tree -> (unit, [> `OutOfBounds ], syn list) SM.Result.t

val to_syn : t -> syn list

val consume_bound : Sptr.DecayMap.SM.Value.Expr.t -> t option -> (t option, 'a) Sptr.DecayMap.SM.Consumer.t

val produce_bound : Expr.t -> t option -> t option Sptr.DecayMap.SM.Producer.t

val produce_mem_val : Expr.t -> Expr.t -> MemVal.syn -> t option -> t option Sptr.DecayMap.SM.Producer.t

val consume_mem_val : Expr.t -> Expr.t -> MemVal.syn -> t option -> (t option, syn list) Sptr.DecayMap.SM.Consumer.t

val consume : syn -> t option -> (t option, syn list) Sptr.DecayMap.SM.Consumer.t

val produce : syn -> t option -> t option Sptr.DecayMap.SM.Producer.t

module Tree : sig ... end

val lift_symex : 'a Sptr.DecayMap.SM.Symex.t -> 'a SM.t

val sint_to_int : [< Soteria__Bv_values__Typed.T.any ] Soteria__Bv_values__Typed.t -> int Sptr.DecayMap.SM.t

val mk_fix_typed : 'a Sptr.DecayMap.SM.Value.t -> Charon.Types.ty -> unit -> Soteria__Sym_states__Pmap.Direct_access_patricia_tree(Rustsymex)(Soteria_rust_lib__Sptr.DecayMap.MapKey)(Soteria_rust_lib__Sptr.DecayMap.EntryExcl).t option -> (syn list list * Soteria__Sym_states__Pmap.Direct_access_patricia_tree(Rustsymex)(Soteria_rust_lib__Sptr.DecayMap.MapKey)(Soteria_rust_lib__Sptr.DecayMap.EntryExcl).t option) Rustsymex.t

val mk_fix_any : 'a Sptr.DecayMap.SM.Value.t -> 'b Sptr.DecayMap.SM.Value.t -> unit -> syn list list

val mk_fix_any_s : 'a Sptr.DecayMap.SM.Value.t -> 'b Sptr.DecayMap.SM.Value.t -> unit -> syn list list Sptr.DecayMap.SM.t

val mk_fix_tb : 'a Sptr.DecayMap.SM.Value.t -> 'b Sptr.DecayMap.SM.Value.t -> unit -> syn list list Sptr.DecayMap.SM.t

val collect_leaves : uninit:[< `Error | `Ignore ] -> Tree.t -> ((rust_val * [> Soteria__Bv_values__Typed.T.sint ] Soteria__Bv_values__Typed.t) list, [> `UninitializedMemoryAccess ], syn list) Sptr.DecayMap.SM.Result.t

val decode_mem_val : ty:Charon.Types.ty -> MemVal.leaf -> (Encoder.rust_val, [> `InvalidLayout of Charon.Types.ty | `UninitializedMemoryAccess ], MemVal.syn list) Sptr.DecayMap.SM.Result.t

val decode_lazy : ty:Charon.Types.ty -> Tree.t -> Sptr.DecayMap.SM.st -> ((Encoder.rust_val, [> `UninitializedMemoryAccess ], syn list) Soteria.Symex.Compo_res.t * Sptr.DecayMap.SM.st) Rustsymex.t

val decode_tree : ty:Charon.Types.ty -> Tree.t -> (Encoder.rust_val, [> `InvalidLayout of Charon.Types.ty | `UninitializedMemoryAccess ], syn list) Sptr.DecayMap.SM.Result.t

val merge_Borrows : Tree.t -> (Borrows.State.t option, 'a, syn list) Sptr.DecayMap.SM.Result.t

val init : (sint * sint) -> rust_val -> Borrows.State.t option -> Tree.t

val uninit : (sint * sint) -> Borrows.State.t option -> Tree.t

val zeros : (sint * sint) -> Borrows.State.t option -> Tree.t

val as_owned : ?mk_fixes: (unit -> Soteria__Sym_states__Pmap.Direct_access_patricia_tree(Rustsymex)(Soteria_rust_lib__Sptr.DecayMap.MapKey)(Soteria_rust_lib__Sptr.DecayMap.EntryExcl).t option -> ('a list * Soteria__Sym_states__Pmap.Direct_access_patricia_tree(Rustsymex)(Soteria_rust_lib__Sptr.DecayMap.MapKey)(Soteria_rust_lib__Sptr.DecayMap.EntryExcl).t option) Rustsymex.t) -> ('b, 'c) tree -> (('b, 'c) tree -> ('d, 'e, 'a) Sptr.DecayMap.SM.Result.t) -> ('d, 'e, 'a) Sptr.DecayMap.SM.Result.t

val check_owned : Soteria_rust_lib.Typed.T.sint Typed.t -> [< Soteria_rust_lib.Typed.T.nonzero ] Typed.t -> (unit, [> `OutOfBounds ], syn list) SM.Result.t

val load : ignore_borrow:bool -> Soteria_rust_lib.Typed.T.sint Typed.t -> Charon.Types.ty -> Borrows.Tag.t option -> Borrows.Tree.t option -> (Encoder.rust_val, [> `AliasingError | `InvalidLayout of Charon.Types.ty | `OutOfBounds | `UninitializedMemoryAccess ], syn list) Soteria.Symex.Compo_res.t SM.t

val store : [ `NonZero | `Zero ] Typed.t -> rust_val -> Borrows.Tag.t option -> Borrows.Tree.t option -> (unit, [> `AliasingError | `InvalidLayout of Charon.Types.ty | `OutOfBounds ], syn list) SM.Result.t

val get_init_leaves : Soteria_rust_lib.Typed.T.sint Typed.t -> [< Soteria_rust_lib.Typed.T.nonzero ] Typed.t -> ((rust_val * Soteria_rust_lib.Typed.T.sint Typed.t) list, [> `OutOfBounds | `UninitializedMemoryAccess ], syn list) SM.Result.t

val uninit_range : Soteria_rust_lib.Typed.T.sint Typed.t -> Soteria_rust_lib.Typed.T.sint Typed.t -> (unit, [> `OutOfBounds ], syn list) SM.Result.t

val zero_range : Soteria_rust_lib.Typed.T.sint Typed.t -> Soteria_rust_lib.Typed.T.sint Typed.t -> (unit, [> `OutOfBounds ], syn list) SM.Result.t

val alloc : ?zeroed:bool -> sint -> Sptr.DecayMap.SM.st -> (t * Sptr.DecayMap.SM.st) Rustsymex.t

val with_tb_access : Soteria_rust_lib.Typed.T.sint Typed.t -> Soteria_rust_lib.Typed.T.sint Typed.t -> (Borrows.State.t option -> Soteria__Sym_states__Pmap.Direct_access_patricia_tree(Rustsymex)(Soteria_rust_lib__Sptr.DecayMap.MapKey)(Soteria_rust_lib__Sptr.DecayMap.EntryExcl).t option -> ((Borrows.State.t option, [> `OutOfBounds ] as 'a, [< `State of Borrows.State.syn | `Structure of Borrows.Tree.syn ] list) Soteria.Symex.Compo_res.t * Soteria__Sym_states__Pmap.Direct_access_patricia_tree(Rustsymex)(Soteria_rust_lib__Sptr.DecayMap.MapKey)(Soteria_rust_lib__Sptr.DecayMap.EntryExcl).t option) Rustsymex.t) -> (unit, 'a, syn list) SM.Result.t

val unprotect : Soteria_rust_lib.Typed.T.sint Typed.t -> Soteria_rust_lib.Typed.T.sint Typed.t -> Borrows.Tag.t -> Borrows.Tree.t option -> (unit, [> `OutOfBounds ], syn list) SM.Result.t

val tb_access : Soteria_rust_lib.Typed.T.sint Typed.t -> Soteria_rust_lib.Typed.T.sint Typed.t -> Borrows.Tag.t -> Borrows.Tree.t option -> (unit, [> `AliasingError | `OutOfBounds ], syn list) SM.Result.t