Module S_int.T

type sint = [

| `NonZero

| `Zero

]

A symbolic integer; can either be `NonZero if it is known to not be 0, `Zero if it is 0.

type sint_ovf = [

| `NonZero

| `Zero

| `Overflowed

]

Any symbolic integer; it may be the result of an overflowing operation

type nonzero = [

| `NonZero

]

A symbolic integer known to be non-zero.

type zero = [

| `Zero

]

A symbolic integer known to be zero.

type sfloat = [

| `Float

]

type sbool = [

| `Bool

]

type sptr = [

| `Ptr

]

type sloc = [

| `Loc

]

type 'a sseq = [

| `List of 'a

]

type cval = [

| sint

| sptr

| sfloat

]

val pp_sint : Stdlib.Format.formatter -> sint -> unit

val pp_sint_ovf : Stdlib.Format.formatter -> sint_ovf -> unit

val pp_nonzero : Stdlib.Format.formatter -> nonzero -> unit

val pp_zero : Stdlib.Format.formatter -> zero -> unit

val pp_sfloat : Stdlib.Format.formatter -> sfloat -> unit

val pp_sbool : Stdlib.Format.formatter -> sbool -> unit

val pp_sptr : Stdlib.Format.formatter -> sptr -> unit

val pp_sloc : Stdlib.Format.formatter -> sloc -> unit

val pp_cval : Stdlib.Format.formatter -> cval -> unit

val pp_sseq : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a sseq -> unit

val hash_sint : sint -> int

val hash_sint_ovf : sint_ovf -> int

val hash_nonzero : nonzero -> int

val hash_zero : zero -> int

val hash_sfloat : sfloat -> int

val hash_sbool : sbool -> int

val hash_sptr : sptr -> int

val hash_sloc : sloc -> int

val hash_cval : cval -> int

val hash_sseq : 'a sseq -> int

val hash_any : {any}10/shadowed/(f0c29383912ef003ee744f4fcb6e86f5) -> int

type sptr_f = [

| `FullPtr

]

type sptr_t = [

| `ThinPtr

]

type tuple = [

| `Tuple

]

type enum = [

| `Enum

]

type union = [

| `Union

]

type poly = [

| `Poly

]

type ptr_meta = [

| sptr_t

| sint

]

type any = [

| sint

| sfloat

| sptr_f

| tuple

| enum

| union

| poly

]

The values that are allowed within the Rust interpreter as standalone values.

val pp_sptr_f : Stdlib.Format.formatter -> sptr_f -> unit

val pp_sptr_t : Stdlib.Format.formatter -> sptr_t -> unit

val pp_tuple : Stdlib.Format.formatter -> tuple -> unit

val pp_enum : Stdlib.Format.formatter -> enum -> unit

val pp_union : Stdlib.Format.formatter -> union -> unit

val pp_poly : Stdlib.Format.formatter -> poly -> unit

val pp_any : Stdlib.Format.formatter -> any -> unit