Module Soteria_rust_lib.Builtins

module Core : sig ... end

Core operations, that may be used in several parts of the interpreter, so we share them here for convenience.

module Eval : sig ... end

This file handles stubbing functions in Rust. There are actually three types of functions we need to handle: intrinsics, externs, and stubbed user-functions.

module Extern : sig ... end

module Fixme : sig ... end

This file was generated with scripts/stubs.py -- do not edit it manually, instead modify the script and re-run it.

module Intrinsics : sig ... end

This file was generated with scripts/stubs.py -- do not edit it manually, instead modify the script and re-run it.

module Optim : sig ... end

This file was generated with scripts/stubs.py -- do not edit it manually, instead modify the script and re-run it.

module Soteria_lib : sig ... end

This file was generated with scripts/stubs.py -- do not edit it manually, instead modify the script and re-run it.

module System : sig ... end

This file was generated with scripts/stubs.py -- do not edit it manually, instead modify the script and re-run it.

module Tokio : sig ... end

This file was generated with scripts/stubs.py -- do not edit it manually, instead modify the script and re-run it.