Module StateM.Sptr

include module type of Sptr

module type DecayMapS = sig ... end

A map to store information on "decayed" pointers, namely mapping from locations to integers (their decayed) value.

module DecayMap : DecayMapS

module D_abstr : sig ... end

module Tag = Soteria_rust_lib.Tree_borrows.Ptr_tag

type ('sptr, 'snonzero, 'sint, 'tag) base = {

ptr : 'sptr;

tag : 'tag option;

align : 'snonzero;

size : 'sint;

}

type t = (Typed.T.sptr Typed.t, Typed.T.nonzero Typed.t, Typed.T.sint Typed.t, Tag.t) base

type syn = (Typed.Expr.t, Typed.Expr.t, Typed.Expr.t, Tag.t) base

val pp' : (Stdlib.Format.formatter -> 'a -> unit) -> 'b Fmt.t -> Stdlib.Format.formatter -> ('a, 'c, 'd, 'b) base -> unit

val pp : Stdlib.Format.formatter -> ('a Typed.t, 'b, 'c, Tag.t) base -> unit

val show : ('a Typed.t, 'b, 'c, Tag.t) base -> string

val pp_syn : Stdlib.Format.formatter -> (Typed.Expr.t, 'a, 'b, Tag.t) base -> Ppx_deriving_runtime.unit

val show_syn : (Typed.Expr.t, 'a, 'b, Tag.t) base -> string

val to_syn : ('a Soteria_rust_lib__Typed.t, 'b Soteria_rust_lib__Typed.t, 'c Soteria_rust_lib__Typed.t, 'd) base -> (Typed.Expr.t, Typed.Expr.t, Typed.Expr.t, 'd) base

val in_bound : ([< Soteria_rust_lib__Typed.T.sptr ] Soteria_rust_lib__Typed.t, 'a, [< Soteria_rust_lib__Typed.T.sint ] Soteria_rust_lib__Typed.t, 'b) base -> [> Soteria_rust_lib__Typed.sbool ] Soteria_rust_lib__Typed.t

Returns a symbolic boolean characterising whether the pointer is in bound to its allocation.

val learn_eq : (DecayMap.SM.Value.Expr.t, DecayMap.SM.Value.Expr.t, DecayMap.SM.Value.Expr.t, Tag.t) base -> ('a DecayMap.SM.Value.t, 'b DecayMap.SM.Value.t, 'c DecayMap.SM.Value.t, Tag.t) base -> (unit, 'd) DecayMap.SM.Consumer.t

val exprs_syn : ('a, 'a, 'a, 'b) base -> 'a list

val fresh : unit -> 'a

val subst : (Typed.Expr.t -> 'a Soteria_rust_lib__Typed.t) -> (Typed.Expr.t, Typed.Expr.t, Typed.Expr.t, 'b) base -> ('c Soteria_rust_lib__Typed.t, 'd Soteria_rust_lib__Typed.t, 'e Soteria_rust_lib__Typed.t, 'b) base

val null : unit -> ([> Soteria_rust_lib__Typed.T.sptr ] Soteria_rust_lib__Typed.t, [> Soteria_rust_lib__Typed.T.nonzero ] Soteria_rust_lib__Typed.t, [> Soteria_rust_lib__Typed.T.sint ] Soteria_rust_lib__Typed.t, 'a) base

The null pointer, which always decays to 0, and has no provenance. Equivalent to of_address 0.

val of_address : [< Soteria_rust_lib__Typed.T.sint ] Soteria_rust_lib__Typed.t -> ([> Soteria_rust_lib__Typed.T.sptr ] Soteria_rust_lib__Typed.t, [> Soteria_rust_lib__Typed.T.nonzero ] Soteria_rust_lib__Typed.t, [> Soteria_rust_lib__Typed.T.sint ] Soteria_rust_lib__Typed.t, 'a) base

Converts an address into a pointer, without provenance.

val is_null : ([< Soteria_rust_lib__Typed.T.sptr ] Soteria_rust_lib__Typed.t, 'a, 'b, 'c) base -> [> Soteria_rust_lib__Typed.sbool ] Soteria_rust_lib__Typed.t

Whether this is the null pointer, meaning it always decays to 0.

val ofs : ([< Soteria_rust_lib__Typed.T.sptr ] Soteria_rust_lib__Typed.t, 'a, 'b, 'c) base -> [> Soteria_rust_lib__Typed.T.sint ] Soteria_rust_lib__Typed.t

The offset of this pointer within its allocation.

val has_provenance : ([< Soteria_rust_lib__Typed.T.sptr ] Soteria_rust_lib__Typed.t, 'a, 'b, 'c) base -> [> Typed.sbool ] Typed.t

Whether this pointer has provenance, i.e. points to some allocation.

val have_same_provenance : ([< Soteria_rust_lib__Typed.T.sptr ] Soteria_rust_lib__Typed.t, 'a, 'b, 'c) base -> ([< Soteria_rust_lib__Typed.T.sptr ] Soteria_rust_lib__Typed.t, 'd, 'e, 'f) base -> [> Soteria_rust_lib__Typed.sbool ] Soteria_rust_lib__Typed.t

If these two pointers have the same provenance, i.e. point to the same allocation (or if they both have no provenance).

val raw_offset : ([< Soteria_rust_lib__Typed.T.sptr ] Soteria_rust_lib__Typed.t, 'a, 'b, 'c) base -> [< Soteria_rust_lib__Typed.T.sint ] Soteria_rust_lib__Typed.t -> Soteria_rust_lib__Rustsymex.MonadState.t -> ((([> Soteria_rust_lib__Typed.T.sptr ] Soteria_rust_lib__Typed.t, 'a, 'b, 'c) base, [> `UBDanglingPointer ], 'd) Soteria.Soteria_std.Compo_res.t * Soteria_rust_lib__Rustsymex.MonadState.t) Soteria__Symex.Make(Soteria.Bv_values.Bv_solver.Z3_solver(Typed)).t

A simplified (and unsafe) version of offset, that adds a signed bitvector to this pointer's offset.

val _decay : expose:bool -> ([< Soteria_rust_lib__Typed.T.sptr ] Soteria_rust_lib__Typed.t, [< Typed.T.nonzero ] Typed.t, [< Typed.T.sint ] Typed.t, 'a) base -> [> Soteria_rust_lib__Typed.T.sint ] Soteria_rust_lib__Typed.t DecayMap.SM.t

val as_id : ([< Soteria_rust_lib__Typed.T.sptr ] Soteria_rust_lib__Typed.t, 'a, 'b, 'c) base -> 'd Typed.t

For Miri: the allocation ID of this location, as a u64

val allocation_info : ('a, 'b Typed.t, 'c Typed.t, 'd) base -> 'e Typed.t * 'f Typed.t

For Miri: get the allocation info for this pointer: its size and alignment

val nondet : Charon.Types.ty -> Soteria_rust_lib__Rustsymex.MonadState.t -> ((([> Soteria_rust_lib__Typed.T.sptr ] Soteria_rust_lib__Typed.t, Typed.T.nonzero Typed.t, Typed.T.sint Typed.t, 'a) base, [> `InvalidLayout of Charon.Types.ty ], 'b) Soteria.Soteria_std.Compo_res.t * Soteria_rust_lib__Rustsymex.MonadState.t) Soteria__Symex.Make(Soteria.Bv_values.Bv_solver.Z3_solver(Typed)).t

Generates a nondeterministic pointer, that is valid for accesses to the given type. The location of the pointer is nondeterministic; it could point to any allocation.

val offset : ?check_signed:bool -> ?ty:Charon.Types.ty -> [< Typed.T.sint ] Typed.t -> t -> (t, 'env) monad

val check_aligned : Rust_val.full_ptr -> Charon.Types.ty -> (unit, 'env) monad

val check_non_dangling : Rust_val.full_ptr -> Charon.Types.ty -> (unit, 'env) monad

val check_non_dangling_untyped : t -> Typed.T.sint Typed.t -> (unit, 'env) monad

val distance : t -> t -> (Typed.T.sint Typed.t, 'env) monad

val decay : t -> (Typed.T.sint Typed.t, 'env) monad

val expose : t -> (Typed.T.sint Typed.t, 'env) monad

val dangling_if_zst : Charon.Types.ty -> (t option, 'env) monad