Module StateM.Value_codec

include module type of Value_codec

module DecayMap = Sptr.DecayMap

val variant_for_discr : 'a Soteria_rust_lib__Typed.t -> Charon.Types.type_decl_ref -> Soteria_rust_lib__Rustsymex.MonadState.t -> (Charon.Types.variant * Soteria_rust_lib__Rustsymex.MonadState.t) Soteria__Symex.Make(Soteria.Bv_values.Bv_solver.Z3_solver(Typed)).t

Returns the variant id and variant matching the given discriminant.

val iter_fields : ?variant:'a -> ?meta:'a -> Layout.t -> Charon.Types.ty -> (Charon.Types.ty * Typed.T.sint Typed.t) Soteria.Soteria_std.Iter.t

Iterator over the fields and offsets of a type; for primitive types, returns a singleton iterator for that value.

val unwrap_box : Rust_val.t -> Rust_val.full_ptr * Rust_val.t * Rust_val.t

Given a value of type Box<T, A>, returns as a triple the inner pointer, the allocator object, and the marker value.

val ptr_of_box : Rust_val.t -> Rust_val.full_ptr

Given a value of type Box<T, A>, returns the container pointer

val mk_box : Rust_val.t -> Rust_val.t -> Rust_val.t -> Rust_val.t

Given a pointer, allocator and marker, returns a value of type Box<T, A>

module Decoder (State_tys : sig ... end) : sig ... end

val validity : ?check_ref:??? -> Charon.Types.ty -> Rust_val.t -> ([ `Bool ] Typed.t -> [> `RefToUninhabited of Charon.Types.ty | `UBDanglingPointer | `UBTransmute of string ] -> (unit, 'a, 'b) Rustsymex.Result.t) -> (unit, 'a, 'b) Rustsymex.Result.t

Iterates over the validity constraints of this particular value for a given type, traversing it recursively. For every requirement, this associates to it the error to be raised if the requirement is not met.

An optional check_refs function can be provided, to further check the validity of references and boxes; this allows checking that references are not dangling, well-aligned, and that their pointees are valid. Note that this check will raise errors in the outside monad, rather than in the returned list, since it is not possible to represent constraints in the state in first order logic. We provide this possibility here, to avoid re-implementing value traversal elsewhere.

Note that this function doesn't (and can't) check basic validity requirements of references and boxes, even alignment, as e.g. for a &dyn Trait the alignment cannot be known without some auxiliary state.

This doesn't check:

the fact the bytes of the value cannot be undefined, as that is checked when the memory is read; once we have a Rust_val.t, we are guaranteed the data is initialised.

the validity of the discriminant of an enum, as that is checked when the memory is read (otherwise it would be impossible to even decode the value in the first place).

Reference: https://doc.rust-lang.org/reference/behavior-considered-undefined.html#r-undefined.validity

val check_validity : check_ref: ((Sptr.t * (Typed.T.sint Typed.t, (Typed.T.sptr Typed.t, Typed.T.nonzero Typed.t, Typed.T.sint Typed.t, Soteria_rust_lib.Sptr.Tag.t) Sptr.base) Rust_val.meta_raw) -> Charon.Types.ty -> 'a -> Soteria_rust_lib__Rustsymex.MonadState.t -> (((unit, [> `InvalidLayout of Charon.Types.ty | `RefToUninhabited of Charon.Types.ty | `UBDanglingPointer | `UBTransmute of string ] as 'c, 'b) Soteria.Soteria_std.Compo_res.t * 'd) * Soteria_rust_lib__Rustsymex.MonadState.t) Soteria__Symex.Make(Soteria.Bv_values.Bv_solver.Z3_solver(Typed)).t) -> Charon.Types.ty -> Rust_val.t -> 'a -> Soteria_rust_lib__Rustsymex.MonadState.t -> (((unit, 'c, 'b) Soteria.Soteria_std.Compo_res.t * 'a) * Soteria_rust_lib__Rustsymex.MonadState.t) Soteria__Symex.Make(Soteria.Bv_values.Bv_solver.Z3_solver(Typed)).t

Applies a validity check for a value, given a check_ref state monad operation. This assumes check_ref is effect-free: the returned state is the input state. See validity.

val float_to_bv_bits : [< Soteria_rust_lib__Typed.T.sfloat ] Typed.t -> [ `NonZero | `Zero ] Typed.t DecayMap.SM.t

Converts a floating value to a bitvector, preserving it's bit representation. This is a symbolic process, because SMT-Lib has no operation for "float->bv" that preserves the bits, due to NaN.

See https://smt-lib.org/theories-FloatingPoint.shtml, "Conversions to other sorts"

val transmute_one : to_ty:Charon.Types.ty -> Rust_val.rust_val -> Rust_val.rust_val DecayMap.SM.t

Transmutes a singular rust value, without splitting. This is under the assumption that size_of to_ty = size_of v, and both are primitives (literal or pointer).

val nondet_raw : Charon.Types.ty -> (Rust_val.rust_val, [> `InvalidLayout of Charon.Types.ty ], 'f) Rustsymex.Result.t

nondet_raw ty returns a nondeterministic value for ty, by traversing ty: the returned value will have the right structure, and any required nondet variable will have been created. Importantly, the returned value may not uphold the validity invariant of ty*. To ensure the value is also valid, use validity, assume-ing the constraints it returns.

* Much like in validity, this function actually assumes two validity invariants: the data is initialised, and the discriminant of enums corresponds to that a variant.

val size_and_align_of_val : load_vtable: ([> `Align | `Size ] -> 'a -> Soteria_rust_lib__Rustsymex.MonadState.t -> ((Rust_val.t, [> `InvalidLayout of Charon.Types.ty | `Overflow ] as 'b, 'c) Soteria.Soteria_std.Compo_res.t * Soteria_rust_lib__Rustsymex.MonadState.t) Soteria__Symex.Make(Soteria.Bv_values.Bv_solver.Z3_solver(Typed)).t) -> t:Charon.Types.ty -> meta:('d Typed.t, 'a) Rust_val.meta_raw -> Soteria_rust_lib__Rustsymex.MonadState.t -> ((Typed.T.sint Typed.t * Typed.T.nonzero Typed.t, 'b, 'c) Soteria.Soteria_std.Compo_res.t * Soteria_rust_lib__Rustsymex.MonadState.t) Soteria__Symex.Make(Soteria.Bv_values.Bv_solver.Z3_solver(Typed)).t

Calculates the size and alignment of a type t, according to the pointer metadata meta. Receives an arbitrary state and load function, to possibly access a heap to get VTable information.

val encode : offset:Typed.T.sint Typed.t -> Rust_val.t -> Charon.Types.ty -> ((Rust_val.t * Typed.T.sint Typed.t * Typed.T.nonzero Typed.t) Soteria.Soteria_std.Iter.t, 'env) monad

val cast_literal : from_ty:Charon.Values.literal_type -> to_ty:Charon.Values.literal_type -> [< Typed.T.cval ] Typed.t -> Rust_val.t

val nondet_valid : Charon.Types.ty -> (Rust_val.t, 'env) monad

val ref_tys_in : f: ('acc -> Charon.Types.ty -> Rust_val.full_ptr -> (Rust_val.full_ptr * 'acc, 'env) monad) -> init:'acc -> Charon.Types.ty -> Rust_val.t -> (Rust_val.t * 'acc, 'env) monad