Module StateM.Sptr

include module type of Sptr

module type DecayMapS = sig ... end

A map to store information on "decayed" pointers, namely mapping from locations to integers (their decayed) value.

module DecayMap : DecayMapS

type t = Svalue.Typed.T.sptr_t Svalue.Typed.t

val pp : Stdlib.Format.formatter -> 'a Svalue.Typed.t -> unit

val raw_offset : [< Soteria_rust_lib__Svalue__Typed.T.sptr_t ] Soteria_rust_lib__Svalue__Typed.t -> [< `NonZero | `Zero ] Soteria_rust_lib__Svalue__Typed.t -> Soteria_rust_lib__Rustsymex.MonadState.t -> (([> Soteria_rust_lib__Svalue__Typed.T.sptr_t ] Soteria_rust_lib__Svalue__Typed.t, [> `UBDanglingPointer ], 'a) Soteria.Soteria_std.Compo_res.t * Soteria_rust_lib__Rustsymex.MonadState.t) Soteria__Symex.Make(Soteria.Bv_values.Bv_solver.Z3_solver(Svalue.Typed)).t

A simplified, untyped (and unsafe) version of offset, that adds a signed integer to this pointer's offset. This offset doesn't check whether the resulting pointer is dangling after being offset.

val _decay : expose:bool -> [< Soteria_rust_lib__Svalue__Typed.T.sptr_t ] Soteria_rust_lib__Svalue__Typed.t -> [> `NonZero | `Zero ] Soteria_rust_lib__Svalue__Typed.t DecayMap.SM.t

val offset : ?check_signed:bool -> ?ty:Charon.Types.ty -> [< Svalue.Typed.T.sint ] v -> t -> (t, 'env) monad

val check_aligned : [< Svalue.Typed.T.sptr_f ] v -> Charon.Types.ty -> (unit, 'env) monad

val check_non_dangling : [< Svalue.Typed.T.sptr_f ] v -> Charon.Types.ty -> (unit, 'env) monad

val check_non_dangling_untyped : [< Svalue.Typed.T.sptr_t ] v -> Svalue.Typed.T.sint v -> (unit, 'env) monad

val distance : t -> t -> (Svalue.Typed.T.sint v, 'env) monad

val decay : t -> (Svalue.Typed.T.sint v, 'env) monad

val expose : t -> (Svalue.Typed.T.sint v, 'env) monad

val dangling_if_zst : Charon.Types.ty -> (t option, 'env) monad