Module Std.M

Parameters

module StateM : State.StateM.S

Signature

val panic_cleanup : 'a -> (('b, 'c, StateM.Sptr.t) Rust_val.raw, 'd) StateM.t

Stub for __rust_panic_cleanup: returns a dummy &dyn Any pointer so the caller does not crash before we handle the panic.

val panic_impl : 'a -> ('b, 'c) StateM.t

This shoyld be a call that gets resolved to the `#panic_handler` function. For now, we just panic.

val fn_to_stub : fn -> 'a -> (('b, 'c, StateM.Sptr.t) Rust_val.raw, 'd) StateM.t