Module System.M

Parameters

module StateM : State.StateM.S

Signature

val tlv_atexit : fun_exec: (Common.Fun_kind.t -> (Typed.T.sint Typed.t, 'a Typed.t, 'b Typed.t, (Typed.T.sptr Typed.t, Typed.T.nonzero Typed.t, Typed.T.sint Typed.t, Soteria_rust_lib.Sptr.Tag.t) Sptr.base) Rust_val.raw list -> ('c, unit) StateM.t) -> (Typed.T.sint Typed.t, 'a Typed.t, 'b Typed.t, (Typed.T.sptr Typed.t, Typed.T.nonzero Typed.t, Typed.T.sint Typed.t, Soteria_rust_lib.Sptr.Tag.t) Sptr.base) Rust_val.raw list -> (('d, 'e, 'f, 'g) Rust_val.raw, 'h) StateM.t

Used on macOS to register thread local destructors; receives a function pointer and an argument. Should call the destructor with the argument at the end of the thread.

_tlv_atexit(dtor: unsafe extern "C" fn( *mut u8), arg: *mut u8)

val cxa_thread_atexit_impl : fun_sig:Charon.Types.fun_sig -> 'a -> (([> Soteria_rust_lib__Typed.T.sint ] Soteria_rust_lib__Typed.t, 'b, 'c, 'd) Rust_val.raw, 'e) StateM.t

__cxa_thread_atexit_impl: Option<extern "C" fn(...)> is a weak symbol we don't link against, so it reads as None; the destructor registration then takes its fallback (`register_dtor_fallback`) path.

val dso_handle : 'a -> (('b, 'c, 'd, ([> Soteria_rust_lib__Typed.T.sptr ] Soteria_rust_lib__Typed.t, [> Soteria_rust_lib__Typed.T.nonzero ] Soteria_rust_lib__Typed.t, [> Soteria_rust_lib__Typed.T.sint ] Soteria_rust_lib__Typed.t, 'e) StateM.Sptr.base) Rust_val.raw, 'f) StateM.t

__dso_handle: *mut u8 is likewise an unlinked weak symbol, i.e. null.

val pthread_key_create : fun_exec: (Common.Fun_kind.t -> ('a, 'b, 'c, ([> Soteria_rust_lib__Typed.T.sptr ] Soteria_rust_lib__Typed.t, [> Soteria_rust_lib__Typed.T.nonzero ] Soteria_rust_lib__Typed.t, [> Soteria_rust_lib__Typed.T.sint ] Soteria_rust_lib__Typed.t, 'd) StateM.Sptr.base) Rust_val.raw list -> ('e, unit) StateM.t) -> (Typed.T.sint Typed.t, 'f, 'g, Sptr.t) Rust_val.raw list -> (([> Soteria_rust_lib__Typed.T.sint ] Soteria_rust_lib__Typed.t, 'h, 'i, 'j) Rust_val.raw, 'k) StateM.t

pthread_key_create takes a key out-pointer and an optional destructor. We don't model pthread TLS storage, so any non-sentinel key works; std asserts the created key isn't KEY_SENTVAL (0), so we write 1. The key's destructor (which ignores its argument and just drains the thread-local destructor list) is registered to run at thread exit.

val pthread_key_noop : 'a -> (([> Soteria_rust_lib__Typed.T.sint ] Soteria_rust_lib__Typed.t, 'b, 'c, 'd) Rust_val.raw, 'e) StateM.t

pthread_setspecific/pthread_key_delete only manage the (unmodelled) TLS storage and marker; returning success is enough.

val fn_to_stub : Charon.Types.fun_sig -> (Common.Fun_kind.t -> (Typed.T.sint Typed.t, 'a Typed.t, 'b Typed.t, (Typed.T.sptr Typed.t, Typed.T.nonzero Typed.t, Typed.T.sint Typed.t, Soteria_rust_lib.Sptr.Tag.t) StateM.Sptr.base) Rust_val.raw list -> ('c, unit) StateM.t) -> fn -> (Typed.T.sint Typed.t, 'a Typed.t, 'b Typed.t, (Typed.T.sptr Typed.t, Typed.T.nonzero Typed.t, Typed.T.sint Typed.t, Soteria_rust_lib.Sptr.Tag.t) Sptr.base) Rust_val.raw list -> (([> Soteria_rust_lib__Typed.T.sint ] Soteria_rust_lib__Typed.t, 'd, 'e, ([> Soteria_rust_lib__Typed.T.sptr ] Soteria_rust_lib__Typed.t, [> Soteria_rust_lib__Typed.T.nonzero ] Soteria_rust_lib__Typed.t, [> Soteria_rust_lib__Typed.T.sint ] Soteria_rust_lib__Typed.t, 'f) StateM.Sptr.base) Rust_val.raw, 'g) StateM.t