Module Intrinsics_impl.M

Parameters

module State : State_intf.S

Signature

module Core : sig ... end

module Encoder : sig ... end

module Monad : sig ... end

type 'a ret = ('a, unit) Monad.t

val as_ptr : Monad.rust_val -> Monad.Sptr.t Rust_val.full_ptr

val as_base : Charon.Types.literal_type -> Monad.rust_val -> [< Soteria_rust_lib.Typed.T.any NonZero Zero ] Typed.t

val as_base_i : Charon.Values.u_int_ty -> Monad.rust_val -> [< Soteria_rust_lib.Typed.T.any NonZero Zero ] Typed.t

val as_base_f : Charon.Values.float_type -> Monad.rust_val -> [< Soteria_rust_lib.Typed.T.any Float ] Typed.t

val abort : unit ret

val checked_op : Charon.Expressions.binop -> t:Charon.Types.ty -> x:Monad.rust_val -> y:Monad.rust_val -> Monad.rust_val ret

val add_with_overflow : t:Charon.Types.ty -> x:Monad.rust_val -> y:Monad.rust_val -> Monad.rust_val ret

val sub_with_overflow : t:Charon.Types.ty -> x:Monad.rust_val -> y:Monad.rust_val -> Monad.rust_val ret

val mul_with_overflow : t:Charon.Types.ty -> x:Monad.rust_val -> y:Monad.rust_val -> Monad.rust_val ret

val align_of : t:Charon.Types.ty -> ([> Soteria__Bv_values__Typed.T.nonzero ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val align_of_val : t:Charon.Types.ty -> ptr:('a * Monad.Sptr.t Rust_val.meta) -> ([< Soteria_rust_lib.Typed.T.any NonZero Zero ] Typed.t, 'b) Monad.t

val arith_offset : t:Charon.Types.ty -> dst:(Monad.Sptr.t * 'a) -> offset:[< Soteria_rust_lib.Typed.T.sint ] Typed.t -> (Monad.Sptr.t * 'a, 'b) Monad.t

val offset : ptr:Charon.Types.ty -> delta:Charon.Types.ty -> dst:Monad.Sptr.t Rust_val.t -> offset:Monad.rust_val -> (Monad.Sptr.t Rust_val.t, 'a) Monad.t

val assert_inhabited : t:Charon.Types.ty -> (unit, 'a) Monad.t

val assert_mem_uninitialized_valid : t:'a -> (unit, 'b) Monad.t

val assert_zero_valid : t:Charon.Types.ty -> (unit, 'a) Monad.t

val assume : b:[< Soteria_rust_lib.Typed.T.sbool ] Soteria__Bv_values__Typed.t -> 'a -> Monad.State.t -> (unit * 'a * Monad.State.t, [< Error.t StdErr ] Monad.State.err * Monad.State.t, Monad.State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.t

val black_box : t:'a -> dummy:'b -> ('b, 'c) Monad.t

val breakpoint : unit ret

val bitreverse : t:Charon.Types.ty -> x:Monad.rust_val -> ('a Rust_val.t, 'b) Monad.t

val bswap : t:Charon.Types.ty -> x:Monad.rust_val -> ('a Rust_val.t, 'b) Monad.t

val caller_location : Monad.full_ptr ret

val carrying_mul_add : t:Charon.Types.ty -> u:Charon.Types.ty -> multiplier:Monad.rust_val -> multiplicand:Monad.rust_val -> addend:Monad.rust_val -> carry:Monad.rust_val -> ('a Rust_val.t, 'b) Monad.t

val catch_unwind : (Charon.UllbcAst.blocks Charon.UllbcAst.gfun_decl -> Monad.Sptr.t Rust_val.t list -> ('a, 'b) Monad.t) -> _try_fn:Monad.rust_val -> _data:Monad.Sptr.t Rust_val.full_ptr -> _catch_fn:Monad.rust_val -> 'b -> Monad.State.t -> ([> Soteria__Bv_values__Typed.T.sint ] Soteria__Bv_values__Typed.t * 'b * State.t, Error.t Monad.State.err * Monad.State.t, Monad.State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.t

val float_rounding : Soteria.Bv_values.Svalue.RoundingMode.t -> [< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val ceilf16 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val ceilf32 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val ceilf64 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val ceilf128 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val floorf16 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val floorf32 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val floorf64 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val floorf128 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val round_ties_even_f16 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val round_ties_even_f32 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val round_ties_even_f64 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val round_ties_even_f128 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val roundf16 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val roundf32 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val roundf64 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val roundf128 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val truncf16 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val truncf32 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val truncf64 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val truncf128 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val cold_path : unit ret

val compare_bytes : left:(Monad.Sptr.t * 'a) -> right:(Monad.Sptr.t * 'b) -> bytes:[< Soteria_rust_lib.Rust_val.T.sint ] Soteria__Bv_values__Typed.t -> 'c -> Monad.State.t -> ([> Soteria__Bv_values__Typed.T.sint ] Soteria__Bv_values__Typed.t * 'c * State.t, Error.t Monad.State.err * Monad.State.t, Monad.State.serialized) Soteria.Symex.Compo_res.t Rustsymex.t

val check_overlap : string -> Monad.Sptr.t -> Monad.Sptr.t -> [< Soteria_rust_lib.Typed.T.sint ] Typed.t -> 'a -> Monad.State.t -> (unit * 'a * Monad.State.t, [< Error.t StdErr UBDanglingPointer ] Monad.State.err * Monad.State.t, Monad.State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.t

check_overlap name l r size ensures the pointers l and r do not overlap for a range of size size; otherwise errors, with `StdErr (name ^ " overlapped").

val copy_ : bool -> t:Charon.Types.ty -> src:Monad.full_ptr -> dst:Monad.full_ptr -> count:[< Soteria__Bv_values__Typed.T.sint ] Typed.t -> unit ret

val copy : t:Charon.Types.ty -> src:Monad.full_ptr -> dst:Monad.full_ptr -> count:[< Soteria__Bv_values__Typed.T.sint ] Typed.t -> unit ret

val copy_nonoverlapping : t:Charon.Types.ty -> src:Monad.full_ptr -> dst:Monad.full_ptr -> count:[< Soteria__Bv_values__Typed.T.sint ] Typed.t -> unit ret

val typed_swap_nonoverlapping : t:Charon.Types.ty -> x:(Monad.State.Sptr.t * Monad.State.Sptr.t Rust_val.meta) -> y:(Monad.State.Sptr.t * Monad.State.Sptr.t Rust_val.meta) -> 'a -> Monad.State.t -> (unit * 'a * Monad.State.t, Error.t Monad.State.err * Monad.State.t, Monad.State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.t

val copy_sign : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> y:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> 'a -> State.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t * 'a * State.t, Error.t State.err * State.t, State.serialized) Soteria.Symex.Compo_res.t Rustsymex.t

val copysignf128 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> y:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> 'a -> State.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t * 'a * State.t, Error.t State.err * State.t, State.serialized) Soteria.Symex.Compo_res.t Rustsymex.t

val copysignf64 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> y:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> 'a -> State.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t * 'a * State.t, Error.t State.err * State.t, State.serialized) Soteria.Symex.Compo_res.t Rustsymex.t

val copysignf32 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> y:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> 'a -> State.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t * 'a * State.t, Error.t State.err * State.t, State.serialized) Soteria.Symex.Compo_res.t Rustsymex.t

val copysignf16 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> y:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> 'a -> State.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t * 'a * State.t, Error.t State.err * State.t, State.serialized) Soteria.Symex.Compo_res.t Rustsymex.t

val binary_int_operation : concrete:(int -> Z.t -> Soteria_rust_lib.Rust_val.T.sint Typed.t) -> symbolic: (int -> [< Soteria_rust_lib.Typed.T.any NonZero Zero ] Typed.t -> Soteria_rust_lib.Rust_val.T.sint Typed.t) -> t:Charon.Types.ty -> x:Monad.rust_val -> Soteria_rust_lib.Rust_val.T.sint Typed.t ret

Applies either concrete or symbolic to a bitvector.

val ctpop : t:Charon.Types.ty -> x:Monad.rust_val -> Soteria_rust_lib.Rust_val.T.sint Typed.t ret

val cttz : t:Charon.Types.ty -> x:Monad.rust_val -> Soteria_rust_lib.Rust_val.T.sint Typed.t ret

val cttz_nonzero : t:Charon.Types.ty -> x:Monad.rust_val -> unit -> Monad.State.t -> (Soteria_rust_lib.Rust_val.T.sint Typed.t * unit * State.t, Error.t Monad.State.err * Monad.State.t, Monad.State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.t

val ctlz : t:Charon.Types.ty -> x:Monad.rust_val -> Soteria_rust_lib.Rust_val.T.sint Typed.t ret

val ctlz_nonzero : t:Charon.Types.ty -> x:Monad.rust_val -> unit -> Monad.State.t -> (Soteria_rust_lib.Rust_val.T.sint Typed.t * unit * State.t, Error.t Monad.State.err * Monad.State.t, Monad.State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.t

val discriminant_value : t:Charon.Types.ty -> v:Monad.State.Sptr.t Rust_val.full_ptr -> ('a Rust_val.t, 'b) Monad.t

val disjoint_bitor : t:Charon.Types.ty -> a:Monad.rust_val -> b:Monad.rust_val -> ('a Rust_val.t, 'b) Monad.t

val exact_div : t:Charon.Types.ty -> x:Monad.rust_val -> y:Monad.rust_val -> 'a -> Monad.State.t -> ('b Rust_val.t * 'a * State.t, Error.t Monad.State.err * Monad.State.t, State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.t

val abs : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val fabsf16 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val fabsf32 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val fabsf64 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val fabsf128 : x:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val float_fast : Charon.Expressions.binop -> t:Charon.Types.ty -> a:Monad.rust_val -> b:Monad.rust_val -> Monad.rust_val ret

val fadd_fast : t:Charon.Types.ty -> a:Monad.rust_val -> b:Monad.rust_val -> Monad.rust_val ret

val fdiv_fast : t:Charon.Types.ty -> a:Monad.rust_val -> b:Monad.rust_val -> Monad.rust_val ret

val fmul_fast : t:Charon.Types.ty -> a:Monad.rust_val -> b:Monad.rust_val -> Monad.rust_val ret

val frem_fast : t:Charon.Types.ty -> a:Monad.rust_val -> b:Monad.rust_val -> Monad.rust_val ret

val fsub_fast : t:Charon.Types.ty -> a:Monad.rust_val -> b:Monad.rust_val -> Monad.rust_val ret

val float_to_int_unchecked : float:Charon.Types.ty -> int:Charon.Types.ty -> value:Monad.rust_val -> 'a -> Monad.State.t -> ('b Rust_val.t * 'a * State.t, Error.t Monad.State.err * Monad.State.t, Monad.State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.t

val fmul_add : a:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> b:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> c:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val fmaf16 : a:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> b:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> c:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val fmaf32 : a:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> b:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> c:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val fmaf64 : a:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> b:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> c:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val fmaf128 : a:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> b:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> c:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val fmuladdf16 : a:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> b:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> c:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val fmuladdf32 : a:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> b:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> c:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val fmuladdf64 : a:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> b:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> c:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val fmuladdf128 : a:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> b:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> c:[< Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t -> ([> Soteria__Bv_values__Typed.T.sfloat ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val forget : t:'a -> arg:'b -> (unit, 'c) Monad.t

val is_val_statically_known : t:'a -> _arg:'b -> ([> Soteria_rust_lib.Typed.T.sbool ] Rustsymex.Value.t, 'c) Monad.t

val likely : b:[< Soteria_rust_lib.Rust_val.T.sbool ] Soteria__Bv_values__Typed.t -> (Soteria_rust_lib.Rust_val.T.sbool Typed.t, 'a) Monad.t

val unlikely : b:[< Soteria_rust_lib.Rust_val.T.sbool ] Soteria__Bv_values__Typed.t -> (Soteria_rust_lib.Rust_val.T.sbool Typed.t, 'a) Monad.t

val float_minmax : is_min:bool -> x:[< Soteria_rust_lib.Rust_val.T.sfloat ] Soteria__Bv_values__Typed.t -> y:[< Soteria_rust_lib.Rust_val.T.sfloat ] Soteria__Bv_values__Typed.t -> Soteria_rust_lib.Rust_val.T.sfloat Typed.t ret

val minnumf16 : x:[< Soteria_rust_lib.Rust_val.T.sfloat ] Soteria__Bv_values__Typed.t -> y:[< Soteria_rust_lib.Rust_val.T.sfloat ] Soteria__Bv_values__Typed.t -> Soteria_rust_lib.Rust_val.T.sfloat Typed.t ret

val minnumf32 : x:[< Soteria_rust_lib.Rust_val.T.sfloat ] Soteria__Bv_values__Typed.t -> y:[< Soteria_rust_lib.Rust_val.T.sfloat ] Soteria__Bv_values__Typed.t -> Soteria_rust_lib.Rust_val.T.sfloat Typed.t ret

val minnumf64 : x:[< Soteria_rust_lib.Rust_val.T.sfloat ] Soteria__Bv_values__Typed.t -> y:[< Soteria_rust_lib.Rust_val.T.sfloat ] Soteria__Bv_values__Typed.t -> Soteria_rust_lib.Rust_val.T.sfloat Typed.t ret

val minnumf128 : x:[< Soteria_rust_lib.Rust_val.T.sfloat ] Soteria__Bv_values__Typed.t -> y:[< Soteria_rust_lib.Rust_val.T.sfloat ] Soteria__Bv_values__Typed.t -> Soteria_rust_lib.Rust_val.T.sfloat Typed.t ret

val maxnumf16 : x:[< Soteria_rust_lib.Rust_val.T.sfloat ] Soteria__Bv_values__Typed.t -> y:[< Soteria_rust_lib.Rust_val.T.sfloat ] Soteria__Bv_values__Typed.t -> Soteria_rust_lib.Rust_val.T.sfloat Typed.t ret

val maxnumf32 : x:[< Soteria_rust_lib.Rust_val.T.sfloat ] Soteria__Bv_values__Typed.t -> y:[< Soteria_rust_lib.Rust_val.T.sfloat ] Soteria__Bv_values__Typed.t -> Soteria_rust_lib.Rust_val.T.sfloat Typed.t ret

val maxnumf64 : x:[< Soteria_rust_lib.Rust_val.T.sfloat ] Soteria__Bv_values__Typed.t -> y:[< Soteria_rust_lib.Rust_val.T.sfloat ] Soteria__Bv_values__Typed.t -> Soteria_rust_lib.Rust_val.T.sfloat Typed.t ret

val maxnumf128 : x:[< Soteria_rust_lib.Rust_val.T.sfloat ] Soteria__Bv_values__Typed.t -> y:[< Soteria_rust_lib.Rust_val.T.sfloat ] Soteria__Bv_values__Typed.t -> Soteria_rust_lib.Rust_val.T.sfloat Typed.t ret

val ptr_guaranteed_cmp : t:'a -> ptr:Core.Sptr.t Rust_val.full_ptr -> other:Core.Sptr.t Rust_val.full_ptr -> 'b -> Monad.State.t -> ([> Soteria__Bv_values__Typed.T.sint ] Typed.t * 'b * Monad.State.t, [< Error.t UBPointerComparison ] Monad.State.err * Monad.State.t, 'c) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.t

val ptr_offset_from_ : unsigned:bool -> t:Charon.Types.ty -> ptr:Monad.full_ptr -> base:Monad.full_ptr -> Soteria_rust_lib.Rust_val.T.sint Typed.t ret

val ptr_offset_from : t:Charon.Types.ty -> ptr:Monad.full_ptr -> base:Monad.full_ptr -> Soteria_rust_lib.Rust_val.T.sint Typed.t ret

val ptr_offset_from_unsigned : t:Charon.Types.ty -> ptr:Monad.full_ptr -> base:Monad.full_ptr -> Soteria_rust_lib.Rust_val.T.sint Typed.t ret

val raw_eq : t:Charon.Types.ty -> a:Monad.State.Sptr.t Rust_val.full_ptr -> b:Monad.State.Sptr.t Rust_val.full_ptr -> 'a -> Monad.State.t -> ([> Typed.sbool ] Typed.t * 'a * State.t, Error.t Monad.State.err * Monad.State.t, Monad.State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.t

val rotate_ : side:[ `Left | `Right ] -> t:Charon.Types.ty -> x:Monad.rust_val -> shift:[< Soteria__Bv_values__Typed.T.sint ] Soteria__Bv_values__Typed.t -> Monad.rust_val ret

val rotate_left : t:Charon.Types.ty -> x:Monad.rust_val -> shift:[< Soteria__Bv_values__Typed.T.sint ] Soteria__Bv_values__Typed.t -> Monad.rust_val ret

val rotate_right : t:Charon.Types.ty -> x:Monad.rust_val -> shift:[< Soteria__Bv_values__Typed.T.sint ] Soteria__Bv_values__Typed.t -> Monad.rust_val ret

val saturating : Charon.Expressions.binop -> t:Charon.Types.ty -> a:Monad.rust_val -> b:Monad.rust_val -> Monad.rust_val ret

val saturating_add : t:Charon.Types.ty -> a:Monad.rust_val -> b:Monad.rust_val -> Monad.rust_val ret

val saturating_sub : t:Charon.Types.ty -> a:Monad.rust_val -> b:Monad.rust_val -> Monad.rust_val ret

val size_of : t:Charon.Types.ty -> ([> Soteria__Bv_values__Typed.T.sint ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val size_of_val : t:Charon.Types.ty -> ptr:('a * Monad.Sptr.t Rust_val.meta) -> 'b -> State.t -> ([ `NonZero | `Zero ] Typed.t * 'b * State.t, Error.t State.err * State.t, State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.t

val transmute : t_src:Charon.Types.ty -> dst:Charon.Types.ty -> src:Encoder.rust_val -> 'a -> Monad.State.t -> (Encoder.rust_val * 'a * Monad.State.t, [< Error.t RefToUninhabited UBDanglingPointer UBTransmute UninitializedMemoryAccess ] Monad.State.err * Monad.State.t, 'b) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.t

val type_id : t:'a -> ('b Rust_val.t, 'c) Monad.t

val type_name : t:Charon.Types.ty -> 'a -> Monad.State.t -> (Monad.State.Sptr.t Rust_val.full_ptr * 'a * State.t, Error.t Monad.State.err * Monad.State.t, Monad.State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.t

val unchecked_op : Charon.Expressions.binop -> t:Charon.Types.ty -> x:Monad.rust_val -> y:Monad.rust_val -> Monad.rust_val ret

val unchecked_add : t:Charon.Types.ty -> x:Monad.rust_val -> y:Monad.rust_val -> Monad.rust_val ret

val unchecked_div : t:Charon.Types.ty -> x:Monad.rust_val -> y:Monad.rust_val -> Monad.rust_val ret

val unchecked_mul : t:Charon.Types.ty -> x:Monad.rust_val -> y:Monad.rust_val -> Monad.rust_val ret

val unchecked_rem : t:Charon.Types.ty -> x:Monad.rust_val -> y:Monad.rust_val -> Monad.rust_val ret

val unchecked_shl : t:Charon.Types.ty -> u:'a -> x:Monad.rust_val -> y:Monad.rust_val -> Monad.rust_val ret

val unchecked_shr : t:Charon.Types.ty -> u:'a -> x:Monad.rust_val -> y:Monad.rust_val -> Monad.rust_val ret

val unchecked_sub : t:Charon.Types.ty -> x:Monad.rust_val -> y:Monad.rust_val -> Monad.rust_val ret

val variant_count : t:Charon.Types.ty -> ([> Soteria__Bv_values__Typed.T.sint ] Soteria__Bv_values__Typed.t, 'a) Monad.t

val read_vtable : slot:int -> ptr:Monad.full_ptr -> Soteria_rust_lib.Rust_val.T.sint Typed.t ret

val vtable_align : ptr:Monad.full_ptr -> Soteria_rust_lib.Rust_val.T.sint Typed.t ret

val vtable_size : ptr:Monad.full_ptr -> Soteria_rust_lib.Rust_val.T.sint Typed.t ret

val wrapping_op : Charon.Expressions.binop -> t:Charon.Types.ty -> a:Monad.rust_val -> b:Monad.rust_val -> Monad.rust_val ret

val wrapping_add : t:Charon.Types.ty -> a:Monad.rust_val -> b:Monad.rust_val -> Monad.rust_val ret

val wrapping_mul : t:Charon.Types.ty -> a:Monad.rust_val -> b:Monad.rust_val -> Monad.rust_val ret

val wrapping_sub : t:Charon.Types.ty -> a:Monad.rust_val -> b:Monad.rust_val -> Monad.rust_val ret

val write_bytes : t:Charon.Types.ty -> dst:(Monad.State.Sptr.t * Monad.State.Sptr.t Rust_val.meta) -> val_:'a Typed.t -> count:[< Soteria__Bv_values__Typed.T.sint ] Soteria__Bv_values__Typed.t -> 'b -> Monad.State.t -> (unit * 'b * Monad.State.t, Error.t Monad.State.err * Monad.State.t, Monad.State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.t

val volatile_load : t:Charon.Types.ty -> src:Monad.State.Sptr.t Rust_val.full_ptr -> 'a -> Monad.State.t -> (Monad.State.Sptr.t Rust_val.rust_val * 'a * Monad.State.t, [> `AccessedFnPointer | `AliasingError | `MisalignedPointer of Soteria_rust_lib.Rust_val.T.nonzero Typed.t * Soteria_rust_lib.Rust_val.T.nonzero Typed.t * Soteria_rust_lib.Rust_val.T.sint Typed.t | `NullDereference | `OutOfBounds | `RefToUninhabited | `UBDanglingPointer | `UBIntToPointerNoProvenance of Soteria_rust_lib.Rust_val.T.sint Typed.t | `UBIntToPointerStrict | `UBTransmute of string | `UninitializedMemoryAccess | `UseAfterFree ] Monad.State.err * Monad.State.t, Monad.State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.t

val volatile_set_memory : t:Charon.Types.ty -> dst:(Monad.State.Sptr.t * Monad.State.Sptr.t Rust_val.meta) -> val_:'a Typed.t -> count:[< Soteria__Bv_values__Typed.T.sint ] Soteria__Bv_values__Typed.t -> 'b -> Monad.State.t -> (unit * 'b * Monad.State.t, Error.t Monad.State.err * Monad.State.t, Monad.State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.t

val volatile_copy_memory : t:Charon.Types.ty -> dst:Monad.full_ptr -> src:Monad.full_ptr -> count:[< Soteria__Bv_values__Typed.T.sint ] Typed.t -> unit ret

val volatile_copy_nonoverlapping_memory : t:Charon.Types.ty -> dst:Monad.full_ptr -> src:Monad.full_ptr -> count:[< Soteria__Bv_values__Typed.T.sint ] Typed.t -> unit ret

val volatile_store : t:Charon.Types.ty -> dst:Monad.State.Sptr.t Rust_val.full_ptr -> val_:Monad.State.Sptr.t Rust_val.rust_val -> 'a -> Monad.State.t -> (unit * 'a * Monad.State.t, [> `AccessedFnPointer | `AliasingError | `MisalignedPointer of Soteria_rust_lib.Rust_val.T.nonzero Typed.t * Soteria_rust_lib.Rust_val.T.nonzero Typed.t * Soteria_rust_lib.Rust_val.T.sint Typed.t | `NullDereference | `OutOfBounds | `RefToUninhabited | `UBDanglingPointer | `UBIntToPointerNoProvenance of Soteria_rust_lib.Rust_val.T.sint Typed.t | `UBIntToPointerStrict | `UBTransmute of string | `UninitializedMemoryAccess | `UseAfterFree ] Monad.State.err * Monad.State.t, Monad.State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.t