Module Soteria_rust_lib.Call_graph

Call graph for Soteria-Rust.

Records a directed edge from caller to callee on each function call in the interpreter. The graph is stored in a single global variable and is NOT concurrent-safe.

module FunNode : sig ... end

Node type: a fully-qualified function name rendered as a string.

module G : sig ... end

type Stdlib.Effect.t +=

| Add_edge : Charon.Types.name option * Charon.Types.name option -> unit Stdlib.Effect.t

val with_callgraph : unit -> (unit -> 'a) -> 'a * G.t

val add_edge : Charon.Types.name option -> Charon.Types.name option -> unit

Adds an edge to the global callgraph. Always needs to be called within a crate context!

val dump : string -> G.Node_set.t G.Hashtbl.t -> unit

Dump the call graph as a DOT file at path.

val with_dumped_callgraph : unit -> (unit -> 'a) -> 'a