Module Common.Charon_util

val size_of_int_ty : Charon.Values.int_ty -> int

val size_of_uint_ty : Charon.Values.u_int_ty -> int

val size_of_literal_ty : Charon.Types.literal_type -> int

val is_signed : Charon.Types.literal_type -> bool

val z_of_scalar : Charon.Values.scalar_value -> Soteria.Soteria_std.Z.t

val z_of_literal : Charon.Values.literal -> Soteria.Soteria_std.Z.t

val type_of_operand : Charon.Expressions.operand -> Charon.Types.ty

val lit_to_string : Charon.Types.literal_type -> string

val ty_as_float : Charon.Types.ty -> Charon.Values.float_type

val pp_literal_ty : Charon.Types.literal_type Fmt.t

val lit_of_int_ty : Charon.Types.integer_type -> Charon.Types.literal_type

val lit_of_scalar : Charon.Values.scalar_value -> Charon.Types.literal_type

val lit_ty_of_lit : Charon.Values.literal -> Charon.Types.literal_type

val int_ty_of_lit_ty : Charon.Types.literal_type -> Charon.Types.integer_type

val z_of_constant_expr : Charon.Types.constant_expr -> Soteria.Soteria_std.Z.t

val int_of_constant_expr : Charon.Types.constant_expr -> int

val pp_ty : Charon.Types.ty Fmt.t

val pp_type_pattern : Charon.Types.type_pattern Fmt.t

val field_tys : Charon.Types.field list -> Charon.Types.ty list

val empty_span_data : Charon.Meta.span_data

val empty_span : Charon.Meta.span

val fields_of_tys : Charon.Types.ty list -> Charon.Types.field list

val mk_array_ty : Charon.Types.ty -> Charon.Generated_Values.big_int -> Charon.Types.ty

val mk_tuple_ty : Charon.Types.ty list -> Charon.Types.ty

val unit_ptr : Charon.Types.ty_kind

The type *const ()

val unit_ref : Charon.Types.ty_kind

The type &()

val u8_ty : Charon.Types.ty_kind

The type u8

val decl_has_attr : Charon.GAst.fun_decl -> string -> bool

val meta_get_attr : Charon.Types.item_meta -> string -> string option

val decl_get_attr : Charon.GAst.fun_decl -> string -> string option

val adt_is_lang_item : string -> Charon.Types.type_decl_ref -> bool

val adt_is_box : Charon.Types.type_decl_ref -> bool

val adt_is_unsafe_cell : Charon.Types.type_decl_ref -> bool

val get_pointee : Charon.Types.ty -> Charon.Types.ty

val ty_as_adt : Charon.Types.ty -> Charon.Types.type_decl_ref

val ty_as_adt_opt : Charon.Types.ty -> Charon.Types.type_decl_ref option

val return_place : 'a Charon.GAst.gexpr_body -> Charon.Expressions.place

val ty_is_monomorphic : Charon.Types.ty -> bool

Whether the given type is monomorphic, i.e. contains no type variables. This is a conservative estimate: struct Foo<T> {} is considered polymorphic despite the generics being unused.

val tyref_is_substituted : Charon.Types.type_decl_ref -> bool

Whether the given type decl ref is properly substituted, i.e. it contains no bound type variables. This can still return true for polymorphic type decl refs, e.g. Option<@Free0>, but we guarantee all type variables are free.

val pp_span_data : Stdlib.Format.formatter -> Charon.Meta.span_data -> unit