Module Core.M

Parameters

module State : State_intf.S

Signature

module Sptr = State.Sptr

type nonrec rust_val = Sptr.t Rust_val.rust_val

val pp_rust_val : Sptr.t Rust_val.t Fmt.t

val cmp : signed:bool -> ([< Soteria__Bv_values__Typed.T.sint ] as 'a) Soteria__Bv_values__Typed.t -> 'a Soteria__Bv_values__Typed.t -> [> Soteria__Bv_values__Typed.T.sint ] Soteria__Bv_values__Typed.t Soteria_rust_lib__Rustsymex.SYMEX.t

val cmp_of_int : [ `NonZero | `MaybeZero ] Typed.t -> [> Soteria__Bv_values__Typed.T.sint ] Soteria__Bv_values__Typed.t Soteria_rust_lib__Rustsymex.SYMEX.t

val equality_check : ([< Soteria__Bv_values__Typed.T.sfloat ] as 'a) Typed.t -> 'b Typed.t -> ([> Soteria__Bv_values__Typed.T.sint ] Soteria.Bv_values.Typed.t, [> `UBPointerComparison ], 'c) Soteria_rust_lib.Rustsymex.Result.t

val binop_fn : Charon.Expressions.binop -> bool -> [< Soteria__Bv_values__Typed.T.sint ] Soteria__Bv_values__Typed.t -> [< Soteria__Bv_values__Typed.T.sint ] Soteria__Bv_values__Typed.t -> [> Soteria__Bv_values__Typed.T.sint ] Soteria__Bv_values__Typed.t

val normalise_shift_r : Charon.Expressions.binop -> 'a Typed.t -> ([> Soteria__Bv_values__Typed.T.sint ] as 'b) Typed.t -> 'b Typed.t

Rust allows shift operations on integers of differents sizes, which isn't possible in SMT-Lib, so we normalise the righthand side to match the left hand side.

val eval_lit_binop : Charon.Expressions.binop -> Charon.Types.literal_type -> Soteria_rust_lib.Rust_val.T.cval Typed.t -> Soteria_rust_lib.Rust_val.T.cval Typed.t -> ([> `Float | `MaybeZero | `NonZero ] Soteria__Bv_values__Typed.t, [> `DivisionByZero | `InvalidShift | `Overflow ], 'a) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

Evaluates a binary operator of +,-,/,*,rem, and ensures the result is within the type's constraints, else errors

val wrapping_binop : Charon.Expressions.binop -> Charon.Types.literal_type -> [< Soteria_rust_lib.Typed.T.any ] Typed.t -> [< Soteria_rust_lib.Typed.T.any MaybeZero NonZero ] Typed.t -> [> Soteria__Bv_values__Typed.T.sint ] Soteria__Bv_values__Typed.t

Wraps a given value to make it fit within the constraints of the given type

val eval_checked_lit_binop : Charon.Expressions.binop -> Charon.Types.literal_type -> [< Soteria_rust_lib.Typed.T.any ] Typed.t -> [< Soteria_rust_lib.Typed.T.any MaybeZero NonZero ] Typed.t -> ('a Rust_val.t, 'b, 'c) Soteria_rust_lib.Rustsymex.Result.t

Evaluates the checked operation, returning (wrapped value, overflowed).

val eval_ptr_binop : Charon.Expressions.binop -> Sptr.t Rust_val.t -> Sptr.t Rust_val.t -> ([> Soteria_rust_lib.Rust_val.T.cval ] Typed.t, [> `UBPointerComparison ], 'm) Soteria_rust_lib.Rustsymex.Result.t