Module Encoder.Make

Parameters

module Sptr : Sptr.S

Signature

type nonrec rust_val = Sptr.t Rust_val.t

val pp_rust_val : Sptr.t Rust_val.t Fmt.t

module ParserMonad : sig ... end

type cval_info = {

value : rust_val;

ty : Charon.Types.ty;

offset : Soteria_rust_lib.Rust_val.T.sint Typed.t;

}

val pp_cval_info : Ppx_deriving_runtime.Format.formatter -> cval_info -> Ppx_deriving_runtime.unit

val show_cval_info : cval_info -> Ppx_deriving_runtime.string

val rust_to_cvals : ?offset:Soteria_rust_lib.Rust_val.T.sint Typed.t -> rust_val -> Charon.Types.ty -> cval_info list

Converts a Rust value of the given type into a list of sub values, along with their size and offset, and whether they are interiorly mutable.

val variant_of_enum : offset:[< Soteria__Bv_values__Typed.T.sint ] Soteria__Bv_values__Typed.t -> Charon.Types.ty -> (Charon.Types.variant_id, 'state, [> `UBTransmute of string ], 'fix) ParserMonad.t

Parses the current variant of the enum at the given offset. This handles cases such as niches, where the discriminant isn't directly encoded as a tag.

type ('e, 'fix, 'state) parser = (rust_val, 'state, 'e, 'fix) ParserMonad.t

val rust_of_cvals : ?meta:'a Typed.t -> offset:Soteria_rust_lib.Rust_val.T.sint Typed.t -> Charon.Types.ty -> (rust_val, 'state, [> `RefToUninhabited | `UBDanglingPointer | `UBTransmute of string ], 'fix) ParserMonad.t

Converts a Rust type into a list of types to read, along with their offset; once these are read, symbolically decides whether we must keep reading. offset is the initial offset to read from, meta is the optional metadata, that originates from a fat pointer.

val transmute : ?verify_ptr: (Sptr.t Rust_val.full_ptr -> Charon.Types.ty -> bool DecayMapMonad.Symex.t) -> ?try_splitting:bool -> from_ty:Charon.Types.ty -> to_ty:Charon.Types.ty -> Sptr.t Rust_val.t -> (Sptr.t Rust_val.t, [> `RefToUninhabited | `UBDanglingPointer | `UBTransmute of string | `UninitializedMemoryAccess ], 'a) DecayMapMonad.Result.t

Transmute a value of the given type into the other type.

Accepts an optional verify_ptr function, that symbolically checks if a pointer can be used to read a value of the given type. This verification is a *ghost read*, and should not have side-effects.

val transmute_many : to_ty:Charon.Types.ty -> cval_info list -> (Sptr.t Rust_val.t, [> `RefToUninhabited | `UBDanglingPointer | `UBTransmute of string | `UninitializedMemoryAccess ], 'a) DecayMapMonad.Result.t

type 'a split_tree = [

| `Node of Soteria_rust_lib.Rust_val.T.sint Typed.t * 'a split_tree * 'a split_tree

| `Leaf of 'a

]

val split : Sptr.t Rust_val.t -> Charon.Types.ty -> 'a Typed.t -> ((rust_val * Charon.Types.ty) split_tree * (rust_val * Charon.Types.ty) split_tree) DecayMapMonad.t