Module Soteria_rust_lib.Error

type t = [

| `DoubleFree
(*
Tried freeing the same allocation twice
*)

| `InvalidFree
(*
Tried freeing memory at a non-0 offset
*)

| `MemoryLeak
(*
Dynamically allocated memory was not freed
*)

| `MisalignedPointer of Soteria_rust_lib.Typed.T.nonzero Typed.t * Soteria_rust_lib.Typed.T.nonzero Typed.t * Soteria_rust_lib.Typed.T.sint Typed.t
(*
Tried accessing memory with a misaligned pointer (expected, received, offset)
*)

| `NullDereference
(*
Dereferenced a null pointer
*)

| `OutOfBounds
(*
Tried accessing memory outside the allocation bounds
*)

| `UninitializedMemoryAccess
(*
Accessed uninitialised memory
*)

| `UseAfterFree
(*
Use a location in memory after it was freed
*)

| `MisalignedFnPointer
(*
Tried calling a function pointer with a non-0 offset
*)

| `NotAFnPointer
(*
Tried calling a function pointer, but it doesn't represent a function
*)

| `AccessedFnPointer
(*
Tried accessing a function pointer's pointee
*)

| `InvalidFnArgCount of int * int
(*
Invalid argument count for function (function pointers only) (expected, received)
*)

| `InvalidFnArgTys of Charon.Types.ty * Charon.Types.ty
(*
Invalid arguments for function (function pointers only) (expected, received)
*)

| `DivisionByZero
(*
Integer division by zero
*)

| `InvalidShift
(*
Binary shift that is less than 0 or that exceeds the bit size of the type
*)

| `Overflow
(*
Arithmetic over or underflow
*)

| `RefToUninhabited of Charon.Types.ty
(*
Attempt to have a reference to an uninhabited value
*)

| `InvalidRef of t
(*
Attempt to have a reference that is not valid, e.g. because it points to uninitialised memory. Keeps track of the inner error
*)

| `StdErr of string
(*
Error caused by the std library (mainly intrinsics); kind of equivalent to `Panic
*)

| `UBAbort
(*
Abort caused by an UB trap being triggered
*)

| `UBDanglingPointer
(*
Pointer was offset outside of its allocation
*)

| `UBPointerArithmetic
(*
Arithmetics on two pointers
*)

| `UBPointerComparison
(*
Comparison of pointers with different provenance
*)

| `UBIntToPointerNoProvenance of Soteria_rust_lib.Typed.T.sint Typed.t
(*
Integer to pointer cast with no provenance
*)

| `UBIntToPointerStrict
(*
Integer to pointer cast with strict provenance
*)

| `UBTransmute of string
(*
Invalid transmute, e.g. null reference, wrong enum discriminant
*)

| `AliasingError
(*
Tree borrow violation that lead to UB
*)

| `RefInvalidatedEarly
(*
A protected reference was invalidated before the end of the function
*)

| `InvalidFreeStrongProtector
(*
Tried freeing an allocation when a strongly protected reference to it still exists
*)

| `FailedAssert of string option
(*
Failed assert!(cond)
*)

| `Panic of string option
(*
Regular panic, with a message
*)

| `UnwindTerminate
(*
Unwinding terminated
*)

| `DeadVariable
(*
Local was not initialised (impossible without `mir!`)
*)

| `Breakpoint
(*
Breakpoint intrinsic
*)

| `MetaExpectedError
(*
Function was marked as expecting an error; none happened
*)

| `InvalidLayout of Charon.Types.ty
(*
Type is too large for memory
*)

| `InvalidAlloc
(*
Error in alloc/realloc; a wrong alignment or size was provided
*)

]

type warn_reason =

| InvalidReference of Soteria_rust_lib.Typed.T.sloc Typed.t

val hash_warn_reason : warn_reason -> int

val is_unwindable : [> t ] -> bool

val pp : Stdlib.Format.formatter -> t -> unit

val severity : t -> Soteria.Terminal.Diagnostic.severity

type with_trace = t * Charon.Meta.span_data Soteria.Terminal.Call_trace.t

val pp_with_trace : Stdlib.Format.formatter -> (t * Charon.Meta.span_data Soteria.Terminal.Call_trace.element list) -> unit

val compare_with_trace : (t * 'a) -> (t * 'a) -> int

val add_to_call_trace : with_trace -> Charon.Meta.span_data Soteria.Terminal.Call_trace.element -> t * Charon.Meta.span_data Soteria.Terminal.Call_trace.element list

val log_at : Trace.t -> t -> unit

val decorate : Trace.t -> t -> with_trace

module Diagnostic : sig ... end