Module Soteria_rust_lib.Layout

module BV = Typed.BitVec

exception CantComputeLayout of string * Charon.Types.ty

exception InvalidLayout of Charon.Types.ty

module Tag_layout : sig ... end

Layout of enum tags in memory. Note tags are distinct from discriminants: a discriminant is user specified and is what Rvalue.Discriminant returns, whereas a tag is specific to variant layouts, and may be of smaller size than the discriminant, or not be encoded at all if it is the untagged variant of a niche-optimised enum.

module Fields_shape : sig ... end

We use a custom type for the member offsets for layouts; this allows us to use a more efficient representation for arrays T; N, that doesn't require N offsets.

type layout = {

size : int;

align : int;

fields : Fields_shape.t;

}

val pp_layout : Stdlib.Format.formatter -> layout -> unit

module Session : sig ... end

include module type of struct include Layout_common end

val size_of_int_ty : Charon.Values.int_ty -> int

val size_of_uint_ty : Charon.Values.u_int_ty -> int

val size_of_literal_ty : Charon.Types.literal_type -> int

val is_signed : Charon.Types.literal_type -> bool

val is_int : Charon.Types.ty -> bool

val align_of_literal_ty : Charon.Types.literal_type -> int

val empty_generics : Charon.Types.generic_args

type meta_kind =

| LenKind

| VTableKind

| NoneKind

val dst_kind : Charon.Types.ty -> meta_kind

val dst_slice_ty : Charon.Types.ty -> Charon.Types.ty option

If this is a DST type with a slice tail, return the type of the slice's element.

val is_dst : Charon.Types.ty -> bool

If this is a dynamically sized type (requiring a fat pointer)

val size_to_fit : size:int -> align:int -> int

val max_array_len : int -> Z.t

val layout_of : Charon.Types.ty -> layout

val layout_of_members : ?fst_size:int -> ?fst_align:int -> Charon.Types.ty list -> layout

val layout_of_struct : Charon.Types.type_decl -> Charon.Types.field list -> layout

val layout_of_enum : Charon.Types.type_decl -> Charon.Types.variant list -> layout

val resolve_trait_ty : Charon.Types.trait_ref -> Charon.Types.trait_item_name -> Charon.Types.ty

val offset_in_array : Charon.Types.ty -> int -> int

val layout_of_s : Charon.Types.ty -> layout Rustsymex.t

val size_of_s : Charon.Types.ty -> [> Soteria__Bv_values__Typed.T.sint ] Soteria__Bv_values__Typed.t Soteria_rust_lib__Rustsymex.t

val align_of_s : Charon.Types.ty -> [> Soteria__Bv_values__Typed.T.nonzero ] Soteria__Bv_values__Typed.t Soteria_rust_lib__Rustsymex.t

val min_value_z : Charon.Types.literal_type -> Z.t

val max_value_z : Charon.Types.literal_type -> Z.t

val size_to_uint : int -> Charon.Types.ty

val lit_to_unsigned : Charon.Types.literal_type -> Charon.Types.ty

val constraints : Charon.Types.literal_type -> [< Soteria_rust_lib.Rust_val.T.cval ] Typed.t -> Soteria_rust_lib.Rust_val.T.sbool Typed.t list

val nondet_literal_ty : Charon.Types.literal_type -> Soteria_rust_lib.Rust_val.T.cval Typed.t Rustsymex.t

val nondet : Charon.Types.ty -> 'a Rust_val.rust_val Rustsymex.t

nondet ~extern ~init ty returns a nondeterministic value for ty, along with some "state". It receives a function extern to get an optional external function that computes the arbitrary value; it tries using it, and otherwise guesses the valid values. init is the initial "state", that is modified and returned by extern.

val nondets : Charon.Types.ty list -> 'a Rust_val.rust_val list Rustsymex.t

val zeroed : null_ptr:'a -> Charon.Types.ty -> 'a Rust_val.rust_val Soteria.Soteria_std.Monad.OptionM.t

val is_inhabited : Charon.Types.ty -> bool

val as_zst : Charon.Types.ty -> 'a Rust_val.rust_val Soteria.Soteria_std.Monad.OptionM.t

Returns the given type as it's unique representant if it's a ZST; otherwise None.

FIXME: giltho: this plays awfully with polymorphism.

val apply_attribute : 'a Rust_val.t -> Charon.Meta.attribute -> (unit, [> `StdErr of string ], 'b) Rustsymex.Result.t

Apply the compiler-attribute to the given value

val apply_attributes : 'a Rust_val.t -> Charon.Meta.attribute list -> (unit, [> `StdErr of string ], 'b) Rustsymex.Result.t

val is_unsafe_cell : Charon.Types.ty -> bool

val ref_tys_in : ?include_ptrs:bool -> 'a Rust_val.rust_val -> Charon.Types.ty -> ('a Rust_val.full_ptr * Charon.Types.ty) Soteria.Soteria_std.List.t

Traverses the given type and rust value, and returns all findable references with their type (ignores pointers, except if include_ptrs is true). This is needed e.g. when needing to get the pointers along with the size of their pointee, in particular in nested cases.

val update_ref_tys_in : ('acc -> 'a Rust_val.full_ptr -> Charon.Types.ty -> Charon.Types.ref_kind -> ('a Rust_val.full_ptr * 'acc, 'e, 'f) Rustsymex.Result.t) -> 'acc -> 'a Rust_val.rust_val -> Charon.Types.ty -> ('a Rust_val.rust_val * 'acc, 'e, 'f) Rustsymex.Result.t

val is_abi_compatible : Charon.Types.ty -> Charon.Types.ty -> bool

is_abi_compatible ty1 ty2 is true if a function expecting an argument of type ty1 can be called with an argument of type ty2.

The full specification is available at: https://doc.rust-lang.org/nightly/std/primitive.fn.html#abi-compatibility