Module Layout_common.Fields_shape

We use a custom type for the member offsets for layouts; this allows us to use a more efficient representation for arrays T; N, that doesn't require N offsets.

type discriminator =

| Invalid

| Known of Charon.Types.variant_id

| Branch of {

offset : Typed.T.sint Typed.t;

tag_ty : Charon.Types.literal_type;

children : (Typed.T.sint Typed.t * Typed.T.sint Typed.t * discriminator) list;

fallback : discriminator;

}

Mirror's MiniRust and Charon's discriminator

val pp_discriminator : Ppx_deriving_runtime.Format.formatter -> discriminator -> Ppx_deriving_runtime.unit

val show_discriminator : discriminator -> Ppx_deriving_runtime.string

type tagger = (Typed.T.sint Typed.t * Typed.T.sint Typed.t) option

The (offset * tag) associated to a variant. If None, the variant is either uninhabited or the untagged variant.

Note tags are distinct from discriminants: a discriminant is user specified and is what Rvalue.Discriminant returns, whereas a tag is specific to variant layouts, and may be of smaller size than the discriminant, or not be encoded at all if it is the untagged variant of a niche-optimised enum.

val pp_tagger : (Typed.T.sint Typed.t * Typed.T.sint Typed.t) option Fmt.t

type t =

| Primitive
(*
No fields present
*)

| Arbitrary of Charon.Types.variant_id * Typed.T.sint Typed.t Stdlib.Array.t
(*
Arbitrary field placement (structs, unions...), with the variant (e.g. enums with a single inhabited variant)
*)

| Enum of discriminator * (tagger * t) Stdlib.Array.t
(*
Enum fields: for each variant, a possible tag with tagger, along with an array of field shapes for each variant (indexed by variant ID). The discriminator allows calculating the current variant. Using offset_of on this isn't valid; one must first retrieve the fields shape of the corresponding variant.
*)

| Array of {

stride : Typed.T.sint Typed.t;

is_ptr : bool;

}
(*
All fields are equally spaced (arrays, slices). We have a is_ptr flag, as we also use this layout for fat pointers.
*)

val pp : t Fmt.t

val offset_of : int -> t -> Typed.T.sint Typed.t

val shape_for_variant : Charon.Types.VariantId.id -> t -> t

val iter_vars_discriminator : discriminator -> ((Soteria.Symex.Var.t * 'a Typed.ty) -> unit) -> unit

val iter_vars_tagger : tagger -> ((Soteria.Symex.Var.t * 'a Typed.ty) -> unit) -> unit

val iter_vars : t -> ((Soteria.Symex.Var.t * 'a Typed.ty) -> unit) -> unit