Module Layout_common.Tag_layout

Layout of enum tags in memory. Note tags are distinct from discriminants: a discriminant is user specified and is what Rvalue.Discriminant returns, whereas a tag is specific to variant layouts, and may be of smaller size than the discriminant, or not be encoded at all if it is the untagged variant of a niche-optimised enum.

type encoding =

| Direct

| Niche of Charon.Types.variant_id

and t = {

offset : Soteria_rust_lib.Typed.T.sint Typed.t;

ty : Charon.Types.literal_type;

encoding : encoding;

tags : Soteria_rust_lib.Typed.T.sint Typed.t option Stdlib.Array.t;
(*
The tag associated to each variant, indexed by variant ID. If None, the variant is either uninhabited or the untagged variant.
*)

}

val pp_encoding : Ppx_deriving_runtime.Format.formatter -> encoding -> Ppx_deriving_runtime.unit

val show_encoding : encoding -> Ppx_deriving_runtime.string

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

val iter_vars : t -> ((Soteria.Symex.Var.t * 'a Typed.ty) -> unit) -> unit