Module Soteria_rust_lib.Rust_val

type 'ptr meta =

| Thin

| Len of Soteria_rust_lib.Typed.T.sint Typed.t

| VTable of 'ptr

type 'ptr full_ptr = 'ptr * 'ptr meta

type 'ptr t =

| Int of Soteria_rust_lib.Typed.T.sint Typed.t

| Float of Soteria_rust_lib.Typed.T.sfloat Typed.t

| Ptr of 'ptr full_ptr
(*
pointer, parametric to enable Ruxt, with optional meta
*)

| Enum of Soteria_rust_lib.Typed.T.sint Typed.t * 'ptr t list
(*
discriminant * values
*)

| Tuple of 'ptr t list
(*
contains ordered values
*)

| Union of ('ptr t * Soteria_rust_lib.Typed.T.sint Typed.t) list
(*
list of blocks in the union, with their offset
*)

| PolyVal of Charon.Types.type_var_id
(*
The opaque value of a type variable, identified by (type variable index, unique identifier).
*)

type 'ptr rust_val = 'ptr t

val is_empty : 'a t -> bool

is_empty v is true if the value is "empty"; ie. it doesn't contain any Base or Ptr value. We also consider Enum to be non-empty, because of the discriminant, though this *may* be wrong if it's a ZST.

Used in unsizing, to find the field with the pointer to modify.

val pp_meta : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a meta -> unit

val pp_full_ptr : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> ('a * 'a meta) -> unit

val pp_rust_val : (Stdlib.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> 'a t Fmt.t

val pp : (Stdlib.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> 'a t Fmt.t

val ppa_rust_val : Stdlib.Format.formatter -> 'a t -> unit

val unit_ : 'a t

val as_ptr : 'a t -> 'a full_ptr

val as_ptr_or : make:([< Soteria_rust_lib.Typed.T.any NonZero Zero ] Typed.t -> 'a) -> 'a t -> 'a full_ptr

val as_base_f : Charon.Types.float_type -> 'a t -> Soteria_rust_lib.Typed.T.sfloat Typed.t

val as_base : Charon.Types.literal_type -> 'a t -> [< Soteria_rust_lib.Typed.T.any NonZero Zero ] Typed.t

val as_base_i : Charon.Values.u_int_ty -> 'a t -> [< Soteria_rust_lib.Typed.T.any NonZero Zero ] Typed.t

val as_tuple : 'a t -> 'a t list

val flatten : 'a t -> 'a t list

val iter_vars : ('a -> ((Soteria.Symex.Var.t * 'b Typed.ty) -> unit) -> unit) -> 'a t -> ((Soteria.Symex.Var.t * 'b Typed.ty) -> unit) -> unit

val subst : ((Soteria.Symex.Var.t -> Soteria.Symex.Var.t) -> 'a -> 'a) -> (Soteria.Symex.Var.t -> Soteria.Symex.Var.t) -> 'a t -> 'a t