Module Soteria_rust_lib.Sptr

The base type of pointers, permitting simple operations on the pointer type. The majority of relevant operations are exposed via the state monad's pointer module, StateM.Sptr, or the Svalue.Ptr module.

module type DecayMapS = sig ... end

A map to store information on "decayed" pointers, namely mapping from locations to integers (their decayed) value.

module DecayMap : DecayMapS

type t = Svalue.Typed.T.sptr_t Svalue.Typed.t

val pp : Stdlib.Format.formatter -> 'a Svalue.Typed.t -> unit

val dangling_if_zst : Charon.Types.ty -> Soteria_rust_lib__Rustsymex.MonadState.t -> (([> Soteria_rust_lib__Svalue__Typed.T.sptr_t ] Soteria_rust_lib__Svalue__Typed.t option, [> `InvalidLayout of Charon.Types.ty ], 'a) Soteria.Soteria_std.Compo_res.t * Soteria_rust_lib__Rustsymex.MonadState.t) Soteria__Symex.Make(Soteria.Bv_values.Bv_solver.Z3_solver(Svalue.Typed)).t

Creates a dangling pointer to the given type, if that type is a ZST; returns None otherwise.

val raw_offset : [< Soteria_rust_lib__Svalue__Typed.T.sptr_t ] Soteria_rust_lib__Svalue__Typed.t -> [< `NonZero | `Zero ] Soteria_rust_lib__Svalue__Typed.t -> Soteria_rust_lib__Rustsymex.MonadState.t -> (([> Soteria_rust_lib__Svalue__Typed.T.sptr_t ] Soteria_rust_lib__Svalue__Typed.t, [> `UBDanglingPointer ], 'a) Soteria.Soteria_std.Compo_res.t * Soteria_rust_lib__Rustsymex.MonadState.t) Soteria__Symex.Make(Soteria.Bv_values.Bv_solver.Z3_solver(Svalue.Typed)).t

A simplified, untyped (and unsafe) version of offset, that adds a signed integer to this pointer's offset. This offset doesn't check whether the resulting pointer is dangling after being offset.

val _decay : expose:bool -> [< Soteria_rust_lib__Svalue__Typed.T.sptr_t ] Soteria_rust_lib__Svalue__Typed.t -> [> `NonZero | `Zero ] Soteria_rust_lib__Svalue__Typed.t DecayMap.SM.t

val decay : [< Soteria_rust_lib__Svalue__Typed.T.sptr_t ] Soteria_rust_lib__Svalue__Typed.t -> [> `NonZero | `Zero ] Soteria_rust_lib__Svalue__Typed.t DecayMap.SM.t

Decay a pointer into an integer value, losing provenance. This does not expose the address of the allocation; for that, use expose

val expose : [< Soteria_rust_lib__Svalue__Typed.T.sptr_t ] Soteria_rust_lib__Svalue__Typed.t -> [> `NonZero | `Zero ] Soteria_rust_lib__Svalue__Typed.t DecayMap.SM.t

Decay a pointer into an integer value, exposing the address of the allocation, allowing it to be retrieved with DecayMapS.from_exposed later.

val distance : [< Soteria_rust_lib__Svalue__Typed.T.sptr_t ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.T.sptr_t ] Soteria_rust_lib__Svalue__Typed.t -> [> Svalue.Typed.T.sint ] Svalue.Typed.t DecayMap.SM.t

The distance, in bytes, between two pointers; if they point to different allocations, they are decayed and substracted.