Module type Sptr.DecayMapS

A map to store information on "decayed" pointers, namely mapping from locations to integers (their decayed) value.

It exposes two functions: decay, to convert a location to an integer (given the size and alignment of its allocation), and from_exposed which does the reverse operation. Importantly, while decay is total, from_exposed is partial: if a provenance cannot be guessed from the integer, None is returned.

include Soteria.Sym_states.Base.M(Rustsymex).S

type t

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

module SM : Soteria.Sym_states.State_monad.S with module Symex = Rustsymex and module Value = Rustsymex.Value and type st = t option

type syn

val pp_syn : Ppx_deriving_runtime.Format.formatter -> syn -> Ppx_deriving_runtime.unit

val show_syn : syn -> Ppx_deriving_runtime.string

val to_syn : t -> syn list

val ins_outs : syn -> Rustsymex.Value.Expr.t list * Rustsymex.Value.Expr.t list

val produce : syn -> t option -> t option Rustsymex.Producer.t

val consume : syn -> t option -> (t option, syn list) Rustsymex.Consumer.t

val empty : t

val decay : expose:bool -> size:[< Soteria_rust_lib.Typed.T.sint ] Typed.t -> align:[< Soteria_rust_lib.Typed.T.nonzero ] Typed.t -> [< Soteria_rust_lib.Typed.T.sloc ] Typed.t -> Soteria_rust_lib.Typed.T.sint Typed.t SM.t

Decays the given location into an integer, updating the decay map accordingly. If expose is true, the provenance is marked as exposed, and can be retrieved later with from_exposed. Returns the decayed integer, along with the updated decay map.

val from_exposed : [< Soteria_rust_lib.Typed.T.sint ] Typed.t -> (Soteria_rust_lib.Typed.T.sloc Typed.t * Soteria_rust_lib.Typed.T.sint Typed.t) option SM.t

Tries finding, for the given integer, the matching provenance in the decay map. If found, it returns that provenance, along with the exposed address for that allocation at offset 0. Otherwise returns None.