Module State.Sptr

include module type of Sptr

module type DecayMapS = sig ... end

A map to store information on "decayed" pointers, namely mapping from locations to integers (their decayed) value.

module DecayMap : DecayMapS

type t = Svalue.Typed.T.sptr_t Svalue.Typed.t

val pp : Stdlib.Format.formatter -> 'a Svalue.Typed.t -> unit

val dangling_if_zst : Charon.Types.ty -> Soteria_rust_lib__Rustsymex.MonadState.t -> (([> Soteria_rust_lib__Svalue__Typed.T.sptr_t ] Soteria_rust_lib__Svalue__Typed.t option, [> `InvalidLayout of Charon.Types.ty ], 'a) Soteria.Soteria_std.Compo_res.t * Soteria_rust_lib__Rustsymex.MonadState.t) Soteria__Symex.Make(Soteria.Bv_values.Bv_solver.Z3_solver(Svalue.Typed)).t

Creates a dangling pointer to the given type, if that type is a ZST; returns None otherwise.

val raw_offset : [< Soteria_rust_lib__Svalue__Typed.T.sptr_t ] Soteria_rust_lib__Svalue__Typed.t -> [< `NonZero | `Zero ] Soteria_rust_lib__Svalue__Typed.t -> Soteria_rust_lib__Rustsymex.MonadState.t -> (([> Soteria_rust_lib__Svalue__Typed.T.sptr_t ] Soteria_rust_lib__Svalue__Typed.t, [> `UBDanglingPointer ], 'a) Soteria.Soteria_std.Compo_res.t * Soteria_rust_lib__Rustsymex.MonadState.t) Soteria__Symex.Make(Soteria.Bv_values.Bv_solver.Z3_solver(Svalue.Typed)).t

A simplified, untyped (and unsafe) version of offset, that adds a signed integer to this pointer's offset. This offset doesn't check whether the resulting pointer is dangling after being offset.

val _decay : expose:bool -> [< Soteria_rust_lib__Svalue__Typed.T.sptr_t ] Soteria_rust_lib__Svalue__Typed.t -> [> `NonZero | `Zero ] Soteria_rust_lib__Svalue__Typed.t DecayMap.SM.t

val decay : [< Soteria_rust_lib__Svalue__Typed.T.sptr_t ] Soteria_rust_lib__Svalue__Typed.t -> [> `NonZero | `Zero ] Soteria_rust_lib__Svalue__Typed.t DecayMap.SM.t

Decay a pointer into an integer value, losing provenance. This does not expose the address of the allocation; for that, use expose

val expose : [< Soteria_rust_lib__Svalue__Typed.T.sptr_t ] Soteria_rust_lib__Svalue__Typed.t -> [> `NonZero | `Zero ] Soteria_rust_lib__Svalue__Typed.t DecayMap.SM.t

Decay a pointer into an integer value, exposing the address of the allocation, allowing it to be retrieved with DecayMapS.from_exposed later.

val distance : [< Soteria_rust_lib__Svalue__Typed.T.sptr_t ] Soteria_rust_lib__Svalue__Typed.t -> [< Soteria_rust_lib__Svalue__Typed.T.sptr_t ] Soteria_rust_lib__Svalue__Typed.t -> [> Svalue.Typed.T.sint ] Svalue.Typed.t DecayMap.SM.t

The distance, in bytes, between two pointers; if they point to different allocations, they are decayed and substracted.

val offset : ?check_signed:bool -> ?ty:Charon.Types.ty -> [< Svalue.Typed.T.sint ] v -> t -> t ret

offset ?check_signed ?ty ptr off Offsets ptr by the size of ty * off. ty defaults to u8.

If check_signed is Some s, the offset is checked for overflows and dangling pointers. s represents whether the offset is signed or unsigned (this matters for overflow checks on the multiplication with the pointee type).

val check_non_dangling : [< Svalue.Typed.T.sptr_f ] v -> Charon.Types.ty -> unit ret

Checks this pointer isn't dangling for the given pointee type, i.e. it points to an allocation and doesn't range outside of it. This also checks the allocation is live, and that the type is inhabited,

val check_non_dangling_untyped : [< Svalue.Typed.T.sptr_t ] v -> Svalue.Typed.T.sint v -> unit ret

Same as check_non_dangling, but for untyped ranges, where only a size is known. The size is signed: if less than 0, the range preceding the pointer is checked.

val check_aligned : [< Svalue.Typed.T.sptr_f ] v -> Charon.Types.ty -> unit ret

Checks this pointer is sufficiently aligned for the given type. This takes into account the metadata: for a &dyn Trait, it will access the VTable to check alignment.