Module Tree_state.Tree_block

module Encoder : sig ... end

type rust_val = Sptr.t Rust_val.t

val pp_rust_val : Sptr.t Rust_val.t Fmt.t

module MemVal : sig ... end

type nonrec sint = MemVal.S_bounded_int.t Sptr.DecayMapMonad.Value.t

type nonrec (!'a, !'sint) tree = {

node : 'a Soteria__Sym_states__Tree_block.node;

range : 'sint * 'sint;

children : (('a, 'sint) Soteria__Sym_states__Tree_block.tree * ('a, 'sint) Soteria__Sym_states__Tree_block.tree) option;

}

module Range : sig ... end

module Split_tree : sig ... end

A Split_tree is a simplified representation of a tree, that has no offset. It however indicates, on Nodes, at what offset the split occurs, relative to that node's start.

module Node : sig ... end

module Tree : sig ... end

type t = {

root : Tree.t;

bound : sint option;

}

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

module SM : sig ... end

val is_empty : t -> bool

val pp_pretty : Stdlib.Format.formatter -> t -> unit

type serialized =

| MemVal of {

offset : sint;

len : sint;

v : MemVal.serialized;

}

| Bound of sint

val pp_serialized : Ppx_deriving_runtime.Format.formatter -> serialized -> Ppx_deriving_runtime.unit

val show_serialized : serialized -> Ppx_deriving_runtime.string

val lift_miss : offset:sint -> len:sint -> ('a, 'b, MemVal.serialized) Sptr.DecayMapMonad.Result.t -> ('a, 'b, serialized list) Sptr.DecayMapMonad.Result.t

val iter_values_serialized : serialized list -> (MemVal.serialized -> unit) -> unit

val of_opt : ?mk_fixes:(unit -> 'a list Sptr.DecayMapMonad.t) -> 'b option -> ('b, 'c, 'a) Sptr.DecayMapMonad.Result.t

val to_opt : t -> t option

val with_bound_check : ?mk_fixes:(unit -> serialized list list Sptr.DecayMapMonad.t) -> sint -> (Tree.t -> ('a * Tree.t, [> `OutOfBounds ] as 'b, serialized list) Sptr.DecayMapMonad.Result.t) -> ('a, 'b, serialized list) SM.Result.t

val assert_exclusively_owned : t option -> (unit, 'a, serialized list) Sptr.DecayMapMonad.Result.t

val get_raw_tree_owned : MemVal.S_bounded_int.t Sptr.DecayMapMonad.Value.t -> MemVal.S_bounded_int.t Sptr.DecayMapMonad.Value.t -> ((MemVal.t, MemVal.S_bounded_int.t Sptr.DecayMapMonad.Value.t) Soteria__Sym_states__Tree_block.tree, [> `OutOfBounds ], serialized list) SM.Result.t

val put_raw_tree : MemVal.S_bounded_int.t Sptr.DecayMapMonad.Value.t -> Tree.t -> (unit, [> `OutOfBounds ], serialized list) SM.Result.t

val subst_serialized : (Soteria.Symex.Var.t -> Soteria.Symex.Var.t) -> serialized -> serialized

val iter_vars_serialized : serialized -> ((Soteria.Symex.Var.t * 'a Sptr.DecayMapMonad.Value.ty) -> unit) -> unit

val serialize : t -> serialized list

val consume_bound : MemVal.S_bounded_int.t Sptr.DecayMapMonad.Value.t -> t option -> (t option, 'a, 'b) Sptr.DecayMapMonad.Result.t

val produce_bound : sint -> unit SM.t

val produce_mem_val : sint -> MemVal.S_bounded_int.t Sptr.DecayMapMonad.Value.t -> MemVal.serialized -> unit SM.t

val consume_mem_val : sint -> sint -> MemVal.serialized -> t option -> (t option, 'a, serialized list) Sptr.DecayMapMonad.Result.t

val produce : serialized -> unit SM.t

val lift_symex : 'a Sptr.DecayMapMonad.Symex.t -> 'a SM.t

val sint_to_int : [< Soteria__Bv_values__Typed.T.any ] Soteria__Bv_values__Typed.t -> int Sptr.DecayMapMonad.t

val collect_leaves : Tree.t -> ((rust_val * [> Soteria__Bv_values__Typed.T.sint ] Soteria__Bv_values__Typed.t) list, 'a, 'b) Sptr.DecayMapMonad.Result.t

val decode_mem_val : ty:Charon.Types.ty -> MemVal.value -> Soteria_rust_lib__Sptr.DecayMap.t -> ((Encoder.rust_val, [> `InvalidLayout of Charon.Types.ty | `UninitializedMemoryAccess ], 'a) Soteria.Symex.Compo_res.t * Soteria_rust_lib__Sptr.DecayMap.t) Rustsymex.t

val decode_lazy : ty:Charon.Types.ty -> Tree.t -> Soteria_rust_lib__Sptr.DecayMap.t -> ((Encoder.rust_val, 'a, 'b) Soteria.Symex.Compo_res.t * Soteria_rust_lib__Sptr.DecayMap.t) Rustsymex.t

val decode_tree : ty:Charon.Types.ty -> Tree.t -> (Encoder.rust_val, [> `InvalidLayout of Charon.Types.ty | `UninitializedMemoryAccess ], 'a) Sptr.DecayMapMonad.Result.t

val uninit : (sint * sint) -> Tree_borrow.tb_state -> Tree.t

val zeros : (sint * sint) -> Tree_borrow.tb_state -> Tree.t

val mk_fix_typed : sint -> Charon.Types.ty -> unit -> Soteria_rust_lib__Sptr.DecayMap.t -> (serialized list list * Soteria_rust_lib__Sptr.DecayMap.t) Rustsymex.t

val mk_fix_any : sint -> sint -> unit -> serialized list list

val mk_fix_any_s : sint -> sint -> unit -> serialized list list Sptr.DecayMapMonad.t

val as_owned : ?mk_fixes: (unit -> Soteria_rust_lib__Sptr.DecayMap.t -> ('a list * Soteria_rust_lib__Sptr.DecayMap.t) Rustsymex.t) -> ('b * 'c, 'd) tree -> (('b * 'c) -> ('e, 'f, 'a) Sptr.DecayMapMonad.Result.t) -> ('e, 'f, 'a) Sptr.DecayMapMonad.Result.t

val check_owned : Soteria_rust_lib.Typed.T.sint Typed.t -> [< Soteria_rust_lib.Typed.T.nonzero ] Typed.t -> (unit, [> `OutOfBounds ], serialized list) SM.Result.t

val load : ignore_borrow:bool -> Soteria_rust_lib.Typed.T.sint Typed.t -> Charon.Types.ty -> Tree_borrow.tag option -> Tree_borrow.t -> (Encoder.rust_val, [> `AliasingError | `InvalidLayout of Charon.Types.ty | `OutOfBounds | `UninitializedMemoryAccess ], serialized list) Soteria.Symex.Compo_res.t SM.t

val store : [ `NonZero | `Zero ] Typed.t -> rust_val -> Tree_borrow.tag option -> Tree_borrow.t -> (unit, [> `AliasingError | `InvalidLayout of Charon.Types.ty | `OutOfBounds ], serialized list) SM.Result.t

val get_init_leaves : Soteria_rust_lib.Typed.T.sint Typed.t -> [< Soteria_rust_lib.Typed.T.nonzero ] Typed.t -> ((rust_val * Soteria_rust_lib.Typed.T.sint Typed.t) list, [> `OutOfBounds ], serialized list) SM.Result.t

val uninit_range : Soteria_rust_lib.Typed.T.sint Typed.t -> Soteria_rust_lib.Typed.T.sint Typed.t -> (unit, [> `OutOfBounds ], serialized list) SM.Result.t

val zero_range : Soteria_rust_lib.Typed.T.sint Typed.t -> Soteria_rust_lib.Typed.T.sint Typed.t -> (unit, [> `OutOfBounds ], serialized list) SM.Result.t

val alloc : ?zeroed:bool -> sint -> t

val with_tb_access : Soteria_rust_lib.Typed.T.sint Typed.t -> Soteria_rust_lib.Typed.T.sint Typed.t -> (Tree_borrow.tb_state -> (Tree_borrow.tb_state, [> `OutOfBounds ] as 'a, serialized list) Soteria.Symex.Compo_res.t Rustsymex.t) -> (unit, 'a, serialized list) SM.Result.t

val protect : Soteria_rust_lib.Typed.T.sint Typed.t -> Soteria_rust_lib.Typed.T.sint Typed.t -> Tree_borrow.tag -> Tree_borrow.t -> (unit, [> `AliasingError | `OutOfBounds ], serialized list) SM.Result.t

val unprotect : Soteria_rust_lib.Typed.T.sint Typed.t -> Soteria_rust_lib.Typed.T.sint Typed.t -> Tree_borrow.tag -> Tree_borrow.t -> (unit, [> `OutOfBounds ], serialized list) SM.Result.t

val tb_access : Soteria_rust_lib.Typed.T.sint Typed.t -> Soteria_rust_lib.Typed.T.sint Typed.t -> Tree_borrow.tag -> Tree_borrow.t -> (unit, [> `AliasingError | `OutOfBounds ], serialized list) SM.Result.t