Module Soteria_rust_lib.Store

module Map : sig ... end

module Place : sig ... end

module Binding : sig ... end

type t = Binding.t Map.t

val pp : Stdlib.Format.formatter -> Binding.t Map.t -> unit

val reserve : Map.key -> Charon.Types.ty -> Binding.t Map.t -> Binding.t Map.t

val declare : Map.key -> Binding.kind -> Binding.t Map.t -> Binding.t Map.t

val declare_value : Map.key -> Rust_val.t -> Binding.t Map.t -> Binding.t Map.t

val declare_ptr : Map.key -> Rust_val.full_ptr -> Binding.t Map.t -> Binding.t Map.t

val declare_uninit : Map.key -> Binding.t Map.t -> Binding.t Map.t

val dealloc : Map.key -> Binding.t Map.t -> Binding.t Map.t

val get_ty : Map.key -> Binding.t Map.t -> Charon.Types.ty

val find : Map.key -> t -> Binding.t

val empty : 'a Map.t

val bindings : t -> (Map.key * Binding.t) list

val iter_bindings : t -> ((Map.key * Binding.t) -> unit) -> unit

val try_load : Place.t -> t -> Binding.kind option

try_load p s tries loading p from the s, returning Some v if it succeded, and None if it has to be spilled into the heap.

val try_store : Place.t -> t -> (Typed.T.sint Typed.t, Typed.T.nonzero Typed.t, Typed.T.sfloat Typed.t, Sptr.t) Rust_val.raw -> Binding.t Map.t option