Module Store.Binding

type kind =

| Stackptr of Rust_val.full_ptr

| Value of Rust_val.t

| Uninit

| Dead

We have four kinds of bindings:

Stackptr: the symbol is bound to a stack pointer, that lives in the heap.

Value: the symbol is bound to an immediate value; it does not have an address.

Uninit: the symbol is bound to an immediate, uninitialized value.

Dead: the symbol is dead; it doesn't exist (e.g. after a StorageDead).

val pp_kind : Ppx_deriving_runtime.Format.formatter -> kind -> Ppx_deriving_runtime.unit

val show_kind : kind -> Ppx_deriving_runtime.string

type t = {

kind : kind;

ty : Charon.Types.ty;

}

val pp : Stdlib.Format.formatter -> t -> unit

val as_value : t -> Rust_val.t option

val bind_value : (Rust_val.t -> kind) -> kind option -> kind option