Module Store.Place

type kind =

| Local of Charon.Expressions.local_id

| Field of t * Charon.Expressions.field_proj_kind * Charon.Types.field_id

| Index of t * int

| Metadata of t

and t = {

kind : kind;

origin : Charon.Expressions.place;

}

A simple place that lives on the store and doesn't need to be decayed to the heap for reading or writing. Almost identical to Expressions.place, but with only field and concrete index access.

val pp : Stdlib.Format.formatter -> t -> Ppx_deriving_runtime.unit

val root : t -> Charon.Expressions.local_id

Local root of the place

val update_val : t -> f:(Rust_val.t -> Rust_val.t) -> Rust_val.t -> Rust_val.t option

update_val sp inner' v updates the root value v with the value at sp replaced by inner'; None if sp is not navigable.

val is_local : t -> bool

val local : Charon.Expressions.local_id -> Charon.Generated_Types.ty -> t