Module Svalue.Ext

module Sv = Soteria.Bv_values.Svalue

type 'ghost sv = ('ghost, 'ghost ext_t, 'ghost ext_ty) Sv.t

and 'ghost svty = 'ghost ext_ty Sv.ty

and 'ghost ptr = {

ptr : 'ghost sv;

tag : Svalue.Ptr_tag.t option;

size : 'ghost sv;

align : 'ghost sv;

}

and ('sc, 'ag) block_value =

| Scalar of 'sc

| Aggregate of 'ag * Charon.Types.ty

The value inside a block: either a scalar or an aggregate that can be split further.

and ('sc, 'ag, 'ofs, 'sz) block = {

value : ('sc, 'ag) block_value;

offset : 'ofs;

size : 'sz;

}

A block of an encoded value: a value, along with the offset it is at and the size it spans. ty is Some iff the value is a whole aggregate of that type.

and 'ghost ext_ty =

| TEnum of Charon.Types.type_decl_ref
(*
the type decl ref of an enum
*)

| TUnion of Charon.Types.type_decl_ref
(*
the type decl ref of a union
*)

| TTuple of 'ghost svty list
(*
structs and tuples (ordered fields)
*)

| TArray of 'ghost svty * Soteria.Soteria_std.Z.t
(*
arrays (all elements share the same type)
*)

| TThinPtr

| TFullPtr

| TPtrMeta
(*
the type of a pointer's metadata (unit, length or metadata). this is not exposed to the user!
*)

| TPolyType

and 'g ptr_meta =

| MetaLen of 'g sv

| MetaVTable of 'g sv

| MetaUnit

and 'ghost ext_t =

| Ptr of 'ghost sv * 'ghost sv
(*
pointer, with meta
*)

| PtrMeta of 'ghost ptr_meta

| ThinPtr of 'ghost ptr
(*
thin pointer, without metadata but with extra info on the pointer
*)

| Enum of Charon.Types.variant_id * 'ghost sv list
(*
variant id * values
*)

| Tuple of 'ghost sv list
(*
structs and tuples: ordered values
*)

| Array of 'ghost sv Soteria.Soteria_std.Iarray.t
(*
arrays: ordered values, all of the same type
*)

| Union of ('ghost sv, 'ghost sv, 'ghost sv, 'ghost sv) block list
(*
list of blocks in the union
*)

| PolyVal of Charon.Types.type_var_id
(*
The opaque value of a type variable, identified by (type variable index, unique identifier).
*)

val equal_sv : 'ghost. ('ghost -> 'ghost -> Ppx_deriving_runtime.bool) -> 'ghost sv -> 'ghost sv -> Ppx_deriving_runtime.bool

val equal_svty : 'ghost. ('ghost -> 'ghost -> Ppx_deriving_runtime.bool) -> 'ghost svty -> 'ghost svty -> Ppx_deriving_runtime.bool

val equal_ptr : 'ghost. ('ghost -> 'ghost -> Ppx_deriving_runtime.bool) -> 'ghost ptr -> 'ghost ptr -> Ppx_deriving_runtime.bool

val equal_block_value : 'sc 'ag. ('sc -> 'sc -> Ppx_deriving_runtime.bool) -> ('ag -> 'ag -> Ppx_deriving_runtime.bool) -> ('sc, 'ag) block_value -> ('sc, 'ag) block_value -> Ppx_deriving_runtime.bool

val equal_block : 'sc 'ag 'ofs 'sz. ('sc -> 'sc -> Ppx_deriving_runtime.bool) -> ('ag -> 'ag -> Ppx_deriving_runtime.bool) -> ('ofs -> 'ofs -> Ppx_deriving_runtime.bool) -> ('sz -> 'sz -> Ppx_deriving_runtime.bool) -> ('sc, 'ag, 'ofs, 'sz) block -> ('sc, 'ag, 'ofs, 'sz) block -> Ppx_deriving_runtime.bool

val equal_ext_ty : 'ghost. ('ghost -> 'ghost -> Ppx_deriving_runtime.bool) -> 'ghost ext_ty -> 'ghost ext_ty -> Ppx_deriving_runtime.bool

val equal_ptr_meta : 'g. ('g -> 'g -> Ppx_deriving_runtime.bool) -> 'g ptr_meta -> 'g ptr_meta -> Ppx_deriving_runtime.bool

val equal_ext_t : 'ghost. ('ghost -> 'ghost -> Ppx_deriving_runtime.bool) -> 'ghost ext_t -> 'ghost ext_t -> Ppx_deriving_runtime.bool

val compare_sv : 'ghost. ('ghost -> 'ghost -> Ppx_deriving_runtime.int) -> 'ghost sv -> 'ghost sv -> Ppx_deriving_runtime.int

val compare_svty : 'ghost. ('ghost -> 'ghost -> Ppx_deriving_runtime.int) -> 'ghost svty -> 'ghost svty -> Ppx_deriving_runtime.int

val compare_ptr : 'ghost. ('ghost -> 'ghost -> Ppx_deriving_runtime.int) -> 'ghost ptr -> 'ghost ptr -> Ppx_deriving_runtime.int

val compare_block_value : 'sc 'ag. ('sc -> 'sc -> Ppx_deriving_runtime.int) -> ('ag -> 'ag -> Ppx_deriving_runtime.int) -> ('sc, 'ag) block_value -> ('sc, 'ag) block_value -> Ppx_deriving_runtime.int

val compare_block : 'sc 'ag 'ofs 'sz. ('sc -> 'sc -> Ppx_deriving_runtime.int) -> ('ag -> 'ag -> Ppx_deriving_runtime.int) -> ('ofs -> 'ofs -> Ppx_deriving_runtime.int) -> ('sz -> 'sz -> Ppx_deriving_runtime.int) -> ('sc, 'ag, 'ofs, 'sz) block -> ('sc, 'ag, 'ofs, 'sz) block -> Ppx_deriving_runtime.int

val compare_ext_ty : 'ghost. ('ghost -> 'ghost -> Ppx_deriving_runtime.int) -> 'ghost ext_ty -> 'ghost ext_ty -> Ppx_deriving_runtime.int

val compare_ptr_meta : 'g. ('g -> 'g -> Ppx_deriving_runtime.int) -> 'g ptr_meta -> 'g ptr_meta -> Ppx_deriving_runtime.int

val compare_ext_t : 'ghost. ('ghost -> 'ghost -> Ppx_deriving_runtime.int) -> 'ghost ext_t -> 'ghost ext_t -> Ppx_deriving_runtime.int

val pp_block_value : (Stdlib.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> (Stdlib.Format.formatter -> 'b -> Ppx_deriving_runtime.unit) -> Stdlib.Format.formatter -> ('a, 'b) block_value -> Ppx_deriving_runtime.unit

val pp_block : (Stdlib.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> (Stdlib.Format.formatter -> 'b -> Ppx_deriving_runtime.unit) -> (Stdlib.Format.formatter -> 'c -> Ppx_deriving_runtime.unit) -> (Stdlib.Format.formatter -> 'd -> Ppx_deriving_runtime.unit) -> Stdlib.Format.formatter -> ('a, 'b, 'c, 'd) block -> Ppx_deriving_runtime.unit

val pp_ext_ty : Ppx_deriving_runtime.Format.formatter -> 'ghost ext_ty -> Ppx_deriving_runtime.unit

val pp_svty : 'ghost svty Fmt.t

module Rust_ext : Sv.Value_ext with type 'ghost t = 'ghost ext_t and type 'ghost ty = 'ghost ext_ty