Module Soteria_rust_lib.Tree_borrows

module type T = sig ... end

type access =

| Read

| Write

and state =

| Reserved of bool

| Unique

| Frozen

| ReservedIM

| Cell

| Disabled

| UB

and protector =

| Strong

| Weak

and locality =

| Local

| Foreign

val pp_access : Ppx_deriving_runtime.Format.formatter -> access -> Ppx_deriving_runtime.unit

val show_access : access -> Ppx_deriving_runtime.string

val pp_state : Ppx_deriving_runtime.Format.formatter -> state -> Ppx_deriving_runtime.unit

val show_state : state -> Ppx_deriving_runtime.string

val pp_protector : Ppx_deriving_runtime.Format.formatter -> protector -> Ppx_deriving_runtime.unit

val show_protector : protector -> Ppx_deriving_runtime.string

val pp_locality : Ppx_deriving_runtime.Format.formatter -> locality -> Ppx_deriving_runtime.unit

val show_locality : locality -> Ppx_deriving_runtime.string

val equal_access : access -> access -> Ppx_deriving_runtime.bool

val equal_state : state -> state -> Ppx_deriving_runtime.bool

val equal_protector : protector -> protector -> Ppx_deriving_runtime.bool

val equal_locality : locality -> locality -> Ppx_deriving_runtime.bool

module M (Symex : sig ... end) : sig ... end

module Concrete : sig ... end

Concrete implementation of tree borrows, for testing and debugging. Does not support compositionality, just lifts Raw to fit the state model shape.

module Raw : sig ... end

A simple Tree Borrows implementation. It doesn't support any sort of symbolic reasoning or anything fancy, it's just the raw logic.