Module Tree_borrows.Raw

A simple Tree Borrows implementation. It doesn't support any sort of symbolic reasoning or anything fancy, it's just the raw logic.

module Tag : sig ... end

type node = {

protector : Soteria_rust_lib__Tree_borrows__.Common.protector option;

parents : Tag.t list;

initial_state : Soteria_rust_lib__Tree_borrows__.Common.state;

}

Whether this node has a protector (this is distinct from having the protector toggled!), its parents (including this node's ID!), and its initial state if it doesn't exist in the state.

val pp_node : Ppx_deriving_runtime.Format.formatter -> node -> Ppx_deriving_runtime.unit

val show_node : node -> Ppx_deriving_runtime.string

val equal_node : node -> node -> Ppx_deriving_runtime.bool

val pp_state : Stdlib.Format.formatter -> Soteria_rust_lib__Tree_borrows__.Common.state -> unit

val meet : Soteria_rust_lib__Tree_borrows__.Common.state -> Soteria_rust_lib__Tree_borrows__.Common.state -> Soteria_rust_lib__Tree_borrows__.Common.state

val meet' : (bool * Soteria_rust_lib__Tree_borrows__.Common.state) -> (bool * Soteria_rust_lib__Tree_borrows__.Common.state) -> bool * Soteria_rust_lib__Tree_borrows__.Common.state

val transition : protected:bool -> Soteria_rust_lib__Tree_borrows__.Common.state -> (Soteria_rust_lib__Tree_borrows__.Common.locality * Soteria_rust_lib__Tree_borrows__.Common.access) -> Soteria_rust_lib__Tree_borrows__.Common.state

type t = node Tag.Map.t

val pp : Stdlib.Format.formatter -> node Tag.Map.t -> unit

val show : node Tag.Map.t -> string

val init : ?initial_state:Soteria_rust_lib__Tree_borrows__.Common.state -> unit -> Tag.t * node Tag.Map.t

val ub_state : Tag.t

val borrow : ?protector:Soteria_rust_lib__Tree_borrows__.Common.protector -> Tag.Map.key -> state:Soteria_rust_lib__Tree_borrows__.Common.state -> node Tag.Map.t -> node Tag.Map.t * Tag.t

val unprotect : Tag.Map.key -> node Tag.Map.t -> node Tag.Map.t

val strong_protector_exists : node Tag.Map.t -> bool

type tb_state = (bool * Soteria_rust_lib__Tree_borrows__.Common.state) Tag.Map.t

tag -> (protected * state), protected indicating the tag's protector (managed outside tb_state) was toggled.

val pp_tb_state : (bool * Soteria_rust_lib__Tree_borrows__.Common.state) Tag.Map.t Fmt.t

val empty_state : 'a Tag.Map.t

val is_empty_state : 'a Tag.Map.t -> bool

val equal_state : (bool * Soteria_rust_lib__Tree_borrows__.Common.state) Tag.Map.t -> (bool * Soteria_rust_lib__Tree_borrows__.Common.state) Tag.Map.t -> bool

val set_protector : protected:'a -> Tag.Map.key -> node Tag.Map.t -> ('a * Soteria_rust_lib__Tree_borrows__.Common.state) Tag.Map.t -> ('a * Soteria_rust_lib__Tree_borrows__.Common.state) Tag.Map.t

val access : Tag.Map.key -> Soteria_rust_lib__Tree_borrows__.Common.access -> t -> tb_state -> ((bool * Soteria_rust_lib__Tree_borrows__.Common.state) Tag.Map.t, [> `AliasingError ]) Stdlib.result

access accessed im e structure state: Update all nodes in the mapping state for the tree structure structure with an event e, that happened at accessed.

val merge : (bool * Soteria_rust_lib__Tree_borrows__.Common.state) Tag.Map.t -> (bool * Soteria_rust_lib__Tree_borrows__.Common.state) Tag.Map.t -> (bool * Soteria_rust_lib__Tree_borrows__.Common.state) Tag.Map.t