Module T.State

Tree Borrows state (the per-byte information)

type t

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

module SM : sig ... end

type syn

val pp_syn : Ppx_deriving_runtime.Format.formatter -> syn -> Ppx_deriving_runtime.unit

val show_syn : syn -> Ppx_deriving_runtime.string

val to_syn : t -> syn list

val ins_outs : syn -> Symex.Value.Expr.t list * Symex.Value.Expr.t list

val produce : syn -> t option -> t option Symex.Producer.t

val consume : syn -> t option -> (t option, syn list) Symex.Consumer.t

type syn_full := [

| `Structure of Tree.syn

| `State of syn

]

val fix_empty : unit -> syn list

Makes a fix to create an empty TB state

val init : unit -> t Symex.t

val equal : t option -> t option -> bool

Operations on the state

val set_protector : protected:bool -> Tag.t -> Tree.t option -> t option -> (t option, 'e, syn_full list) Symex.Result.t

val access : Tag.t -> Soteria_rust_lib__Tree_borrows__.Common.access -> Tree.t option -> t option -> (t option, [> `AliasingError ], syn_full list) Symex.Result.t

access root accessed e state: Update all nodes in the mapping state for the tree rooted at root with an event e, that happened at accessed.

val merge : t -> t -> t Symex.t

val assert_exclusively_owned : t option -> (unit, 'e, syn list) Symex.Result.t