Module Soteria_rust_lib.Typed

include module type of struct include Soteria.Bv_values.Typed end

Phantom types

module T = Soteria.Bv_values.Typed.T

type +'a ty = 'a Soteria.Bv_values.Typed.ty

Types

val pp_ty : (Stdlib.Format.formatter -> 'a ty -> unit) -> Stdlib.Format.formatter -> 'a ty -> unit

val ppa_ty : Stdlib.Format.formatter -> 'a ty -> unit

val equal_ty : 'a ty -> 'b ty -> bool

val t_bool : [> T.sbool ] ty

val t_int : int -> [> T.sint ] ty

val t_seq : ([< T.any ] as 'a) ty -> [> 'a T.sseq ] ty

val t_f16 : [> T.sfloat ] ty

val t_f32 : [> T.sfloat ] ty

val t_f64 : [> T.sfloat ] ty

val t_f128 : [> T.sfloat ] ty

Typed svalues

type +'a t = 'a Soteria.Bv_values.Typed.t

type sbool = T.sbool

Basic value operations

val get_ty : 'a t -> Soteria.Bv_values.Svalue.ty

val type_type : Soteria.Bv_values.Svalue.ty -> 'a ty

val untype_type : 'a ty -> Soteria.Bv_values.Svalue.ty

val kind : 'a t -> Soteria.Bv_values.Svalue.t_kind

val mk_var : Soteria.Symex.Var.t -> 'a ty -> 'a t

val iter_vars : 'a t -> ((Soteria.Symex.Var.t * 'b ty) -> unit) -> unit

val subst : (Soteria.Symex.Var.t -> Soteria.Symex.Var.t) -> 'a t -> 'a t

val type_ : Soteria.Bv_values.Svalue.t -> 'a t

val type_checked : Soteria.Bv_values.Svalue.t -> 'a ty -> 'a t option

val cast : 'a t -> 'b t

val is_float : 'a ty -> bool

val size_of_int : [< T.sint ] t -> int

val untyped : 'a t -> Soteria.Bv_values.Svalue.t

val untyped_list : 'a t list -> Soteria.Bv_values.Svalue.t list

val pp : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a t -> unit

val ppa : Stdlib.Format.formatter -> 'a t -> unit

val equal : ([< T.any ] as 'a) t -> 'a t -> bool

val compare : ([< T.any ] as 'a) t -> 'a t -> int

val hash : ('a -> int) -> 'a t -> int

val hasha : 'a t -> int

val unique_tag : [< T.any ] t -> int

Typed constructors

val sem_eq : 'a t -> 'a t -> sbool t

val sem_eq_untyped : 'a t -> 'a t -> [> sbool ] t

Boolean operations

module type Bool_ := sig ... end

include Bool_

val v_true : [> sbool ] t

val v_false : [> sbool ] t

val of_bool : bool -> [> sbool ] t

val to_bool : 'a t -> bool option

val and_ : [< sbool ] t -> [< sbool ] t -> [> sbool ] t

val and_lazy : [< sbool ] t -> (unit -> [< sbool ] t) -> [> sbool ] t

Similar to and_, but the rhs is only evaluated if the lhs is not the concrete false. In other words, this is a short-circuiting and. Avoids some errors, like a division by zero in 0 != x && n / x when x is 0.

val conj : [< sbool ] t list -> [> sbool ] t

val split_ands : [< sbool ] t -> ([> sbool ] t -> unit) -> unit

val or_ : [< sbool ] t -> [< sbool ] t -> [> sbool ] t

val or_lazy : [< sbool ] t -> (unit -> [< sbool ] t) -> [> sbool ] t

Similar to or_, but the rhs is only evaluated if the lhs is not the concrete true. In other words, this is a short-circuiting or. Avoids some errors, like a division by zero in 0 == x || n / x when x is 0.

val not : [< sbool ] t -> [> sbool ] t

val distinct : 'a t list -> [> sbool ] t

val ite : [< sbool ] t -> 'a t -> 'a t -> 'a t

module Bool = Soteria.Bv_values.Typed.Bool

Bit vector operations

Floating point operations

module SSeq = Soteria.Bv_values.Typed.SSeq

module Infix = Soteria.Bv_values.Typed.Infix

exception CastError of T.any t * T.any ty * T.any ty

CastError (value, expected, got)

val cast_error : [< `Bool | `Float | `List of T.any & T.any | `Loc | `NonZero | `Overflowed | `Ptr | `Zero ] t -> [< `Bool | `Float | `List of T.any & T.any | `Loc | `NonZero | `Overflowed | `Ptr | `Zero ] ty -> 'a

val t_ptr : unit -> [> T.sptr ] ty

val t_loc : unit -> [> T.sloc ] ty

val t_usize : unit -> [> T.sint ] ty

val t_lit : Charon.Types.literal_type -> [< T.cval Float NonZero Zero ] ty

val t_float : Charon.Types.float_type -> T.sfloat ty

val cast_checked : ty: ([< `Bool | `Float | `List of T.any & T.any | `Loc | `NonZero | `Overflowed | `Ptr | `Zero ] as 'a) ty -> [< `Bool | `Float | `List of T.any & T.any | `Loc | `NonZero | `Overflowed | `Ptr | `Zero ] t -> 'a t

val cast_checked2 : [< `Bool | `Float | `List of T.any & T.any | `Loc | `NonZero | `Overflowed | `Ptr | `Zero ] t -> 'a t -> 'b t * 'c t * 'd ty

val cast_lit : Charon.Types.literal_type -> [< `Bool | `Float | `List of T.any & T.any | `Loc | `NonZero | `Overflowed | `Ptr | `Zero ] t -> [< `Bool | `Float | `List of T.any & T.any | `Loc | `NonZero | `Overflowed | `Ptr | `Zero NonZero Zero ] t

val cast_i : Charon.Values.u_int_ty -> [< `Bool | `Float | `List of T.any & T.any | `Loc | `NonZero | `Overflowed | `Ptr | `Zero ] t -> [< `Bool | `Float | `List of T.any & T.any | `Loc | `NonZero | `Overflowed | `Ptr | `Zero NonZero Zero ] t

val cast_f : Charon.Types.float_type -> [< `Bool | `Float | `List of T.any & T.any | `Loc | `NonZero | `Overflowed | `Ptr | `Zero ] t -> T.sfloat t

val cast_float : [< `Bool | `Float | `List of T.any & T.any | `Loc | `NonZero | `Overflowed | `Ptr | `Zero ] t -> [> T.sfloat ] t

val cast_int : [< `Bool | `Float | `List of T.any & T.any | `Loc | `NonZero | `Overflowed | `Ptr | `Zero ] t -> [> T.sint ] t * int

DEPRECATED: it is unlikely you need this; the interpreter should be well typed

module BitVec : sig ... end

module BV = BitVec

module Float : sig ... end

module Ptr : sig ... end

module Syntax : sig ... end