Module Value.Expr

Expr.t describes a syntactic representation of a value of type 'a v. Syntactic objects, unlike semantic values, can contain "free" variables and operations such as substitution have meaning.

Expr.t objects are used in, for instance, assertions (Logic). They can be substituted through a Symex.S.Producer monad or learned through a Symex.S.Consumer monad.

type t

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

val of_value : 'a t -> t

Obtain a syntactic representation from a semantic value. This implicitly uses an identity substitution.

val ty : t -> 'a ty

Gets the type associated to a syntactic values.

val subst : (t -> 'a t) -> t -> 'b t

Convenience function

module Subst : sig ... end

A susbtitution projects the syntactic to the semantic world.