Parameter Encoder.Sptr

type t

pointer type

include Sptr.D_abstr.S_with_syn with type t := t

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

val fresh : unit -> t Sptr.DecayMap.SM.t

type syn

val pp_syn : Ppx_deriving_runtime.Format.formatter -> syn -> Ppx_deriving_runtime.unit

val show_syn : syn -> Ppx_deriving_runtime.string

val to_syn : t -> syn

val subst : (Sptr.DecayMap.SM.Value.Expr.t -> 'a Sptr.DecayMap.SM.Value.t) -> syn -> t

val learn_eq : syn -> t -> (unit, 'a) Sptr.DecayMap.SM.Consumer.t

val exprs_syn : syn -> Sptr.DecayMap.SM.Value.Expr.t list

val of_address : [< Soteria_rust_lib.Typed.T.sint ] Typed.t -> t

Converts an address into a pointer, without provenance.

val null : unit -> t

The null pointer, which always decays to 0, and has no provenance. Equivalent to of_address 0.

val is_null : t -> Soteria_rust_lib.Typed.T.sbool Typed.t

Whether this is the null pointer, meaning it always decays to 0.

val has_provenance : t -> Soteria_rust_lib.Typed.T.sbool Typed.t

Whether this pointer has provenance, i.e. points to some allocation.

val have_same_provenance : t -> t -> Soteria_rust_lib.Typed.T.sbool Typed.t

If these two pointers have the same provenance, i.e. point to the same allocation (or if they both have no provenance).

val distance : t -> t -> Soteria_rust_lib.Typed.T.sint Typed.t Sptr.DecayMap.SM.t

The distance, in bytes, between two pointers; if they point to different allocations, they are decayed and substracted.

val nondet : Charon.Types.ty -> (t, [> Error.t ], 'a) Rustsymex.Result.t

Generates a nondeterministic pointer, that is valid for accesses to the given type. The location of the pointer is nondeterministic; it could point to any allocation.

val decay : t -> [> Soteria_rust_lib.Typed.T.sint ] Typed.t Sptr.DecayMap.SM.t

Decay a pointer into an integer value, losing provenance. This does not expose the address of the allocation; for that, use expose

val expose : t -> [> Soteria_rust_lib.Typed.T.sint ] Typed.t Sptr.DecayMap.SM.t

Decay a pointer into an integer value, exposing the address of the allocation, allowing it to be retrieved with DecayMapS.from_exposed later.

val as_id : t -> [> Soteria_rust_lib.Typed.T.sint ] Typed.t

For Miri: the allocation ID of this location, as a u64

val allocation_info : t -> [> Soteria_rust_lib.Typed.T.sint ] Typed.t * [> Soteria_rust_lib.Typed.T.nonzero ] Typed.t

For Miri: get the allocation info for this pointer: its size and alignment