Module Soteria_rust_lib.Builtins

module Alloc : sig ... end

Builtins related to the default allocator.

module Core : sig ... end

Core operations, that may be used in several parts of the interpreter, so we share them here for convenience.

module Eval : sig ... end

module Intrinsics : sig ... end

This file was generated with scripts/intrinsics.py -- do not edit it manually, instead modify the script and re-run it.

module Intrinsics_impl : sig ... end

module Intrinsics_intf : sig ... end

This file was generated with scripts/intrinsics.py -- do not edit it manually, instead modify the script and re-run it.

module Intrinsics_stubs : sig ... end

This file was generated with scripts/intrinsics.py -- do not edit it manually, instead modify the script and re-run it.

module Miri : sig ... end

Miri-related intrinsics.

module Soteria_lib : sig ... end

Builtins for Soteria Rust; provide the building blocks necessary for symbolic execution!

module Std : sig ... end

Builtins, for functions we stub manually, either because the originals suffer from poor performance, or are not properly computable in our engine.

module System : sig ... end

Builtins relating to particular operating systems. There is no central documentation for these; we must instead implement them as needed.