Module Intrinsics_impl.M

Parameters

module State : State_intf.S

Signature

module Sptr = State.Sptr

module Core : sig ... end

module Encoder : sig ... end

type nonrec rust_val = Sptr.t Rust_val.rust_val

type ret = (rust_val * State.t, Error.t State.err * State.t, State.serialized) Soteria_rust_lib.Rustsymex.Result.t

val as_ptr : ?null_ok:bool -> rust_val -> Sptr.t Rust_val.full_ptr Rustsymex.t

val as_base : ?ty:'ty Soteria.C_values.Typed.ty -> rust_val -> 'ty Soteria.C_values.Typed.t Rustsymex.t

val as_float : rust_val -> [> Soteria.C_values.Typed.T.sfloat ] Soteria.C_values.Typed.t Rustsymex.t

val (&&*) : 'a Soteria_rust_lib__Rustsymex.SYMEX.t -> 'b Soteria_rust_lib__Rustsymex.SYMEX.t -> ('a * 'b) Soteria_rust_lib__Rustsymex.SYMEX.t

val abort : State.t -> ('a, [< Error.t Panic ] State.err * State.t, State.serialized) Soteria_rust_lib.Rustsymex.Result.t

val checked_op : Charon.Expressions.binop -> t:Charon.Types.ty -> x:rust_val -> y:rust_val -> State.t -> ret

val add_with_overflow : t:Charon.Types.ty -> x:rust_val -> y:rust_val -> State.t -> ret

val sub_with_overflow : t:Charon.Types.ty -> x:rust_val -> y:rust_val -> State.t -> ret

val mul_with_overflow : t:Charon.Types.ty -> x:rust_val -> y:rust_val -> State.t -> ret

val align_of : t:Charon.Types.ty -> 'a -> ('b Rust_val.t * 'a, 'c, 'd) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val align_of_val : t:Charon.Types.ty -> ptr:'a -> 'b -> ('c Rust_val.t * 'b, 'd, 'e) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val ptr_add : check:bool -> t:Charon.Types.ty -> dst:Sptr.t Rust_val.t -> offset:rust_val -> State.t -> ret

val arith_offset : t:Charon.Types.ty -> dst:Sptr.t Rust_val.t -> offset:rust_val -> State.t -> ret

val offset : ptr:Charon.Types.ty -> delta:'a -> dst:Sptr.t Rust_val.t -> offset:rust_val -> State.t -> ret

val assert_inhabited : t:Charon.Types.ty -> State.t -> ('a Rust_val.t * State.t, [< Error.t Panic ] State.err * State.t, State.serialized) Soteria_rust_lib.Rustsymex.Result.t

val assert_mem_uninitialized_valid : t:'a -> 'b -> ('c Rust_val.t * 'b, 'd, 'e) Soteria_rust_lib.Rustsymex.Result.t

val assert_zero_valid : t:Charon.Types.ty -> State.t -> ('a Rust_val.t * State.t, [< Error.t Panic ] State.err * State.t, State.serialized) Soteria_rust_lib.Rustsymex.Result.t

val assume : b:rust_val -> State.t -> ('a Rust_val.t * State.t, [< Error.t StdErr ] State.err * State.t, State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val black_box : t:'a -> dummy:'b -> 'c -> ('b * 'c, 'd, 'e) Soteria_rust_lib.Rustsymex.Result.t

val breakpoint : State.t -> ret

val bswap : t:Charon.Types.ty -> x:Sptr.t Rust_val.t -> State.t -> (Sptr.t Rust_val.t * State.t, [< Error.t RefToUninhabited UBDanglingPointer UBTransmute UninitializedMemoryAccess ] State.err * State.t, 'a) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val catch_unwind : (Charon.UllbcAst.blocks Charon.UllbcAst.gfun_decl -> args:Sptr.t Rust_val.t list -> State.t -> ('a * 'b, Error.t State.err * State.t, State.serialized) Soteria_rust_lib.Rustsymex.Result.t) -> _try_fn:rust_val -> _data:Sptr.t Rust_val.t -> _catch_fn:rust_val -> State.t -> ('c Rust_val.t * 'b, Error.t State.err * State.t, State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val float_rounding : Soteria.C_values.Svalue.FloatRoundingMode.t -> x:rust_val -> State.t -> ret

val ceilf16 : x:rust_val -> State.t -> ret

val ceilf32 : x:rust_val -> State.t -> ret

val ceilf64 : x:rust_val -> State.t -> ret

val ceilf128 : x:rust_val -> State.t -> ret

val floorf16 : x:rust_val -> State.t -> ret

val floorf32 : x:rust_val -> State.t -> ret

val floorf64 : x:rust_val -> State.t -> ret

val floorf128 : x:rust_val -> State.t -> ret

val round_ties_even_f16 : x:rust_val -> State.t -> ret

val round_ties_even_f32 : x:rust_val -> State.t -> ret

val round_ties_even_f64 : x:rust_val -> State.t -> ret

val round_ties_even_f128 : x:rust_val -> State.t -> ret

val roundf16 : x:rust_val -> State.t -> ret

val roundf32 : x:rust_val -> State.t -> ret

val roundf64 : x:rust_val -> State.t -> ret

val roundf128 : x:rust_val -> State.t -> ret

val truncf16 : x:rust_val -> State.t -> ret

val truncf32 : x:rust_val -> State.t -> ret

val truncf64 : x:rust_val -> State.t -> ret

val truncf128 : x:rust_val -> State.t -> ret

val cold_path : 'a -> ('b Rust_val.t * 'a, 'c, 'd) Soteria_rust_lib.Rustsymex.Result.t

val compare_bytes : left:rust_val -> right:rust_val -> bytes:rust_val -> State.t -> ('a Rust_val.t * State.t, [< Error.t AliasingError MisalignedPointer NullDereference OutOfBounds RefToUninhabited UBDanglingPointer UBTransmute UninitializedMemoryAccess UseAfterFree ] State.err * State.t, State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val copy_ : bool -> t:Charon.Types.ty -> src:rust_val -> dst:rust_val -> count:rust_val -> State.t -> ret

val copy : t:Charon.Types.ty -> src:rust_val -> dst:rust_val -> count:rust_val -> State.t -> ret

val copy_nonoverlapping : t:Charon.Types.ty -> src:rust_val -> dst:rust_val -> count:rust_val -> State.t -> ret

val copy_sign : x:rust_val -> y:rust_val -> 'a -> ('b Rust_val.t * 'a, 'c, 'd) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val copysignf128 : x:rust_val -> y:rust_val -> 'a -> ('b Rust_val.t * 'a, 'c, 'd) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val copysignf64 : x:rust_val -> y:rust_val -> 'a -> ('b Rust_val.t * 'a, 'c, 'd) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val copysignf32 : x:rust_val -> y:rust_val -> 'a -> ('b Rust_val.t * 'a, 'c, 'd) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val copysignf16 : x:rust_val -> y:rust_val -> 'a -> ('b Rust_val.t * 'a, 'c, 'd) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val binary_int_operation : concrete: (int -> Z.t -> [< Soteria_rust_lib.Rust_val.T.cval ] Soteria__C_values__Typed.t) -> symbolic: (int -> Soteria.C_values.Typed.T.sint Soteria.C_values.Typed.t -> [< Soteria_rust_lib.Rust_val.T.cval ] Soteria__C_values__Typed.t) -> t:Charon.Types.ty -> x:rust_val -> State.t -> ret

Converts the input value to its unsigned representation, and either calls concrete or symbolic with it accordingly.

val ctpop : t:Charon.Types.ty -> x:rust_val -> State.t -> ret

val cttz : t:Charon.Types.ty -> x:rust_val -> State.t -> ret

val cttz_nonzero : t:Charon.Types.ty -> x:rust_val -> State.t -> (rust_val * State.t, Error.t State.err * State.t, State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val discriminant_value : t:Charon.Types.ty -> v:rust_val -> State.t -> ('a Rust_val.t * State.t, [> `AliasingError | `MisalignedPointer | `NullDereference | `OutOfBounds | `RefToUninhabited | `UBDanglingPointer | `UBTransmute | `UninitializedMemoryAccess | `UseAfterFree ] State.err * State.t, State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val exact_div : t:Charon.Types.ty -> x:rust_val -> y:rust_val -> State.t -> ('a Rust_val.t * State.t, [< Error.t DivisionByZero InvalidShift Overflow StdErr UBPointerArithmetic ] State.err * State.t, State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val abs : x:rust_val -> 'a -> ('b Rust_val.t * 'a, 'c, 'd) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val fabsf16 : x:rust_val -> 'a -> ('b Rust_val.t * 'a, 'c, 'd) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val fabsf32 : x:rust_val -> 'a -> ('b Rust_val.t * 'a, 'c, 'd) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val fabsf64 : x:rust_val -> 'a -> ('b Rust_val.t * 'a, 'c, 'd) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val fabsf128 : x:rust_val -> 'a -> ('b Rust_val.t * 'a, 'c, 'd) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val float_fast : Charon.Expressions.binop -> t:Charon.Types.ty -> a:rust_val -> b:rust_val -> State.t -> ret

val fadd_fast : t:Charon.Types.ty -> a:rust_val -> b:rust_val -> State.t -> ret

val fdiv_fast : t:Charon.Types.ty -> a:rust_val -> b:rust_val -> State.t -> ret

val fmul_fast : t:Charon.Types.ty -> a:rust_val -> b:rust_val -> State.t -> ret

val frem_fast : t:Charon.Types.ty -> a:rust_val -> b:rust_val -> State.t -> ret

val fsub_fast : t:Charon.Types.ty -> a:rust_val -> b:rust_val -> State.t -> ret

val float_to_int_unchecked : float:'a -> int:Charon.Types.ty -> value:rust_val -> State.t -> ('b Rust_val.t * State.t, [< Error.t StdErr ] State.err * State.t, State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val fmul_add : a:rust_val -> b:rust_val -> c:rust_val -> 'a -> ('b Rust_val.t * 'a, 'c, 'd) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val fmaf16 : a:rust_val -> b:rust_val -> c:rust_val -> 'a -> ('b Rust_val.t * 'a, 'c, 'd) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val fmaf32 : a:rust_val -> b:rust_val -> c:rust_val -> 'a -> ('b Rust_val.t * 'a, 'c, 'd) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val fmaf64 : a:rust_val -> b:rust_val -> c:rust_val -> 'a -> ('b Rust_val.t * 'a, 'c, 'd) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val fmaf128 : a:rust_val -> b:rust_val -> c:rust_val -> 'a -> ('b Rust_val.t * 'a, 'c, 'd) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val fmuladdf16 : a:rust_val -> b:rust_val -> c:rust_val -> 'a -> ('b Rust_val.t * 'a, 'c, 'd) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val fmuladdf32 : a:rust_val -> b:rust_val -> c:rust_val -> 'a -> ('b Rust_val.t * 'a, 'c, 'd) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val fmuladdf64 : a:rust_val -> b:rust_val -> c:rust_val -> 'a -> ('b Rust_val.t * 'a, 'c, 'd) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val fmuladdf128 : a:rust_val -> b:rust_val -> c:rust_val -> 'a -> ('b Rust_val.t * 'a, 'c, 'd) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val forget : t:'a -> arg:'b -> 'c -> ('d Rust_val.t * 'c, 'e, 'f) Soteria_rust_lib.Rustsymex.Result.t

val is_val_statically_known : t:'a -> _arg:'b -> 'c -> ('d Rust_val.t * 'c, 'e, 'f) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val likely : b:'a -> 'b -> ('a * 'b, 'c, 'd) Soteria_rust_lib.Rustsymex.Result.t

val unlikely : b:'a -> 'b -> ('a * 'b, 'c, 'd) Soteria_rust_lib.Rustsymex.Result.t

val float_minmax : is_min:bool -> x:rust_val -> y:rust_val -> State.t -> ret

val minnumf16 : x:rust_val -> y:rust_val -> State.t -> ret

val minnumf32 : x:rust_val -> y:rust_val -> State.t -> ret

val minnumf64 : x:rust_val -> y:rust_val -> State.t -> ret

val minnumf128 : x:rust_val -> y:rust_val -> State.t -> ret

val maxnumf16 : x:rust_val -> y:rust_val -> State.t -> ret

val maxnumf32 : x:rust_val -> y:rust_val -> State.t -> ret

val maxnumf64 : x:rust_val -> y:rust_val -> State.t -> ret

val maxnumf128 : x:rust_val -> y:rust_val -> State.t -> ret

val ptr_guaranteed_cmp : t:'a -> ptr:Core.Sptr.t Rust_val.t -> other:Core.Sptr.t Rust_val.t -> State.t -> ('b Rust_val.t * State.t, [< Error.t UBPointerComparison ] State.err * State.t, 'c) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val ptr_offset_from_ : unsigned:bool -> t:Charon.Types.ty -> ptr:rust_val -> base:rust_val -> State.t -> ret

val ptr_offset_from : t:Charon.Types.ty -> ptr:rust_val -> base:rust_val -> State.t -> ret

val ptr_offset_from_unsigned : t:Charon.Types.ty -> ptr:rust_val -> base:rust_val -> State.t -> ret

val raw_eq : t:Charon.Types.ty -> a:rust_val -> b:rust_val -> State.t -> ('a Rust_val.t * State.t, [> `AliasingError | `MisalignedPointer | `NullDereference | `OutOfBounds | `RefToUninhabited | `UBDanglingPointer | `UBTransmute | `UninitializedMemoryAccess | `UseAfterFree ] State.err * State.t, State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val saturating : Charon.Expressions.binop -> t:Charon.Types.ty -> a:rust_val -> b:rust_val -> State.t -> ret

val saturating_add : t:Charon.Types.ty -> a:rust_val -> b:rust_val -> State.t -> ret

val saturating_sub : t:Charon.Types.ty -> a:rust_val -> b:rust_val -> State.t -> ret

val size_of : t:Charon.Types.ty -> 'a -> ('b Rust_val.t * 'a, 'c, 'd) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val size_of_val : t:Charon.Types.ty -> ptr:'a Rust_val.t -> 'b -> ('c Rust_val.t * 'b, 'd, 'e) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val transmute : t_src:Charon.Types.ty -> dst:Charon.Types.ty -> src:Sptr.t Rust_val.t -> State.t -> (Sptr.t Rust_val.t * State.t, [< Error.t RefToUninhabited UBDanglingPointer UBTransmute UninitializedMemoryAccess ] State.err * State.t, 'a) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val type_id : t:'a -> 'b -> ('c Rust_val.t * 'b, 'd, 'e) Soteria_rust_lib.Rustsymex.Result.t

val type_name : t:Charon.Types.ty -> State.t -> (State.Sptr.t Rust_val.t * State.t, [> `AliasingError | `MisalignedPointer | `NullDereference | `OutOfBounds | `UBDanglingPointer | `UseAfterFree ] State.err * State.t, State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val typed_swap_nonoverlapping : t:Charon.Types.ty -> x:rust_val -> y:rust_val -> State.t -> ('a Rust_val.t * State.t, [< Error.t AliasingError MisalignedPointer NullDereference OutOfBounds RefToUninhabited StdErr UBDanglingPointer UBTransmute UninitializedMemoryAccess UseAfterFree ] State.err * State.t, State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t

val unchecked_op : Charon.Expressions.binop -> t:Charon.Types.ty -> x:rust_val -> y:rust_val -> State.t -> ret

val unchecked_add : t:Charon.Types.ty -> x:rust_val -> y:rust_val -> State.t -> ret

val unchecked_div : t:Charon.Types.ty -> x:rust_val -> y:rust_val -> State.t -> ret

val unchecked_mul : t:Charon.Types.ty -> x:rust_val -> y:rust_val -> State.t -> ret

val unchecked_rem : t:Charon.Types.ty -> x:rust_val -> y:rust_val -> State.t -> ret

val unchecked_shl : t:Charon.Types.ty -> u:'a -> x:rust_val -> y:rust_val -> State.t -> ret

val unchecked_shr : t:Charon.Types.ty -> u:'a -> x:rust_val -> y:rust_val -> State.t -> ret

val unchecked_sub : t:Charon.Types.ty -> x:rust_val -> y:rust_val -> State.t -> ret

val variant_count : t:Charon.Types.ty -> State.t -> ('a Rust_val.t * State.t, [< Error.t StdErr ] State.err * State.t, State.serialized) Soteria_rust_lib.Rustsymex.Result.t

val wrapping_op : Charon.Expressions.binop -> t:Charon.Types.ty -> a:rust_val -> b:rust_val -> State.t -> ret

val wrapping_add : t:Charon.Types.ty -> a:rust_val -> b:rust_val -> State.t -> ret

val wrapping_mul : t:Charon.Types.ty -> a:rust_val -> b:rust_val -> State.t -> ret

val wrapping_sub : t:Charon.Types.ty -> a:rust_val -> b:rust_val -> State.t -> ret

val write_bytes : t:Charon.Types.ty -> dst:rust_val -> val_:rust_val -> count:rust_val -> State.t -> ('a Rust_val.t * State.t, [< Error.t AliasingError MisalignedPointer NullDereference OutOfBounds UBDanglingPointer UseAfterFree ] State.err * State.t, State.serialized) Soteria.Symex.Compo_res.t Soteria_rust_lib__Rustsymex.SYMEX.t