Module Soteria_rust_lib.Config

type t = {

cleanup : bool;
(*
Clean up compiled files after execution
*)

log_compilation : bool;
(*
Log the compilation process
*)

monomorphize_old : bool;
(*
Use Charon's old monomorphization, which may be slower and less powerful.
*)

no_compile : bool;
(*
Do not compile the Rust code, as it is already compiled
*)

no_compile_plugins : bool;
(*
Do not compile the plugins, as they are already compiled
*)

target : string option;
(*
The compilation target triple to use, e.g. x86_64-unknown-linux-gnu. If not provided, the default target for the current machine is used.
*)

output_crate : bool;
(*
Pretty-print the compiled crate to a file
*)

rustc_flags : string list;
(*
Additional flags to pass to the Rustc compiler
*)

with_obol : bool;
(*
Compile the code using Obol, rather than Charon
*)

with_kani : bool;
(*
Use the Kani library
*)

with_miri : bool;
(*
Use the Miri library
*)

filter : string list;
(*
Filter the entrypoints to run, by name. If empty, all entrypoints are run. Multiple filters can be provided; tests matching any will be selected. The filters are treated as regexes.
*)

no_timing : bool;
(*
Do not display execution times
*)

print_summary : bool;
(*
If a summary of all test cases should be printed at the end of execution
*)

ignore_leaks : bool;
(*
Ignore memory leaks
*)

ignore_aliasing : bool;
(*
Ignore pointer aliasing rules (tree borrows)
*)

step_fuel : int option;
(*
The default step fuel for each entrypoint -- every control flow jump counts as one fuel. Defaults to infinite fuel.
*)

branch_fuel : int option;
(*
The default branch fuel for each entrypoint -- every symbolic execution branching point counts as one fuel. Defaults to infinite fuel.
*)

}

val make : ?cleanup:bool -> ?log_compilation:bool -> ?monomorphize_old:bool -> ?no_compile:bool -> ?no_compile_plugins:bool -> ?target:string -> ?output_crate:bool -> ?rustc_flags:string list -> ?with_obol:bool -> ?with_kani:bool -> ?with_miri:bool -> ?filter:string list -> ?no_timing:bool -> ?print_summary:bool -> ?ignore_leaks:bool -> ?ignore_aliasing:bool -> ?step_fuel:int -> ?branch_fuel:int -> unit -> t

val cmdliner_term : unit -> t Cmdliner.Term.t

val term : t Cmdliner.Term.t

type global = {

logs : (Soteria.Logs.Config.t, string) Stdlib.result;

terminal : Soteria.Terminal.Config.t;

solver : Soteria.Solvers.Config.t;

rusteria : t;

}

val make_global : logs:(Soteria.Logs.Config.t, string) Stdlib.result -> terminal:Soteria.Terminal.Config.t -> solver:Soteria.Solvers.Config.t -> rusteria:t -> global

val global_cmdliner_term : unit -> global Cmdliner.Term.t

val global_term : global Cmdliner.Term.t

val default : t

val current : t Stdlib.ref

val set : global -> unit