Module Soteria_rust_lib.Config

type frontend =

| Charon

| Obol

val frontend_cmdliner_conv : unit -> frontend Cmdliner.Arg.conv

val pp_frontend : Ppx_deriving_runtime.Format.formatter -> frontend -> Ppx_deriving_runtime.unit

val show_frontend : frontend -> Ppx_deriving_runtime.string

type provenance =

| Strict

| Permissive

val provenance_cmdliner_conv : unit -> provenance Cmdliner.Arg.conv

type t = {

cleanup : bool;
(*
Clean up compiled files after execution
*)

log_compilation : bool;
(*
Log the compilation process
*)

no_compile : bool;
(*
Do not compile the Rust code, as it is already compiled
*)

no_compile_plugins : bool;
(*
Do not compile the plugins, as they are already compiled
*)

plugin_directory : string option;
(*
The directory in which plugins are and should be compiled; defaults to the current dune-managed site.
*)

target : string option;
(*
The compilation target triple to use, e.g. x86_64-unknown-linux-gnu. If not provided, the default target for the current machine is used.
*)

output_crate : bool;
(*
Pretty-print the compiled crate to a file
*)

rustc_flags : string list;
(*
Additional flags to pass to the Rustc compiler
*)

frontend : frontend;
(*
Choose the frontend to use: Charon or Obol
*)

sysroot : string option;
(*
The sysroot to use for compilation. If not provided, the default sysroot is used.
*)

with_kani : bool;
(*
Use the Kani library
*)

with_miri : bool;
(*
Use the Miri library
*)

filter : string list;
(*
Filter the entrypoints to run, by name. If empty, all entrypoints are run. Multiple filters can be provided; tests matching any will be selected. The filters are treated as regexes.
*)

print_summary : bool;
(*
If a summary of all test cases should be printed at the end of execution
*)

ignore_leaks : bool;
(*
Ignore memory leaks
*)

ignore_aliasing : bool;
(*
Ignore pointer aliasing rules (tree borrows)
*)

provenance : provenance;
(*
The provenance model to use for pointers. If not provided, the default is permissive.
*)

step_fuel : int option;
(*
The default step fuel for each entrypoint -- every control flow jump counts as one fuel. Defaults to infinite fuel.
*)

branch_fuel : int option;
(*
The default branch fuel for each entrypoint -- every symbolic execution branching point counts as one fuel. Defaults to infinite fuel.
*)

}

val make : ?cleanup:bool -> ?log_compilation:bool -> ?no_compile:bool -> ?no_compile_plugins:bool -> ?plugin_directory:string -> ?target:string -> ?output_crate:bool -> ?rustc_flags:string list -> ?frontend:frontend -> ?sysroot:string -> ?with_kani:bool -> ?with_miri:bool -> ?filter:string list -> ?print_summary:bool -> ?ignore_leaks:bool -> ?ignore_aliasing:bool -> ?provenance:provenance -> ?step_fuel:int -> ?branch_fuel:int -> unit -> t

val cmdliner_term : unit -> t Cmdliner.Term.t

val term : t Cmdliner.Term.t

val default : t

val get : unit -> t

val set_and_lock : t -> unit

type global = {

logs : (Soteria.Logs.Config.t, string) Stdlib.result;

terminal : Soteria.Terminal.Config.t;

solver : Soteria.Solvers.Config.t;

stats : Soteria.Stats.Config.t;

soteria_rust : t;

}

val make_global : logs:(Soteria.Logs.Config.t, string) Stdlib.result -> terminal:Soteria.Terminal.Config.t -> solver:Soteria.Solvers.Config.t -> stats:Soteria.Stats.Config.t -> soteria_rust:t -> global

val global_cmdliner_term : unit -> global Cmdliner.Term.t

val global_term : global Cmdliner.Term.t

val set_and_lock_global : global -> unit