Module Soteria_rust_lib.Config

type t = {

no_compile : bool;
(*
Do not compile the Rust code, as it is already compiled
*)

no_timing : bool;
(*
Do not display execution times
*)

cleanup : bool;
(*
Clean up compiles files after execution
*)

ignore_leaks : bool;
(*
Ignore memory leaks
*)

ignore_aliasing : bool;
(*
Ignore pointer aliasing rules (tree borrows)
*)

monomorphize_experimental : bool;
(*
Use Charon's new monomorphization, which may cause unexpected results but resolves drops.
*)

with_kani : bool;
(*
Use the Kani library
*)

with_miri : bool;
(*
Use the Miri library
*)

log_compilation : bool;
(*
Log the compilation process
*)

step_fuel : int;
(*
The default step fuel for each entrypoint -- every control flow jump counts as one fuel
*)

branch_fuel : int;
(*
The default branch fuel for each entrypoint -- every symbolic execution branching point counts as one fuel
*)

}

val make : ?no_compile:bool -> ?no_timing:bool -> ?cleanup:bool -> ?ignore_leaks:bool -> ?ignore_aliasing:bool -> ?monomorphize_experimental:bool -> ?with_kani:bool -> ?with_miri:bool -> ?log_compilation:bool -> ?step_fuel:int -> ?branch_fuel:int -> unit -> t

val cmdliner_term : unit -> t Cmdliner.Term.t

type global = {

logs : (Soteria_logs.Config.t, string) Stdlib.result;

terminal : Soteria_terminal.Config.t;

solver : Soteria_c_values.Solver_config.t;

rusteria : t;

}

val make_global : logs:(Soteria_logs.Config.t, string) Stdlib.result -> terminal:Soteria_terminal.Config.t -> solver:Soteria_c_values.Solver_config.t -> rusteria:t -> global

val default : t

val current : t Stdlib.ref

val set : t -> unit