Module Core.M

Parameters

module State : State_intf.S

Signature

module Sptr = State.Sptr

type nonrec rust_val = Sptr.t Rust_val.rust_val

val pp_rust_val : Sptr.t Rust_val.t Fmt.t

val cmp : signed:bool -> [< Soteria__Bv_values__Typed.T.sint ] Soteria__Bv_values__Typed.t -> [< Soteria__Bv_values__Typed.T.sint ] Soteria__Bv_values__Typed.t -> Soteria_rust_lib__Sptr.DecayMap.t option -> ([> Soteria__Bv_values__Typed.T.sint ] Soteria__Bv_values__Typed.t * Soteria_rust_lib__Sptr.DecayMap.t option) Rustsymex.t

val cmp_of_int : [< Soteria__Bv_values__Typed.T.sint ] Typed.t -> Soteria_rust_lib__Sptr.DecayMap.t option -> ([> Soteria__Bv_values__Typed.T.sint ] Soteria__Bv_values__Typed.t * Soteria_rust_lib__Sptr.DecayMap.t option) Rustsymex.t

val equality_check : ([< `NonZero | `Ptr | `Zero ] as 'a) Typed.t -> 'b Typed.t -> ([> Soteria__Bv_values__Typed.T.sint ] Soteria.Bv_values.Typed.t, [> `UBPointerComparison ], 'c) Sptr.DecayMapMonad.Result.t

val normalise_shift_r : Charon.Expressions.binop -> 'a Typed.t -> ([> Soteria__Bv_values__Typed.T.sint ] as 'b) Typed.t -> 'b Typed.t

Rust allows shift operations on integers of differents sizes, which isn't possible in SMT-Lib, so we normalise the righthand side to match the left hand side.

val eval_lit_binop : Charon.Expressions.binop -> Charon.Types.literal_type -> [< Soteria__Bv_values__Typed.T.sint ] Typed.t -> [ `NonZero | `Zero ] Typed.t -> ([> Soteria__Bv_values__Typed.T.sint_ovf ] Typed.t, [> `DivisionByZero | `InvalidShift | `Overflow ], 'f) Sptr.DecayMapMonad.Result.t

Evaluates a binary operator of +,-,/,*,rem, and ensures the result is within the type's constraints, else errors

val eval_checked_lit_binop : Charon.Expressions.binop -> Charon.Types.literal_type -> [< Soteria_rust_lib.Typed.T.any ] Typed.t -> [< Soteria_rust_lib.Typed.T.any ] Typed.t -> ('a Rust_val.t, 'b, 'c) Sptr.DecayMapMonad.Result.t

Evaluates the checked operation, returning (wrapped value, overflowed).

val meta_as_int : Sptr.t Rust_val.meta -> Soteria_rust_lib.Rust_val.T.sint Typed.t option Sptr.DecayMapMonad.t

val eval_meta_eq : Sptr.t Rust_val.meta -> Sptr.t Rust_val.meta -> Soteria_rust_lib__Sptr.DecayMap.t option -> ([> Typed.sbool ] Typed.t * Soteria_rust_lib__Sptr.DecayMap.t option) Rustsymex.t

val eval_ptr_binop : Charon.Expressions.binop -> Sptr.t Rust_val.t -> Sptr.t Rust_val.t -> Soteria_rust_lib__Sptr.DecayMap.t option -> (([> Soteria__Bv_values__Typed.T.sint ] Typed.t, [> `UBPointerComparison ], 'a) Soteria.Symex.Compo_res.t * Soteria_rust_lib__Sptr.DecayMap.t option) Rustsymex.t