Module Encoder.Make

Parameters

module Sptr : Sptr.S

Signature

type nonrec rust_val = Sptr.t Rust_val.t

val pp_rust_val : Sptr.t Rust_val.t Fmt.t

module ParserMonad : sig ... end

val iter_fields : ?variant:Charon.Types.VariantId.id -> ?meta:'a Rust_val.meta -> Layout.t -> Charon.Types.ty -> (Charon.Types.ty * Soteria_rust_lib.Rust_val.T.sint Typed.t) Soteria.Soteria_std.Iter.t

Iterator over the fields and offsets of a type; for primitive types, returns a singleton iterator for that value.

val encode : offset:Soteria_rust_lib.Rust_val.T.sint Typed.t -> rust_val -> Charon.Types.ty -> ((rust_val * Soteria_rust_lib.Rust_val.T.sint Typed.t) Soteria.Soteria_std.Iter.t, [> `InvalidLayout of Charon.Types.ty ], 'f) Rustsymex.Result.t

encode ?offset v ty Converts a Rust_val.t of type ty into an iterator over its sub values, along with their offset. Offsets all blocks by offset if specified

val variant_of_enum : offset:[< Soteria__Bv_values__Typed.T.sint ] Soteria__Bv_values__Typed.t -> Charon.Types.ty -> (Charon.Types.variant_id, 'state, [> `InvalidLayout of Charon.Types.ty | `UBTransmute of string ], 'fix) ParserMonad.t

Parses the current variant of the enum at the given offset. This handles cases such as niches, where the discriminant isn't directly encoded as a tag.

val decode : meta:'a Rust_val.meta -> offset:Soteria_rust_lib.Rust_val.T.sint Typed.t -> Charon.Types.ty -> (rust_val, 'state, [> `InvalidLayout of Charon.Types.ty | `RefToUninhabited | `UBTransmute of string ], 'fix) ParserMonad.t

decode ~meta ~offset ty Parses a rust value of type ty at the given offset, using the provided metadata for DSTs, and returns the associated Rust_val. This does not perform any validity checking, aside from erroring if the type is uninhabited.

val check_valid : fake_read: ((Sptr.t * Sptr.t Rust_val.meta) -> Charon.Types.ty -> 'a -> (([> `InvalidLayout of Charon.Types.ty | `RefToUninhabited | `UBDanglingPointer | `UBTransmute of string ] as 'b) option * 'a) Soteria_rust_lib__Rustsymex.t) -> Sptr.t Rust_val.t -> Charon.Types.ty -> 'a -> ('a, 'b, 'c) Rustsymex.Result.t

Ensures this value is valid for the given type. This includes checking pointer metadata, e.g. slice lengths and vtables. The fake_read function is used to simulate reading from memory to check the validity of a pointee type.

val with_constraints : ty:Charon.Types.literal_type -> Soteria_rust_lib.Rust_val.T.sint Typed.t -> Soteria_rust_lib__Sptr.DecayMap.t option -> (('a Rust_val.t, [> `UBTransmute of string ], 'b) Soteria.Symex.Compo_res.t * Soteria_rust_lib__Sptr.DecayMap.t option) Rustsymex.t

val cast_literal : from_ty:Charon.Types.literal_type -> to_ty:Charon.Types.literal_type -> [< Soteria_rust_lib.Rust_val.T.cval ] Typed.t -> ('a Rust_val.t, [> `UBTransmute of string ], 'b) DecayMapMonad.Result.t

Cast between literals; perform validation of the type's constraints. This is different from a transmute! It doesn't simply reinterpret the bits, but rather converts between types, e.g. rounding, truncating, extending, etc.

See also: https://doc.rust-lang.org/stable/reference/expressions/operator-expr.html#numeric-cast

val float_to_bv_bits : [< Soteria__Bv_values__Typed.T.sfloat ] Typed.t -> [ `NonZero | `Zero ] Typed.t DecayMapMonad.t

Converts a floating value to a bitvector, preserving it's bit representation. This is a symbolic process, because SMT-Lib has no operation for "float->bv" that preserves the bits, due to NaN.

See https://smt-lib.org/theories-FloatingPoint.shtml, "Conversions to other sorts"

val transmute_one : to_ty:Charon.Types.ty -> rust_val -> (rust_val, [> `UBDanglingPointer | `UBTransmute of string ], 'a) DecayMapMonad.Result.t

Transmutes a singular rust value, without splitting. This is under the assumption that size_of to_ty = size_of v, and both are primitives (literal or pointer).