Module Frontend.Lib

val target : string lazy_t

val root : Dune_site.Private_.Helpers.Location.t lazy_t

type t =

| Std

| Kani

| Miri

val name : t -> string

val path : t -> string

val exec_cargo : ?env:string list -> t -> string list -> unit

exec_cargo ?env lib args executes the command cargo <args> for the library lib.

raises PluginError
if the command fails, with the error message from Cargo.

val clean : t -> unit

Deletes the target directory of a target, to avoid duplicate builds

val compile : t -> unit

val with_compiled : t -> ((string * string) -> Cmd.t) -> Cmd.t