Module Soteria_rust_lib.Layout

exception CantComputeLayout of string * Charon.Types.ty

exception InvalidLayout of Charon.Types.ty

module Fields_shape : sig ... end

We use a custom type for the member offsets for layouts; this allows us to use a more efficient representation for arrays T; N, that doesn't require N offsets.

type layout = {

size : int;

align : int;

fields : Fields_shape.t;

}

val pp_layout : Stdlib.Format.formatter -> layout -> Ppx_deriving_runtime.unit

module Session : sig ... end

val is_int : Charon.Types.ty -> bool

val size_of_int_ty : Charon.Values.int_ty -> int

val size_of_uint_ty : Charon.Values.u_int_ty -> int

val size_of_literal_ty : Charon.Types.literal_type -> int

val align_of_literal_ty : Charon.Types.literal_type -> int

val empty_generics : Charon.Types.generic_args

val is_dst : Charon.Types.ty -> bool

If this is a dynamically sized type (requiring a fat pointer)

val size_to_fit : size:int -> align:int -> int

val max_array_len : int -> Z.t

val enum_discr_ty : Charon.Types.TypeDeclId.Map.key -> Charon.Types.ty

val layout_of : Charon.Types.ty -> layout

val layout_of_members : Charon.Types.ty list -> layout

val of_variant : Charon.Types.type_decl_id -> Charon.Types.variant -> layout

val of_enum_variant : Charon.Types.TypeDeclId.Map.key -> Charon.Types.VariantId.id -> layout

val resolve_trait_ty : Charon.Types.trait_ref -> Charon.Types.trait_item_name -> Charon.Types.ty

val offset_in_array : Charon.Types.ty -> int -> int

val layout_of_s : Charon.Types.ty -> layout Rustsymex.t

val size_of_s : Charon.Types.ty -> [> Soteria.C_values.Typed.T.sint ] Soteria.C_values.Typed.t Soteria_rust_lib__Rustsymex.SYMEX.t

val align_of_s : Charon.Types.ty -> [> Soteria.C_values.Typed.T.nonzero ] Soteria.C_values.Typed.t Soteria_rust_lib__Rustsymex.SYMEX.t

val is_signed : Charon.Types.literal_type -> bool

Assumes the input is a literal type of either TInt or TUInt

val min_value_z : Charon.Types.literal_type -> Z.t

val min_value : Charon.Types.literal_type -> [> Soteria.C_values.Typed.T.sint ] Soteria.C_values.Typed.t

val max_value_z : Charon.Types.literal_type -> Z.t

val max_value : Charon.Types.literal_type -> [> Soteria.C_values.Typed.T.nonzero ] Soteria.C_values.Typed.t

val size_to_uint : int -> Charon.Types.ty

val lit_to_unsigned : Charon.Types.literal_type -> Charon.Types.ty

val int_constraints : Charon.Types.literal_type -> [< Soteria__C_values__Typed.T.sint ] Soteria__C_values__Typed.t -> [> Soteria__C_values__Typed.sbool ] Soteria__C_values__Typed.t list

val constraints : Charon.Types.literal_type -> [< Soteria_rust_lib.Rust_val.T.cval ] Soteria.C_values.Typed.t -> Soteria_rust_lib.Rust_val.T.sbool Soteria.C_values.Typed.t list

val nondet_literal_ty : Charon.Types.literal_type -> Soteria_rust_lib.Rust_val.T.cval Soteria.C_values.Typed.t Rustsymex.t

val nondet : Charon.Types.ty -> 'a Rust_val.rust_val Rustsymex.t

nondet ~extern ~init ty returns a nondeterministic value for ty, along with some "state". It receives a function extern to get an optional external function that computes the arbitrary value; it tries using it, and otherwise guesses the valid values. init is the initial "state", that is modified and returned by extern.

val nondets : Charon.Types.ty list -> 'a Rust_val.rust_val list Rustsymex.t

val zeroed_lit : Charon.Types.literal_type -> Soteria_rust_lib.Rust_val.T.cval Soteria.C_values.Typed.t

val zeroed : null_ptr:'a -> Charon.Types.ty -> 'a Rust_val.rust_val Soteria.Soteria_std.Monad.OptionM.t

val is_inhabited : Charon.Types.ty -> bool

val as_zst : Charon.Types.ty -> 'a Rust_val.rust_val Soteria.Soteria_std.Monad.OptionM.t

Returns the given type as it's unique representant if it's a ZST; otherwise None.

FIXME: giltho: this plays awfully with polymorphism.

val apply_attribute : 'a Rust_val.t -> Charon.Meta.attribute -> (unit, [> `StdErr of string ], 'b) Soteria_rust_lib.Rustsymex.Result.t

Apply the compiler-attribute to the given value

val apply_attributes : 'a Rust_val.t -> Charon.Meta.attribute list -> (unit, [> `StdErr of string ], 'b) Soteria_rust_lib.Rustsymex.Result.t

val is_unsafe_cell : Charon.Types.ty -> bool

val ref_tys_in : ?include_ptrs:bool -> 'a Rust_val.rust_val -> Charon.Types.ty -> ('a Rust_val.full_ptr * Charon.Types.ty) Soteria.Soteria_std.List.t

Traverses the given type and rust value, and returns all findable references with their type (ignores pointers, except if include_ptrs is true). This is needed e.g. when needing to get the pointers along with the size of their pointee, in particular in nested cases.

val update_ref_tys_in : ('acc -> 'a Rust_val.full_ptr -> Charon.Types.ty -> Charon.Types.ref_kind -> ('a Rust_val.full_ptr * 'acc, 'e, 'f) Soteria_rust_lib.Rustsymex.Result.t) -> 'acc -> 'a Rust_val.rust_val -> Charon.Types.ty -> ('a Rust_val.rust_val * 'acc, 'e, 'f) Soteria_rust_lib.Rustsymex.Result.t

val is_abi_compatible : Charon.Types.ty -> Charon.Types.ty -> Ppx_deriving_runtime.bool

is_abi_compatible ty1 ty2 is true if a function expecting an argument of type ty1 can be called with an argument of type ty2.

The full specification is available at: https://doc.rust-lang.org/nightly/std/primitive.fn.html#abi-compatibility