Module Soteria_rust_lib.Sptr

module type DecayMapS = sig ... end

A map to store information on "decayed" pointers, namely mapping from locations to integers (their decayed) value.

module DecayMap : sig ... end

module D_abstr : sig ... end

module type S = sig ... end

The base type of pointers, permitting simple operations on the pointer type. The majority of relevant operations are exposed via the state monad's pointer module, Rust_state_m.S.Sptr.