Module Svalue.Ext

include module type of struct include Ext_base end

module Sv = Soteria.Bv_values.Svalue

Type definitions

module Unop = Ext_base.Unop

type 'g sv = ('g, 'g ext_t, 'g ext_ty) Sv.t

and 'g svty = 'g ext_ty Sv.ty

and 'g ptr = 'g Ext_base.ptr = {

ptr : 'g sv;

tag : Svalue.Ptr_tag.t option;

size : 'g sv;

align : 'g sv;

}

and ('sc, 'ag) block_value = ('sc, 'ag) Ext_base.block_value =

| Scalar of 'sc

| Aggregate of 'ag * Charon.Types.ty

The value inside a block: either a scalar or an aggregate that can be split further.

and ('sc, 'ag, 'ofs, 'sz) block = ('sc, 'ag, 'ofs, 'sz) Ext_base.block = {

value : ('sc, 'ag) block_value;

offset : 'ofs;

size : 'sz;

}

A block of an encoded value: a value, along with the offset it is at and the size it spans. ty is Some iff the value is a whole aggregate of that type.

and 'g ext_ty = 'g Ext_base.ext_ty =

| TEnum of Charon.Types.type_decl_ref
(*
the type decl ref of an enum
*)

| TUnion of Charon.Types.type_decl_ref
(*
the type decl ref of a union
*)

| TTuple of 'g svty list
(*
structs and tuples (ordered fields)
*)

| TArray of 'g svty * Soteria.Soteria_std.Z.t
(*
arrays (all elements share the same type)
*)

| TThinPtr

| TFullPtr

| TPtrMeta
(*
the type of a pointer's metadata (unit, length or metadata). this is not exposed to the user!
*)

| TPolyType

and 'g ptr_meta = 'g Ext_base.ptr_meta =

| MetaLen of 'g sv

| MetaVTable of 'g sv

| MetaUnit

and 'g ext_t = 'g Ext_base.ext_t =

| Ptr of 'g sv * 'g sv
(*
pointer, with meta
*)

| PtrMeta of 'g ptr_meta

| ThinPtr of 'g ptr
(*
thin pointer, without metadata but with extra info on the pointer
*)

| Enum of Charon.Types.variant_id * 'g sv list
(*
variant id * values
*)

| Tuple of 'g sv list
(*
structs and tuples: ordered values
*)

| Array of 'g sv Soteria.Soteria_std.Iarray.t
(*
arrays: ordered values, all of the same type
*)

| Union of ('g sv, 'g sv, 'g sv, 'g sv) block list
(*
list of blocks in the union
*)

| PolyVal of Charon.Types.type_var_id
(*
The opaque value of a type variable, identified by (type variable index, unique identifier).
*)

| Unop of Unop.t * 'g sv
(*
unary operation
*)

val equal_sv : 'g. ('g -> 'g -> Ppx_deriving_runtime.bool) -> 'g sv -> 'g sv -> Ppx_deriving_runtime.bool

val equal_svty : 'g. ('g -> 'g -> Ppx_deriving_runtime.bool) -> 'g svty -> 'g svty -> Ppx_deriving_runtime.bool

val equal_ptr : 'g. ('g -> 'g -> Ppx_deriving_runtime.bool) -> 'g ptr -> 'g ptr -> Ppx_deriving_runtime.bool

val equal_block_value : 'sc 'ag. ('sc -> 'sc -> Ppx_deriving_runtime.bool) -> ('ag -> 'ag -> Ppx_deriving_runtime.bool) -> ('sc, 'ag) block_value -> ('sc, 'ag) block_value -> Ppx_deriving_runtime.bool

val equal_block : 'sc 'ag 'ofs 'sz. ('sc -> 'sc -> Ppx_deriving_runtime.bool) -> ('ag -> 'ag -> Ppx_deriving_runtime.bool) -> ('ofs -> 'ofs -> Ppx_deriving_runtime.bool) -> ('sz -> 'sz -> Ppx_deriving_runtime.bool) -> ('sc, 'ag, 'ofs, 'sz) block -> ('sc, 'ag, 'ofs, 'sz) block -> Ppx_deriving_runtime.bool

val equal_ext_ty : 'g. ('g -> 'g -> Ppx_deriving_runtime.bool) -> 'g ext_ty -> 'g ext_ty -> Ppx_deriving_runtime.bool

val equal_ptr_meta : 'g. ('g -> 'g -> Ppx_deriving_runtime.bool) -> 'g ptr_meta -> 'g ptr_meta -> Ppx_deriving_runtime.bool

val equal_ext_t : 'g. ('g -> 'g -> Ppx_deriving_runtime.bool) -> 'g ext_t -> 'g ext_t -> Ppx_deriving_runtime.bool

val compare_sv : 'g. ('g -> 'g -> Ppx_deriving_runtime.int) -> 'g sv -> 'g sv -> Ppx_deriving_runtime.int

val compare_svty : 'g. ('g -> 'g -> Ppx_deriving_runtime.int) -> 'g svty -> 'g svty -> Ppx_deriving_runtime.int

val compare_ptr : 'g. ('g -> 'g -> Ppx_deriving_runtime.int) -> 'g ptr -> 'g ptr -> Ppx_deriving_runtime.int

val compare_block_value : 'sc 'ag. ('sc -> 'sc -> Ppx_deriving_runtime.int) -> ('ag -> 'ag -> Ppx_deriving_runtime.int) -> ('sc, 'ag) block_value -> ('sc, 'ag) block_value -> Ppx_deriving_runtime.int

val compare_block : 'sc 'ag 'ofs 'sz. ('sc -> 'sc -> Ppx_deriving_runtime.int) -> ('ag -> 'ag -> Ppx_deriving_runtime.int) -> ('ofs -> 'ofs -> Ppx_deriving_runtime.int) -> ('sz -> 'sz -> Ppx_deriving_runtime.int) -> ('sc, 'ag, 'ofs, 'sz) block -> ('sc, 'ag, 'ofs, 'sz) block -> Ppx_deriving_runtime.int

val compare_ext_ty : 'g. ('g -> 'g -> Ppx_deriving_runtime.int) -> 'g ext_ty -> 'g ext_ty -> Ppx_deriving_runtime.int

val compare_ptr_meta : 'g. ('g -> 'g -> Ppx_deriving_runtime.int) -> 'g ptr_meta -> 'g ptr_meta -> Ppx_deriving_runtime.int

val compare_ext_t : 'g. ('g -> 'g -> Ppx_deriving_runtime.int) -> 'g ext_t -> 'g ext_t -> Ppx_deriving_runtime.int

val hash_sv : ('g -> int) -> 'g sv -> int

val hash_svty : ('g -> int) -> 'g svty -> int

val hash_ptr : ('g -> int) -> 'g ptr -> int

val hash_block_value : ('g sv -> int) -> ('g sv -> int) -> ('g sv, 'g sv) block_value -> int

val hash_block : ('g sv -> int) -> ('g sv -> int) -> ('g sv -> int) -> ('g sv -> int) -> ('g sv, 'g sv, 'g sv, 'g sv) block -> int

val hash_ext_ty : ('g -> int) -> 'g ext_ty -> int

val hash_ptr_meta : ('g -> int) -> 'g ptr_meta -> int

val hash_ext_t : ('g -> int) -> 'g ext_t -> int

val pp_block_value : (Stdlib.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> (Stdlib.Format.formatter -> 'b -> Ppx_deriving_runtime.unit) -> Stdlib.Format.formatter -> ('a, 'b) block_value -> Ppx_deriving_runtime.unit

val pp_block : (Stdlib.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> (Stdlib.Format.formatter -> 'b -> Ppx_deriving_runtime.unit) -> (Stdlib.Format.formatter -> 'c -> Ppx_deriving_runtime.unit) -> (Stdlib.Format.formatter -> 'd -> Ppx_deriving_runtime.unit) -> Stdlib.Format.formatter -> ('a, 'b, 'c, 'd) block -> Ppx_deriving_runtime.unit

val pp_ptr_meta_kind : Stdlib.Format.formatter -> 'a ptr_meta -> unit

val pp_ext_ty : Ppx_deriving_runtime.Format.formatter -> 'g ext_ty -> Ppx_deriving_runtime.unit

val pp_svty : 'g svty Fmt.t

Types and Rust type conversions

val float_precision : Charon.Values.float_type -> Soteria.Bv_values.Svalue.FloatPrecision.t

val ty_of_rust : Charon.Types.ty -> 'g svty

val t_as_tuple : 'g svty -> 'g svty list

val t_as_array : 'g svty -> 'g svty * Soteria.Soteria_std.Z.t

val array_length : 'g svty -> Soteria.Soteria_std.Z.t

val array_elem_ty : 'g svty -> 'g svty

val t_as_enum : 'g svty -> Charon.Types.type_decl_ref

val t_as_union : 'g svty -> Charon.Types.type_decl_ref

val usize_bits : unit -> int

Smart constructors

We want to have our smart constructors accessible from within the extension, so that the extension's eval and mk functions can use the same smart constructors as the rest of the engine.

type 'g build = ('g, 'g ext_t, 'g ext_ty) Sv.t_kind -> 'g svty -> 'g sv

Thin pointers

val mk_thin_ptr : build:'a build -> 'a ptr -> 'a sv

val thin_ptr_part : build:'ghost build -> Unop.ptr_part -> 'ghost sv -> 'ghost sv

Fat pointers

val mk_full_ptr : build:'a build -> 'a sv -> 'a sv -> 'a sv

val mk_unit_meta : build:'a build -> unit -> 'a sv

val mk_len_meta : build:'a build -> 'a sv -> 'a sv

val mk_vtable_meta : build:'a build -> 'a sv -> 'a sv

val full_ptr_inner : build:'ghost build -> 'ghost sv -> 'ghost sv

val of_thin_ptr : build:'a build -> 'a sv -> 'a sv

val full_ptr_meta_raw : build:'ghost build -> 'ghost sv -> 'ghost sv

val ptr_meta_as : build:'ghost build -> Unop.ptr_meta -> 'ghost sv -> 'ghost sv

val full_ptr_meta : build:'ghost build -> Unop.ptr_meta -> 'ghost sv -> 'ghost sv

val full_ptr_set_inner : build:'ghost build -> 'ghost sv -> 'ghost sv -> 'ghost sv

Tuples

val mk_tuple : build:'a build -> 'a sv list -> 'a sv

val field_of : build:'ghost build -> int -> 'ghost sv -> 'ghost sv

val as_tuple : build:'ghost build -> 'ghost sv -> 'ghost sv list

val set_field : build:'ghost build -> int -> 'ghost sv -> 'ghost sv -> 'ghost sv

Enums

val mk_enum : build:'a build -> Charon.Types.type_decl_ref -> Charon.Types.variant_id -> 'a sv list -> 'a sv

val field_of_variant : build:'ghost build -> Charon.Types.variant_id -> int -> 'ghost sv -> 'ghost sv

val as_enum_of_variant : build:'ghost build -> Charon.Types.variant_id -> 'ghost sv -> 'ghost sv list

val set_field_of_variant : build:'ghost build -> Charon.Types.variant_id -> int -> 'ghost sv -> 'ghost sv -> 'ghost sv

val is_variant : build:'ghost build -> Charon.Types.variant_id -> 'ghost sv -> 'ghost sv

Arrays

val mk_array_of_svty : build:'a build -> 'a svty -> 'a sv Soteria.Soteria_std.Iarray.t -> 'a sv

val mk_array : build:'a build -> Charon.Types.ty -> 'a sv Soteria.Soteria_std.Iarray.t -> 'a sv

val array_field_of : build:'ghost build -> int -> 'ghost sv -> 'ghost sv

val as_array : build:'ghost build -> 'ghost sv -> 'ghost sv Soteria.Soteria_std.Iarray.t

val set_array_field : build:'ghost build -> int -> 'ghost sv -> 'ghost sv -> 'ghost sv

Unions and PolyVal (limited support)

val mk_union : build:'a build -> Charon.Types.type_decl_ref -> ('a sv, 'a sv, 'a sv, 'a sv) block list -> 'a sv

val mk_poly : build:'a build -> Charon.Types.type_var_id -> 'a sv

Operators

val apply_unop : build:'a build -> Unop.t -> 'a sv -> 'a sv

type 'ghost ty = 'ghost ext_ty

type 'ghost t = 'ghost ext_t

val equal : ('a -> 'a -> Ppx_deriving_runtime.bool) -> 'a ext_t -> 'a ext_t -> Ppx_deriving_runtime.bool

val compare : ('a -> 'a -> Ppx_deriving_runtime.int) -> 'a ext_t -> 'a ext_t -> Ppx_deriving_runtime.int

val equal_ty : ('a -> 'a -> Ppx_deriving_runtime.bool) -> 'a ext_ty -> 'a ext_ty -> Ppx_deriving_runtime.bool

val compare_ty : ('a -> 'a -> Ppx_deriving_runtime.int) -> 'a ext_ty -> 'a ext_ty -> Ppx_deriving_runtime.int

val pp_ty : Ppx_deriving_runtime.Format.formatter -> 'ghost ty -> Ppx_deriving_runtime.unit

val pp : 'a sv Fmt.t -> Stdlib.Format.formatter -> 'a ext_t -> Ppx_deriving_runtime.unit

val iter_vars_ptr : ('a sv -> 'b) -> 'a ptr -> 'b

val iter_vars_block_value : ('a -> 'b) -> ('a, 'a) block_value -> 'b

val iter_vars_block : ('a -> 'b) -> ('a, 'a, 'a, 'a) block -> 'b

val iter_vars : ('a sv -> unit) -> 'a ext_t -> unit

val hash_ghost : 'a -> int

val hash : 'a ext_t -> int

val hash_ty : 'a ext_ty -> int

val mk : 'a build -> 'a svty -> 'a ext_t -> 'a sv

val eval : ('a sv -> 'a sv) -> 'a ext_t -> 'a ext_t

val apply_list : (missing_var:'a -> 'b -> 'c -> 'd * 'b) -> missing_var:'a -> 'b -> 'c list -> 'd list * 'b

val apply_iarray : (missing_var:'a -> 'b -> 'c -> 'c * 'b) -> missing_var:'a -> 'b -> 'c iarray -> 'c iarray * 'b

val apply_subst_ptr : (missing_var:'a -> 'b -> 'c sv -> 'd sv * 'b) -> missing_var:'a -> 'b -> 'c ptr -> 'd ptr * 'b

val apply_block_value : (missing_var:'a -> 'b -> 'c -> 'd * 'e) -> missing_var:'a -> 'b -> ('c, 'c) block_value -> ('d, 'd) block_value * 'e

val apply_block : (missing_var:'a -> 'b -> 'c -> 'd * 'b) -> missing_var:'a -> 'b -> ('c, 'c, 'c, 'c) block -> ('d, 'd, 'd, 'd) block * 'b

val apply_subst : (missing_var:'a -> 'b -> 'c sv -> 'c sv * 'b) -> missing_var:'a -> 'b -> 'c ext_t -> 'c ext_t * 'b

val encode_ty : ('a Ext_base.svty -> Soteria.Smt.sexp) -> 'a Ext_base.ext_ty -> Soteria.Smt.sexp

val encode_value : ('a Ext_base.svty -> Soteria.Smt.sexp) -> ('a Ext_base.sv -> Soteria.Smt.sexp) -> ty:'a Ext_base.svty -> 'a Ext_base.ext_t -> Soteria.Smt.sexp