Module Svalue.Ext

module Sv = Soteria.Bv_values.Svalue

type 'ghost sv = ('ghost, 'ghost ext_t, 'ghost ext_ty) Sv.t

and 'ghost svty = 'ghost ext_ty Sv.ty

and 'ghost ptr = {

ptr : 'ghost sv;

tag : Svalue.Ptr_tag.t option;

size : 'ghost sv;

align : 'ghost sv;

}

and 'ghost ext_ty =

| TEnum of Charon.Types.type_decl_ref
(*
the type decl ref of an enum
*)

| TUnion of Charon.Types.type_decl_ref
(*
the type decl ref of a union
*)

| TTuple of 'ghost svty list
(*
structs and tuples (ordered fields)
*)

| TArray of 'ghost svty * Soteria.Soteria_std.Z.t
(*
arrays (all elements share the same type)
*)

| TThinPtr

| TFullPtr

| TPolyType

and 'ghost ext_t =

| Ptr of 'ghost sv * 'ghost sv option
(*
pointer, with optional meta
*)

| ThinPtr of 'ghost ptr
(*
thin pointer, without metadata but with extra info on the pointer
*)

| Enum of 'ghost sv * 'ghost sv list
(*
discriminant * values
*)

| Tuple of 'ghost sv list
(*
structs and tuples: ordered values
*)

| Array of 'ghost sv Soteria.Soteria_std.Iarray.t
(*
arrays: ordered values, all of the same type
*)

| Union of ('ghost sv * 'ghost sv * 'ghost sv) list
(*
list of blocks in the union, with their offset and size
*)

| PolyVal of Charon.Types.type_var_id
(*
The opaque value of a type variable, identified by (type variable index, unique identifier).
*)

val equal_sv : 'ghost. ('ghost -> 'ghost -> Ppx_deriving_runtime.bool) -> 'ghost sv -> 'ghost sv -> Ppx_deriving_runtime.bool

val equal_svty : 'ghost. ('ghost -> 'ghost -> Ppx_deriving_runtime.bool) -> 'ghost svty -> 'ghost svty -> Ppx_deriving_runtime.bool

val equal_ptr : 'ghost. ('ghost -> 'ghost -> Ppx_deriving_runtime.bool) -> 'ghost ptr -> 'ghost ptr -> Ppx_deriving_runtime.bool

val equal_ext_ty : 'ghost. ('ghost -> 'ghost -> Ppx_deriving_runtime.bool) -> 'ghost ext_ty -> 'ghost ext_ty -> Ppx_deriving_runtime.bool

val equal_ext_t : 'ghost. ('ghost -> 'ghost -> Ppx_deriving_runtime.bool) -> 'ghost ext_t -> 'ghost ext_t -> Ppx_deriving_runtime.bool

val compare_sv : 'ghost. ('ghost -> 'ghost -> Ppx_deriving_runtime.int) -> 'ghost sv -> 'ghost sv -> Ppx_deriving_runtime.int

val compare_svty : 'ghost. ('ghost -> 'ghost -> Ppx_deriving_runtime.int) -> 'ghost svty -> 'ghost svty -> Ppx_deriving_runtime.int

val compare_ptr : 'ghost. ('ghost -> 'ghost -> Ppx_deriving_runtime.int) -> 'ghost ptr -> 'ghost ptr -> Ppx_deriving_runtime.int

val compare_ext_ty : 'ghost. ('ghost -> 'ghost -> Ppx_deriving_runtime.int) -> 'ghost ext_ty -> 'ghost ext_ty -> Ppx_deriving_runtime.int

val compare_ext_t : 'ghost. ('ghost -> 'ghost -> Ppx_deriving_runtime.int) -> 'ghost ext_t -> 'ghost ext_t -> Ppx_deriving_runtime.int

val pp_ext_ty : Ppx_deriving_runtime.Format.formatter -> 'ghost ext_ty -> Ppx_deriving_runtime.unit

val pp_svty : 'ghost svty Fmt.t

module Rust_ext : Sv.Value_ext with type 'ghost t = 'ghost ext_t and type 'ghost ty = 'ghost ext_ty