Module Value_codec.Encoder

Parameters

module Sptr : Sptr.S

Signature

type nonrec rust_val = Sptr.t Rust_val.t

val pp_rust_val : Sptr.t Rust_val.t Fmt.t

val encode : offset:Soteria_rust_lib.Layout.T.sint Typed.t -> rust_val -> Charon.Types.ty -> ((rust_val * Soteria_rust_lib.Layout.T.sint Typed.t) Soteria.Soteria_std.Iter.t, [> `InvalidLayout of Charon.Types.ty ], 'f) Rustsymex.Result.t

encode ?offset v ty Converts a Rust_val.t of type ty into an iterator over its sub values, along with their offset. Offsets all blocks by offset if specified

val check_valid : fake_read: ((Sptr.t * Sptr.t Rust_val.meta) -> Charon.Types.ty -> 'a -> Soteria_rust_lib__Rustsymex.MonadState.t -> ((([> `InvalidLayout of Charon.Types.ty | `RefToUninhabited | `UBDanglingPointer | `UBTransmute of string ] as 'b) option * 'a) * Soteria_rust_lib__Rustsymex.MonadState.t) Soteria_rust_lib__Rustsymex.MonoSymex.t) -> Sptr.t Rust_val.t -> Charon.Types.ty -> 'a -> Soteria_rust_lib__Rustsymex.MonadState.t -> (((unit, 'c, 'd) Compo_res.t * 'a) * Soteria_rust_lib__Rustsymex.MonadState.t) Soteria_rust_lib__Rustsymex.MonoSymex.t

Ensures this value is valid for the given type. This includes checking pointer metadata, e.g. slice lengths and vtables. The fake_read function is used to simulate reading from memory to check the validity of a pointee type.

val with_constraints : ty:Charon.Types.literal_type -> Soteria_rust_lib.Rust_val.T.sint Typed.t -> Soteria_rust_lib__Sptr.DecayMap.t option -> (('a Rust_val.t, [> `UBTransmute of string ], 'b) Soteria.Symex.Compo_res.t * Soteria_rust_lib__Sptr.DecayMap.t option) Rustsymex.t

val cast_literal : from_ty:Charon.Types.literal_type -> to_ty:Charon.Types.literal_type -> [< Soteria_rust_lib.Layout.T.cval ] Typed.t -> ('a Rust_val.t, [> `UBTransmute of string ], 'b) DecayMapMonad.Result.t

Cast between literals; perform validation of the type's constraints. This is different from a transmute! It doesn't simply reinterpret the bits, but rather converts between types, e.g. rounding, truncating, extending, etc.

See also: https://doc.rust-lang.org/stable/reference/expressions/operator-expr.html#numeric-cast

val float_to_bv_bits : [< Soteria__Bv_values__Typed.T.sfloat ] Typed.t -> [ `NonZero | `Zero ] Typed.t DecayMapMonad.t

Converts a floating value to a bitvector, preserving it's bit representation. This is a symbolic process, because SMT-Lib has no operation for "float->bv" that preserves the bits, due to NaN.

See https://smt-lib.org/theories-FloatingPoint.shtml, "Conversions to other sorts"

val transmute_one : to_ty:Charon.Types.ty -> rust_val -> (rust_val, [> `UBDanglingPointer | `UBTransmute of string ], 'a) DecayMapMonad.Result.t

Transmutes a singular rust value, without splitting. This is under the assumption that size_of to_ty = size_of v, and both are primitives (literal or pointer).

val nondet : Charon.Types.ty -> (rust_val, [> `InvalidLayout of Charon.Types.ty ], 'f) Rustsymex.Result.t

nondet ~extern ~init ty returns a nondeterministic value for ty, along with some "state". It receives a function extern to get an optional external function that computes the arbitrary value; it tries using it, and otherwise guesses the valid values. init is the initial "state", that is modified and returned by extern.

val nondets : Charon.Types.ty list -> Soteria_rust_lib__Rustsymex.MonadState.t -> ((Sptr.t Rust_val.t list, [> `InvalidLayout of Charon.Types.ty ], 'f) Soteria.Symex.Compo_res.t * Soteria_rust_lib__Rustsymex.MonadState.t) Soteria_rust_lib__Rustsymex.MonoSymex.t

val apply_attribute : 'a Rust_val.t -> Charon.Meta.attribute -> Soteria_rust_lib__Rustsymex.MonadState.t -> ((unit, [> `StdErr of string ], 'b) Soteria.Symex.Compo_res.t * Soteria_rust_lib__Rustsymex.MonadState.t) Soteria_rust_lib__Rustsymex.MonoSymex.t

Apply the compiler-attribute to the given value

val apply_attributes : 'a Rust_val.t -> Charon.Meta.attribute list -> (unit, [> `StdErr of string ], 'b) Rustsymex.Result.t

val ref_tys_in : ?include_ptrs:bool -> rust_val -> Charon.Types.ty -> (Sptr.t Rust_val.full_ptr * Charon.Types.ty) Soteria.Soteria_std.List.t

Traverses the given type and rust value, and returns all findable references with their type (ignores pointers, except if include_ptrs is true). This is needed e.g. when needing to get the pointers along with the size of their pointee, in particular in nested cases.

val update_ref_tys_in : ('acc -> Sptr.t Rust_val.full_ptr -> Charon.Types.ty -> Charon.Types.ref_kind -> (Sptr.t Rust_val.full_ptr * 'acc, 'e, 'f) Rustsymex.Result.t) -> 'acc -> rust_val -> Charon.Types.ty -> (rust_val * 'acc, 'e, 'f) Rustsymex.Result.t